% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-Chemoprotection.R
\name{add_bites_to_trial_files}
\alias{add_bites_to_trial_files}
\title{Add Bite Events to Trial Files}
\usage{
add_bites_to_trial_files(
  baselineFolder,
  outputFolder,
  TimePreDose,
  bite_ADM,
  bite_AMT = 640,
  nsubjs,
  ntrials,
  biteTimeGenerator,
  biteTimeGeneratorArgs
)
}
\arguments{
\item{baselineFolder}{Character string. The directory path where base simulation output files are
stored and will be updated. This should be the folder containing the baseline trial files
before adding bite events; these baseline files should contain PKPD parameters for all subjects}

\item{outputFolder}{Character string. The directory path where the updated simulation output files
with added bite events will be stored. The function will copy the contents of \code{baselineFolder}
to this location and then update the trial files with the new bite events.}

\item{TimePreDose}{Numeric. The time before the first dose (in hours). If greater than zero,
an initial zero dose event is added at time zero. Including \code{TimePreDose} permits infectious
bites to occur before the therapy is given.}

\item{bite_ADM}{Integer. The administration (ADM) code representing the bite input in the
simulation structural model used in the trial files you are updating. Example: if your
structural model has parasites being INPUT3, bite_ADM should = 3}

\item{bite_AMT}{Numeric. The amount of malaria parasites introduced per bite event. Default: 640,
as estimated from sporozoite human challenge studies of DSM265}

\item{nsubjs}{Integer. The number of subjects in each trial. This is used for indexing and
adjusting subject IDs when updating trial files. This must be equal to nsubjs in \code{baselineFolder}}

\item{ntrials}{Integer. The number of trials. This is used for indexing and adjusting trial IDs
when updating trial files. This must be equal to ntrials in \code{baselineFolder}}

\item{biteTimeGenerator}{Function. A function used to generate bite times. This function should
accept arguments specified in \code{biteTimeGeneratorArgs}. The bite time generator function
should be designed to generate bite times according to your simulation requirements.}

\item{biteTimeGeneratorArgs}{List. A named list of arguments to be passed to the \code{biteTimeGenerator}
function. These arguments customize the bite time generation process and should match the
parameters expected by your \code{biteTimeGenerator} function.}
}
\description{
This function adds simulated mosquito bite events to existing trial files using a specified
bite time generator function. It copies a pre-defined set of trial files from a baseline folder
to an output folder, and then updates the trial files with these bite events. The bite times
are generated using the provided bite time generator function, which accepts arguments specified
in \code{biteTimeGeneratorArgs}. This allows for flexible and customized bite time generation
according to the requirements of your simulation.
}
\details{
The function operates in two main steps:

\strong{Step 1:} It copies the baseline trial files from \code{baselineFolder} to \code{outputFolder}. This
ensures that PKPD parameters, regressors, and dosing events are retained.

\strong{Step 2:} It processes each trial file in \code{outputFolder} to add bite events. For each subject
in each trial, it generates bite times using the provided \code{biteTimeGenerator} function and the
arguments specified in \code{biteTimeGeneratorArgs}. The bite events are then added to the trial's
event table.

If \code{TimePreDose} is greater than zero, a zero dose event is added at time zero to ensure that
simulations start at time zero and to allow bites to occur before dosing.
}
\references{
Cherkaoui-Rbati MH, Andenmatten N, Burgert L, et al. (2023). A pharmacokinetic-pharmacodynamic model
for chemoprotective agents against malaria. \emph{CPT Pharmacometrics Syst Pharmacol}, \strong{12}, 50-61. \doi{10.1002/psp4.12875}
}
\seealso{
Other Chemoprotection: 
\code{\link{evaluate_BreakthroughEvent}()},
\code{\link{generate_ChemoModel_FluxLimiter}()},
\code{\link{generate_ChemoModel_UpwindFirstOrder}()},
\code{\link{generate_bite_time_poisson}()}
}
\author{
Sam Jones (MMV, \email{joness@mmv.org})
}
\concept{Chemoprotection}
