% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-GraphicalFunctions.R
\name{adjust_ggplotTheme}
\alias{adjust_ggplotTheme}
\title{Adjust ggplot Theme}
\usage{
adjust_ggplotTheme(gr, style = c("MMV", "MMVvpc")[1])
}
\arguments{
\item{gr}{ggplot object to adjust theme.}

\item{style}{Style to choose between \code{"MMV"} or \code{"MMVvpc"} (Default: \code{"MMV"}).}
}
\value{
`gr` with adjusted theme
}
\description{
Adjust the theme of the ggplot object `gr` to various style (e.g. to MMV style).
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Graphical Functions}
