% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{applyCureThreshold}
\alias{applyCureThreshold}
\title{applyCureThreshold}
\usage{
applyCureThreshold(x, threshold, PDname = "Parasitemia")
}
\arguments{
\item{PDname}{Default: 'Parasitemia'}
}
\description{
applyCureThreshold
}
\seealso{
Other Simulations: 
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Aline Fuchs (MMV)
}
\concept{Simulations}
