% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{assess_MICMPC90PRR48}
\alias{assess_MICMPC90PRR48}
\title{Estimate MIC, MPC90, PRR48 parameters}
\usage{
assess_MICMPC90PRR48(
  modelFolder,
  dataPath,
  outputFolder,
  yaxis = NULL,
  yaxisunit = "",
  GRmodelFolder = NULL,
  PKmodelFolder = "../04-Output/S131-01_PKmodel",
  Apparent = TRUE,
  Tk0 = "Tk0",
  Indv = FALSE,
  ActivityPath = NULL,
  addArgs_sim_IQRmodel = NULL
)
}
\arguments{
\item{modelFolder}{Relative or absolute path to the IQRnlmeProject PD model}

\item{dataPath}{Path to the dataset}

\item{outputFolder}{Folder where the plots of MIC, MPC90 and PRR48 will be saved}

\item{yaxis}{Default: `NULL`}

\item{yaxisunit}{Default: ''}

\item{GRmodelFolder}{Default: `NULL`}

\item{PKmodelFolder}{Default: '../04-Output/S131-01_PKmodel'}

\item{Apparent}{Default: `TRUE`}

\item{Tk0}{Default: 'Tk0'}

\item{Indv}{Default: `FALSE`}

\item{ActivityPath}{Default: `NULL`}

\item{addArgs_sim_IQRmodel}{named list (default NULL) with additional arguments to be passed to sim_IQR_model. This
argument is relevant only if \code{Apparent == TRUE}.}
}
\value{

}
\description{
Required: `Emax` `EC50`, `hill`. Emax, cleareance, delay, turnover and effect compartment models
}
\seealso{
Other Key Parameters: 
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR28}()},
\code{\link{evaluate_APR}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV), Sam Jones (MMV)
}
\concept{Key Parameters}
