% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{assess_PDcomboModel}
\alias{assess_PDcomboModel}
\title{assess_PDcomboModel}
\usage{
assess_PDcomboModel(
  ModelToAssess,
  fileDataGeneral,
  hiDoseGr,
  LLOQ,
  modelFile = NULL,
  PDname = "Parasitemia Total",
  abs0inputs = NULL,
  abs0Tk0param = NULL
)
}
\arguments{
\item{modelFile}{Default: \code{NULL}}

\item{PDname}{Default: 'Parasitemia Total'}

\item{abs0inputs}{Default: \code{NULL}}

\item{abs0Tk0param}{Default: \code{NULL}}
}
\description{
assess_PDcomboModel
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
