% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{aux_constructCIpercentiles}
\alias{aux_constructCIpercentiles}
\title{Construct CI percentiles from CI level}
\usage{
aux_constructCIpercentiles(CIlevel)
}
\arguments{
\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}
}
\value{
The vector \code{CI.percentiles} with names.
}
\description{
Construct CI percentiles from CI level
}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{adjust_Dose}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{create_PKPDsimtime}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{getDoselevel}()},
\code{\link{get_IQRdistribution}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{get_MMVmalariaPath}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{saveActivityInfo}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{General Functions}
