% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{aux_removeEscapeChar}
\alias{aux_removeEscapeChar}
\title{Removal of escape characters}
\usage{
aux_removeEscapeChar(x, escapeChars = c("\\r\\n", "\\n", "\\r", "\\t"))
}
\arguments{
\item{x}{A character vector.}

\item{escapeChars}{Escape character to substituted with space (Default: \code{c("\\r\\n", "\\n", "\\r", "\\t")})}
}
\value{
\code{x} where all escape character defined in \code{escapeChars} were replaced by a space.
}
\description{
Removal of escape characters
}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{adjust_Dose}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{create_PKPDsimtime}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{getDoselevel}()},
\code{\link{get_IQRdistribution}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{get_MMVmalariaPath}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{saveActivityInfo}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{General Functions}
