% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{backwardElimination}
\alias{backwardElimination}
\title{backwardElimination}
\usage{
backwardElimination(
  model,
  dosing,
  data,
  modelSpec,
  covariateToTest,
  COVcentering = NULL,
  projectPath = "PKmodels",
  alpha = 0.05,
  tool = "MONOLIX",
  toolVersion = NULL,
  ncores = 1,
  Nparallel = 1,
  setting = NULL
)
}
\arguments{
\item{COVcentering}{Default: \code{NULL}}

\item{projectPath}{Default: 'PKmodels'}

\item{alpha}{Default: 0.05}

\item{tool}{Default: 'MONOLIX'}

\item{toolVersion}{Default: \code{NULL}}

\item{ncores}{Default: 1}

\item{Nparallel}{Default: 1}

\item{setting}{Default: \code{NULL}}
}
\value{

}
\description{
backwardElimination
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
