% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{barplotACPR}
\alias{barplotACPR}
\title{Create stratified barplot of ACPR}
\usage{
barplotACPR(data, type = "Crude", acprDay = 28, stratifyBy = "BWB")
}
\arguments{
\item{data}{general dataset}

\item{type}{'Crude' or 'PCR-Adjusted'}

\item{acprDay}{ACPR assessment day, can only be 28, 42, or 63}

\item{stratifyBy}{Column to stratify by}
}
\value{
A ggplot object with the barplot of ACPR
}
\description{
Create stratified barplot of ACPR
}
\seealso{
Other DataExploration: 
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{DataExploration}
