% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-MMVnlmeParamSamplingXLS.R
\name{calcIndParamValues_MMVmalariaXLS}
\alias{calcIndParamValues_MMVmalariaXLS}
\title{Calculate individual parameter values by applying sampled random effects
to typical individual parameter values}
\usage{
calcIndParamValues_MMVmalariaXLS(spec, typicalIndParamValues, randomEffects)
}
\arguments{
\item{spec}{A filename (character string) denoting the path to a GPF excel file, or a GPF object, or an IQRnlmeParamSpec object.}

\item{typicalIndParamValues}{a data.frame specifying typical individual
parameter values on an original (not transformed to normal) scale for
the model parameters in \code{spec}. Apart from the
parameter columns, this data.frame should have an \code{ID.POP} and an
\code{ID} column corresponding to the columns \code{ID.POP} and \code{ID}
in \code{randomEffects}. This data.frame
should have the same number of rows as \code{randomEffects}. See
\code{\link{calcTypicalIndParamValues}}.}

\item{randomEffects}{a data.frame specifying random effect values on a
normal (transformed) scale for the model parameters in \code{spec}.
Apart from the parameter columns, this data.frame should have an \code{ID.POP}
and an \code{ID} integer columns corresponding to population id and individual id
within a population respectively. See \code{\link{sampleRandomEffects}}.}
}
\value{
a data.frame with the same column names as the arguments \code{typicalIndParamValues}
and \code{randomEffects}. The individual parameter values in this data.frame are on
the original scale (not transformed to a normal scale).
}
\description{
Calculate individual parameter values by applying sampled random effects
to typical individual parameter values
}
\examples{
\dontrun{
#See the examples for IQRtools::sampleIndParamValues
}
# See examples in \code{\link{sampleIndParamValues_MMVmalariaXLS}}.

}
\seealso{
Other functions in the NLME parameter sampling API: 
\code{\link{calcTypicalIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleIDs_MMVmalariaXLS}()},
\code{\link{sampleIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleParamFromUncertainty_MMVmalariaXLS}()},
\code{\link{sampleRandomEffects_MMVmalariaXLS}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{functions in the NLME parameter sampling API}
