% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{censorGametocyte}
\alias{censorGametocyte}
\title{censorGametocyte}
\usage{
censorGametocyte(
  data,
  threshold,
  nameAll = "Parasitemia asexual + gametocyte",
  nameGam = "Parasitemia female gametocyte",
  FLAGfollowing = TRUE
)
}
\arguments{
\item{data}{Data frame or IQRdataGENERAL object with gametocyte data.}

\item{threshold}{Percentage of gametocytes of total. If this is exceeded total will be censored.}

\item{nameAll}{Default: 'Parasitemia asexual + gametocyte'}

\item{nameGam}{Default: 'Parasitemia female gametocyte'}

\item{FLAGfollowing}{Default: TRUE}
}
\value{
A vector of IXGDF indices where the total parasitemia values should be censored.
}
\description{
This function censors the total parasitemia values based on the percentage of gametocytes present in the total parasitemia.
}
\seealso{
Other Data Preparation: 
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan)
}
\concept{Data Preparation}
