% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ActivityMonitoring.R
\name{check_ActivityInfo_PiNK}
\alias{check_ActivityInfo_PiNK}
\title{Check Activity Information on PiNK}
\usage{
check_ActivityInfo_PiNK(
  ActivityInfoFile = "../01-Data/S001-ActivityInfo",
  tagFile = "../tags.RData"
)
}
\arguments{
\item{ActivityInfoFile}{Path to the activity information file (Default: \code{'../01-Data/S001-ActivityInfo'}).}

\item{tagFile}{Path to the tag file  saved from PiNK (Default: \code{'../tags.RData'}).}
}
\value{
Message indicating if valid or not.
}
\description{
Automatic check that the information provided by the modeler are valid for the activities on PiNK.
}
\seealso{
Other Activity Monitoring: 
\code{\link{check_ActivityInfo_Sdrive}()},
\code{\link{check_ActivityInfo}()},
\code{\link{check_ActivityResults}()},
\code{\link{create_RprofileMMV}()},
\code{\link{generate_ActivityContent}()},
\code{\link{generate_ActivityResults}()},
\code{\link{generate_DuplicatedActivityReport}()},
\code{\link{generate_MissingActivityReport}()},
\code{\link{generate_OverviewFile}()},
\code{\link{get_ActivityInfo}()},
\code{\link{get_ActivityList}()},
\code{\link{list_ActivityType}()},
\code{\link{list_GenericType}()},
\code{\link{set_MMVsettings}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Activity Monitoring}
