% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ActivityMonitoring.R
\name{check_ActivityInfo_Sdrive}
\alias{check_ActivityInfo_Sdrive}
\title{Check Activity Information on S Drive}
\usage{
check_ActivityInfo_Sdrive(ActivityInfoFile = "../01-Data/S001-ActivityInfo")
}
\arguments{
\item{ActivityInfoFile}{Path to the activity information file (Default: \code{'../01-Data/S001-ActivityInfo'}).}
}
\value{
Message indicating if valid or not.
}
\description{
Automatic check that the information provided by the modeler are valid for the activities on the S Drive.
}
\seealso{
Other Activity Monitoring: 
\code{\link{check_ActivityInfo_PiNK}()},
\code{\link{check_ActivityInfo}()},
\code{\link{check_ActivityResults}()},
\code{\link{create_RprofileMMV}()},
\code{\link{generate_ActivityContent}()},
\code{\link{generate_ActivityResults}()},
\code{\link{generate_DuplicatedActivityReport}()},
\code{\link{generate_MissingActivityReport}()},
\code{\link{generate_OverviewFile}()},
\code{\link{get_ActivityInfo}()},
\code{\link{get_ActivityList}()},
\code{\link{list_ActivityType}()},
\code{\link{list_GenericType}()},
\code{\link{set_MMVsettings}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Activity Monitoring}
