% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ActivityMonitoring.R
\name{check_ActivityResults}
\alias{check_ActivityResults}
\title{Check Activity Results}
\usage{
check_ActivityResults(
  ActivityInfoFile = "../01-Data/S001-ActivityInfo",
  ActivityResultsFile = "../01-Data/S999-ActivityResults",
  tagFile = "../tags.RData",
  FLAGMMVazure = ("MMVazure" \%in\% .packages(all.available = TRUE))
)
}
\arguments{
\item{ActivityInfoFile}{Path to an R object containing all the information of the activity of interest.}

\item{ActivityResultsFile}{Path to an R object containing all the results of the activity of interest.}

\item{tagFile}{Path to the tag file  saved from PiNK (Default: \code{'../tags.RData'}).}
}
\value{
Message indicating if valid or not.
}
\description{
The function check if the results for the activity of interest are valid,
i.e. that the expected results for a given Activity Type are furnished.
}
\seealso{
Other Activity Monitoring: 
\code{\link{check_ActivityInfo_PiNK}()},
\code{\link{check_ActivityInfo_Sdrive}()},
\code{\link{check_ActivityInfo}()},
\code{\link{create_RprofileMMV}()},
\code{\link{generate_ActivityContent}()},
\code{\link{generate_ActivityResults}()},
\code{\link{generate_DuplicatedActivityReport}()},
\code{\link{generate_MissingActivityReport}()},
\code{\link{generate_OverviewFile}()},
\code{\link{get_ActivityInfo}()},
\code{\link{get_ActivityList}()},
\code{\link{list_ActivityType}()},
\code{\link{list_GenericType}()},
\code{\link{set_MMVsettings}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Activity Monitoring}
