% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{check_dataGeneralMMV_Center}
\alias{check_dataGeneralMMV_Center}
\title{check_dataGeneralMMV_Center
Checks if CENTER and CENTERNAME is well defined within \code{dataGeneral}}
\usage{
check_dataGeneralMMV_Center(dataGeneral)
}
\arguments{
\item{dataGeneral}{IQRdataGENERAL object}
}
\value{
IQRdataGENERAL object
}
\description{
Checks if CENTER and CENTERNAME (number and name of the center conducting the study) is defined within \code{dataGeneral}, according to MMV's standards.
}
\details{
The name and number of the center conducting the study will be checked according to the up-to-date center list, see \link{list_Center}.
If both CENTERNAME and CENTER columns exist in \code{dataGeneral}, CENTERNAME column is taken as the reference to correct CENTER column.
If only CENTERNAME column exists in \code{dataGeneral}, the function will create CENTER column.
If only CENTER column exists in \code{dataGeneral}, the function will create CENTERNAME column.
}
\seealso{
\link{list_Center}

Other Data Preparation: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{censorGametocyte}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{convert_catCovToTxt}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{list_Center}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()},
\code{\link{summarizeReplicates}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Mohammed H. Cherkaoui (MMV), \email{cherkaouim@mmv.org}
}
\concept{Data Preparation}
