% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{Check_MissingDatabyNAME}
\alias{Check_MissingDatabyNAME}
\title{Check_MissingDatabyNAME}
\usage{
Check_MissingDatabyNAME(dataMaster, NAME, colNAME = c("USUBJID"))
}
\arguments{
\item{dataMaster}{Master dataset in MMV format}

\item{NAME}{list of variable to test}

\item{colNAME}{Column name to use to identify subject (default: "USUBJID")}
}
\value{
dataSummaryMissing, which is a data.frame with the list of missing subject for each variable in \code{NAME}
}
\description{
Check for for each variable defined in \code{NAME}, if it is missing for
any subject defined in the column USUBJID  of the dataset \code{dataMaster}.
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{convert_catCovToTxt}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{list_Center}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()},
\code{\link{summarizeReplicates}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Mohammed H. Cherkaoui (MMV), \email{cherkaouim@mmv.org}
}
\concept{Data Preparation}
