% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{clopperPearsonMMV}
\alias{clopperPearsonMMV}
\title{Clopper-Pearson Test}
\usage{
clopperPearsonMMV(k, n, CI = 0.9)
}
\arguments{
\item{k}{Number of successful observations of a trial}

\item{n}{Number of observation}

\item{CI}{Confidence Interval to estimate (Default: \code{0.9})}
}
\value{
Return a vector with the lower \code{CI_low} and high \code{CI_high} limit of the CI.
}
\description{

}
\examples{
# Probability of success:
p <- 0.8
# Number of observations:
n <- 100
# Number of successful observation:
k <- rbinom(1,n,p)
# Empirical probability of success:
p_Empirical <- k/n
# 90\% confidence interval of the empirical probability of success:
CI <- clopperPearsonMMV(k,n,CI=0.9)

}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{add_FuncToEnv}()},
\code{\link{adjust_Dose}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{cbindMMV}()},
\code{\link{combMMV}()},
\code{\link{convert_Unit}()},
\code{\link{create_PKPDsimtime}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{getDoselevel}()},
\code{\link{getMKLthreadsMMV}()},
\code{\link{get_IQRdistribution}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{get_MMVmalariaPath}()},
\code{\link{libraryMMV}()},
\code{\link{load_RData}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rbindMMV}()},
\code{\link{rectintMMV}()},
\code{\link{reverse_List}()},
\code{\link{saveActivityInfo}()},
\code{\link{saveMMV}()},
\code{\link{setMKLthreadsMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
