% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{collapseMMV}
\alias{collapseMMV}
\title{Customizable Collapse Function}
\usage{
collapseMMV(x, collapseSymbole = ", ", andSymbole = " & ", messageEmpty = NULL)
}
\arguments{
\item{x}{Character vector of length \code{n}.}

\item{collapseSymbole}{Character used as a separator for the \code{n-1} first elements in \code{x} (Default: \code{", "})}

\item{andSymbole}{Character used as a separator for the last element in \code{x} (Default: \code{" & "}).}

\item{messageEmpty}{Character to print if \code{x} is empty (Default: \code{NULL}).}
}
\value{
A character string where the elements of \code{x} were collapsed.
}
\description{
Collapses a character vector of any length into a character string separated by
\code{collapseSymbole}, except for the last element where \code{andSymbole} is used.
}
\examples{
PMXmember <- c("Aline", "Catalina", "Mohammed", "Nathalie")
msg       <- paste0("The PMX group is composed of ", collapseMMV(PMXmember), ".")
cat(msg)

}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions, Editing Functions}
