% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_DataModelMMV}
\alias{compare_DataModelMMV}
\title{compare_DataModelMMV}
\usage{
compare_DataModelMMV(
  modelFolder,
  dataFile = NULL,
  outputFolder = ".",
  ylabel = NULL,
  stratify = "TRTNAME",
  catToStratify = NULL,
  fileName = "DataPredComparison",
  logY = FALSE,
  figLimX = NULL,
  figLimY = NULL,
  predToPlot = NULL,
  doseCOL = "DOSELEVEL",
  doseMultCOL = "DOSEMULT",
  ActivityPath = NULL
)
}
\arguments{
\item{modelFolder}{Path containing the IQRnlmeProject}

\item{dataFile}{Character string with path to IQRdataGeneral object}

\item{outputFolder}{Character string with path where to save output}

\item{ylabel}{Character string defining y axis label.}

\item{stratify}{Character string defining column name on which to stratify
the plots. Must be one of "USUBJID", "ID", "TRTNAME", "STUDY". Define catToStratify
for stratification of other categorical covariates. Default: 'TRTNAME'}

\item{catToStratify}{Character string defining categorical covariate on
which to stratifyDefault: \code{NULL}}

\item{fileName}{Character string defining name of figure output. Default: 'DataPredComparison'}

\item{logY}{Logical: log 10 y axis on plot. Default: \code{FALSE}}

\item{figLimX}{Numeric, xlim to add to plot with \code{ggplot2::coord_cartesian}.
Default: \code{NULL}}

\item{figLimY}{Numeric, ylim to add to plot with \code{ggplot2::coord_cartesian}.
Default: \code{NULL}}

\item{predToPlot}{Which prediction to plot (POP, XPRED, INDIV or IPRED);
default: \code{NULL} in which case the function will plot both population and
individual predictions.}

\item{doseCOL}{Character string defining column name containing doselevel.
Default: 'DOSELEVEL'}

\item{doseMultCOL}{Character string defining column name containing number
of doses Default: 'DOSEMULT'}

\item{ActivityPath}{Default: \code{NULL}}
}
\value{
Figure showing the comparison of data and model predictions.
}
\description{
Compare and visualize model predictions against data,
stratified by chosen parameters
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
