% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_DataModelTimeRecrudCombo}
\alias{compare_DataModelTimeRecrudCombo}
\title{compare_DataPredTimeRecrudCombo
Isobole - plots for Combo model}
\usage{
compare_DataModelTimeRecrudCombo(
  ModelCombo,
  PKparameters,
  doses1,
  doses2,
  LLOQ,
  fileDataGeneral,
  PDname = "Parasitemia Total",
  CONDcol = "TRTNAME",
  filename = NULL,
  FLAGreturnObject = FALSE,
  logScale = FALSE,
  Time2RecDef = "MMV",
  setting = NULL,
  ActivityPath = NULL
)
}
\arguments{
\item{ModelCombo}{Path to an IQRsysProject or a sysfit-object (IQRtools <= 0.9)}

\item{PKparameters}{Named vector of PK parameters}

\item{doses1, doses2}{Numeric vectors of dosing values of drug 1 and 2}

\item{LLOQ}{numeric, the LLOQ of the PD readout}

\item{fileDataGeneral}{filepath to dataset to extract time-to-recrudescence from}

\item{PDname}{Default: 'Parasitemia Total'. Entry in NAME-column of fileDataGeneral to denote PD observations}

\item{CONDcol}{Default: 'TRTNAME'. Name of column to extract the CONDITION from}

\item{filename}{Default: \code{NULL}. filename for graphics output.}

\item{FLAGreturnObject}{Default: \code{FALSE}. Return the plot objects (TRUE) or don't return anything (FALSE)}

\item{logScale}{Default: \code{FALSE}. Plot on log-scale? If yes, doses with AMT = 0 are set to min(doseX)/10 for plotting.}

\item{Time2RecDef}{Default: 'MMV'. Argument going to getTimeRecrudescence}

\item{setting}{Default: \code{NULL}. List of settings going to simTimeRecrudescenceCombo}
}
\value{
Either nothing or a list of plots: list(facetted_plot, list_of_single_plots)
}
\description{
Simulation of combinations with estimates parameter set to obtain
\itemize{
\item predicted concentrations
\item predicted time to recrudescence
Visualization of predicted and observed time to recrudescance in 2D
Calculation of sum of least squares for predicted-versus-observed time profiles and time to recrudescence
}
}
\seealso{
simTimeRecrudescenceCombo
}
\author{
Anne Kümmel (IntiQuan), Mohammed Cherkaoui (MMV)
}
\concept{Combo-model assessment functions}
