% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_IndPredFits}
\alias{compare_IndPredFits}
\title{compare_IndPredFits}
\usage{
compare_IndPredFits(
  modelFolder,
  outputFolder,
  stratify = "TRTNAME",
  fileName = "DataPredComparison",
  logY = FALSE,
  figLimX = NULL,
  figLimY = NULL,
  predToPlot = NULL,
  withVehicle = FALSE,
  ActivityPath = NULL
)
}
\arguments{
\item{modelFolder}{Path to IQRNLME project containing data and model
predictions}

\item{outputFolder}{Character string denoting path to folder in which to save
output}

\item{stratify}{Character string defining column name on which to stratify
the plots. Must be one of "USUBJID", "ID", "TRTNAME", "STUDY".}

\item{fileName}{Character string denoting name with which to save output.
Default: 'DataPredComparison'}

\item{logY}{Logical: log 10 y axis on plot. Default: \code{FALSE}}

\item{figLimX}{Numeric, xlim to add to plot with \code{ggplot2::coord_cartesian}.
Default: \code{NULL}}

\item{figLimY}{Numeric, ylim to add to plot with \code{ggplot2::coord_cartesian}.
Default: \code{NULL}}

\item{predToPlot}{Which prediction to plot (POP, XPRED, INDIV or IPRED);
default: \code{NULL} in which case the function will plot both population and
individual predictions.}

\item{withVehicle}{Logical: Include Vehicle? Default: \code{FALSE}}

\item{ActivityPath}{Default: \code{NULL}}
}
\value{
Nothing is returned directly by the function, visualizations of
model comparisons are saved as .pdf to the specific output folder
}
\description{
Compare model predictions with observations and save
visualizations to pdf
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Aline Fuchs (MMV), Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
