% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_PredFitsSYS}
\alias{compare_PredFitsSYS}
\title{compare_PredFitsSYS}
\usage{
compare_PredFitsSYS(
  fit,
  pathname = "ComparePredObs",
  logY = FALSE,
  stratify = "CONDITION",
  data = NULL,
  modelFile = NULL,
  FLAGrefit = FALSE
)
}
\arguments{
\item{fit}{SysFit model to compare with data, expects object generated by "run_SysFitEstimation"}

\item{pathname}{Character string denoting folder name for output graphs Default: 'ComparePredObs'}

\item{logY}{Logical, flag whether to plot y on log-transformed scale. Default: \code{FALSE}}

\item{stratify}{Character string defining column to stratify panels by. Default: 'CONDITION'}

\item{data}{IQRdataGENERAL object or character string defining file path to IQRdataGENERAL.csv file, containing
data to compare simulations with. If NULL modeling data will be used.Default: \code{NULL}}

\item{modelFile}{Character string defining filepath of IQRmodel. Default: \code{NULL}}

\item{FLAGrefit}{Logical, FLAG whether to use pre-calculated simulations or to regenerate simulations. Default: \code{FALSE}}
}
\value{
Nothing is returned directly by the function, visualizations of
model comparisons are saved as .pdf to \code{pathname}
}
\description{
Compare predictions from a sysfit model to observations.
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
