% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{ConfusionMatrix}
\usage{
ConfusionMatrix(
  data,
  colPred,
  colRef,
  positiveClass,
  negativeClass = paste0("No ", positiveClass)
)
}
\arguments{
\item{data}{a data.frame with at least two columns of logical values named as the arguments colPred and colRef.}

\item{colPred, colRef}{character strings denoting names of logical (TRUE/FALSE) columns in data.}

\item{positiveClass}{character string denoting the human readable meaning of TRUE values in data[\link{colPred}] and data[\link{colRef}].}

\item{negativeClass}{character string (default paste0("No ", positiveClass)), denoting the human readable meaning of FALSE
values in data[\link{colPred}] and data[\link{colRef}].}
}
\value{
a named list as described in documentation of \verb{[caret::confusionMatrix()]}.
}
\description{
Calculate confusion matrix and derived statistics from a predicted and reference binary data
}
\note{
This function is using \verb{[caret::confusionMatrix()]}
}
