% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-OldFunctions.R
\name{convertGametocytes}
\alias{convertGametocytes}
\title{convertGametocytes}
\usage{
convertGametocytes(
  data,
  gamNameOld = "Parasitemia female gametocytes",
  gamNameNew = "Parasitemia female gametocytes",
  oldStdCurve,
  convertu2m,
  elutionFactor = 1,
  femaleGamFactor = 590,
  oldNewStdFactor = 62
)
}
\arguments{
\item{gamNameOld}{Default: 'Parasitemia female gametocytes'}

\item{gamNameNew}{Default: 'Parasitemia female gametocytes'}

\item{elutionFactor}{Default: 1}

\item{femaleGamFactor}{Default: 590}

\item{oldNewStdFactor}{Default: 62}
}
\description{
convertGametocytes
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{list_Center}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()},
\code{\link{summarizeReplicates}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Aline Fuchs (MMV)
}
\concept{Data Preparation}
