% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{create_PKPDsimtime}
\alias{create_PKPDsimtime}
\title{Generate Simulation Times}
\usage{
create_PKPDsimtime(Tend, dt = c(0.25, 1, 2, 6), Tswitch = c(0, 12, 24, 48))
}
\arguments{
\item{Tend}{Final time}

\item{dt}{Time steps (Defaul: \code{c(  0.25,  1,  2, 6)}).}

\item{Tswitch}{Time intervals over which time steps are used. It should be a vector of similar length as \code{dt} (Default: \code{c(0,  12, 24, 48)}).}
}
\value{
Numeric vector with the calcultaed time point
}
\description{
Calculate time points using given time steps between specifed time point. This is useful
to optimize time step for ploting. For example with PK profile the time step is often needed
to be small during absorption and large during the elimination phase.
}
\details{
Calculate timepoints from 0 to Tend separated by regular time steps sepcified in dt. Several time steps can be specified: they will be applied
for the respective time intervals specified in the vector Tswitch.
}
\examples{
create_PKPDsimtime(Tend=1,dt=c(0.1,0.5),Tswitch=c(0,0.4))

}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{add_FuncToEnv}()},
\code{\link{adjust_Dose}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{cbindMMV}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{combMMV}()},
\code{\link{convert_Unit}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{getDoselevel}()},
\code{\link{getMKLthreadsMMV}()},
\code{\link{get_IQRdistribution}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{get_MMVmalariaPath}()},
\code{\link{libraryMMV}()},
\code{\link{load_RData}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rbindMMV}()},
\code{\link{rectintMMV}()},
\code{\link{reverse_List}()},
\code{\link{saveActivityInfo}()},
\code{\link{saveMMV}()},
\code{\link{setMKLthreadsMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
