% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{createDataVPC_IQRnlmeProject_MMV}
\alias{createDataVPC_IQRnlmeProject_MMV}
\title{Create data for VPC}
\usage{
createDataVPC_IQRnlmeProject_MMV(
  project = NULL,
  dataVPC = NULL,
  Ntrials,
  gpf = NULL,
  dosingInfo = NULL,
  regressors = NULL,
  model = NULL,
  modelsSample = NULL,
  FLAGpreparePC = FALSE,
  simtimeOption = c("obs", "simTFD", "simTAD"),
  simtimeN = 200,
  simtimes = NULL,
  simtimeMax = NULL,
  FLAGindivSimtimeMax = FALSE,
  FLAGaddResidualNoise = TRUE,
  FLAGsampleUncertainty = TRUE,
  FLAGlogOutput = FALSE,
  factorMult = NULL,
  ncores = 1,
  seed = NULL,
  FLAGkeepParameters = FALSE,
  opt_initstep = 0,
  opt_abstol = 1e-06,
  opt_reltol = 1e-06,
  opt_maxstep = 0,
  opt_minstep = 0,
  par_lower = NULL,
  par_upper = NULL,
  FLAGfixDoseOverlap = TRUE,
  FLAGstopFailedSim = FALSE
)
}
\arguments{
\item{project}{Path to IQRnlmeProject to be used to generate the VPC. The stuctural model
of this project will be used and parameters will be sampled from this project.}

\item{dataVPC}{Dataset for VPC generation. If undefined, the modeling dataset from
modelSampleSimulate will be used. Mandatory if sampling from GPF.}

\item{Ntrials}{Number of trials to simulate for the VPC}

\item{gpf}{GPF object or path to GPF as xlsx file.}

\item{dosingInfo}{Required for information on lagtimes and zero order absorption when no project is provided.
A list with one field per input in the model. Fields need to be named according to the input names
in the model. Each field is a named vector with the required name "type" and the optional names "Tlag" and
"Tk0". "type" can be "BOLUS", "INFUSION", or "ABSORPTION0". Only difference between INFUSION and
ABSORPTION0 is that in the first case the infusion RATE is provided in the data and in the second the
absorption time is sampled. Lag times can be added to each dosing input by defining "Tlag" as vector
element. The value of this element needs to be the parameter named for this lag time. If type="ABSORPTION0"
then also "Tk0" needs to be present as element in the vector. The value is a parameter name that also needs
to appear in the model. For examples see dosing_IQRest}

\item{regressors}{Required for information on regressors if no project is given. Input is a vector with the names of
the regressors taken from the dataset.}

\item{model}{Required structural model (IQRmodel) if no project is given}

\item{modelsSample}{If sequential model building has been done (e.g. PKPD) model
parameters might need to be sampled from additional IQRnlmeProjects. modelSample
can be a vector with paths to these additional IQRnlmeProjects. It is not allowed that
in these projects and modelSampleSimulate estimated parameter names overlap}

\item{FLAGpreparePC}{Flag whether to prepare prediction correction by corresponding additional simulations}

\item{simtimeOption}{Select for which times, outputs are simulated: 'obs': respective indivdual observation time,
'simTFD': Equidistant or user-profided times after first dose, or 'simTAD': equidistant or
user-provided times after last dose (these will be repeated for each dose)}

\item{simtimeN}{Define number of simulation time steps to add between smallest and largest
event time in the dataset. Equidistant sampling will be used.}

\item{simtimes}{Vector of simulation times. If not NULL, this vector of simulation times overrules simtimeN}

\item{simtimeMax}{Define maximum simulation time for VPC for all subjects.}

\item{FLAGindivSimtimeMax}{If TRUE then the simulation time of a simulated subject will
be limited to the max event time for this subject in the dataset.
A subject here basically defines the dosing schedule and the covariates.
If set to FALSE then the simulation results will not be truncated.}

\item{FLAGaddResidualNoise}{logical. If TRUE then residual noise will be added
to the simulations}

\item{FLAGsampleUncertainty}{logical. If set to TRUE then for each new trial
new population parameters are drawn from the uncertainty distribution}

\item{FLAGlogOutput}{Logical. If set to TRUE then the model outputs are assumed to be
log transformed data and will be back-transformed. The data in dataVPC are assumed to be
in the normal domain - to simplify dose-normalization. If FALSE then data and model outputs
are assumed to be in the same domain.}

\item{factorMult}{named vector allowing to define multiplicative factors to modulate
by name selected sampled parameters for simulation}

\item{ncores}{Number of cores for parallelization.}

\item{seed}{Set seed for reproducible sampling. If the simulations are parallelized, the seed will be
set for the simulation of each population to seed + population number - 1.}

\item{FLAGkeepParameters}{Logical. If set to TRUE, sampled parameters will be added to
returned simulation results. Defaults to FALSE}

\item{opt_initstep}{Double value for initial step-size to be attempted for simulation}

\item{opt_abstol}{Double value for absolute tolerance for simulation}

\item{opt_reltol}{Double value for relative tolerance for simulation}

\item{opt_maxstep}{Double value for maximal integrator step-size for simulation}

\item{opt_minstep}{Double value for minimal integrator step-size for simulation}

\item{par_lower}{Named vector of lower limits for sampled individual parameters.
Applies to individual parameters after potential modulation by factorMult argument.}

\item{par_upper}{Named vector of upper limits for sampled individual parameters.
Applies to individual parameters after potential modulation by factorMult argument.}

\item{FLAGfixDoseOverlap}{Input argument for IQReventTable. If TRUE (default) then ensure
non-overlapping dosing intervals if these occur by re-defining the dosing events.
If FALSE in case of overlapping dose intervals, error is issued. Please inspect correct
behavior. If correction does not work, par_upper could be an alternative to avoid overlapping
dosing events.}

\item{FLAGstopFailedSim}{Logical. If TRUE execution will be stopped if simulation fails for any
of the repeated trials. Defaults to FALSE. Execution will stop anyways if all simulations fail.}
}
\value{
List with data frames \emph{sim} and \emph{obs}. If not prediction correction preparation
is required, \emph{sim} contains individual predictions and \emph{obs} the observation data,
including dosing records. Otherwise \emph{sim} and \emph{obs} contains typical predictions
in addition.
}
\description{
Simulates trial data for producing VPCs
}
\details{
Model from IQR nlme project is used to simulate trial population defined either in contained dataset
or given dataset as dataVPC.
It will also perform additional simulations to performed prediction correction for the VPCs if required.
For the observations, a typical prediction will also be performed using the population
parameter point estimates.
The VPC can subsequently e plotted using \link{plotVPC_IQRdataVPC}.
}
\concept{IQRnlmeProject}
\concept{VPC}
