% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-IQRaux.R
\name{createSimIQRmodelSpec}
\alias{createSimIQRmodelSpec}
\title{Collect and standardize input arguments for sim_IQRmodel}
\usage{
createSimIQRmodelSpec(
  model,
  simtime,
  IC = NULL,
  parameters = NULL,
  dosingTable = NULL,
  eventTable = NULL,
  FLAGsensitivity = FALSE,
  FLAGuseSensEq = TRUE,
  sensParams = NULL,
  FLAGoutputsOnly = FALSE,
  opt_eventTimes = NULL,
  opt_method_stiff = TRUE,
  opt_abstol = 1e-06,
  opt_reltol = 1e-06,
  opt_minstep = 0,
  opt_maxstep = 0,
  opt_initstep = 0,
  opt_maxnumsteps = 1e+05,
  opt_maxerrtestfails = 50,
  opt_maxorder_stiff = 5,
  opt_maxorder_nonstiff = 12,
  opt_maxconvfails = 10,
  opt_maxnonlineariter = 3,
  opt_usesymjac = TRUE,
  opt_sens_simultaneous = FALSE,
  opt_sens_errcon = FALSE,
  opt_sens_maxnonlineariter = 3,
  verbose = FALSE,
  FLAGprogressBar = FALSE,
  opt_timeHandling = c("augmentEventTable", "warn", "stop"),
  mc.cores = 1
)
}
\arguments{
\item{model}{See \code{\link[IQRtools:sim_IQRmodel]{IQRtools::sim_IQRmodel()}}, also for other arguments}

\item{opt_timeHandling}{sanitize times in an eventTable (see description)}

\item{mc.cores}{For parallelization with parallel::mclapply}
}
\value{
List of its input arguments, with class simIQRmodelSpec
}
\description{
This function does four things
}
\details{
\itemize{
\item Collect input arguments differing from the default values for \code{\link[IQRtools:sim_IQRmodel]{IQRtools::sim_IQRmodel()}} in a list
\item Require that \code{simtime} is specified explicitly
\item Replace \code{parameters} and \code{dosingTable} inputs by \code{eventTable}
\item If min(simtime) < min(eventTable$TIME), add a dummy event at min(simtime), so no simtime is cut off ever.
}
}
\examples{
library(IQRtools)
library(ggplot2)
fileURL <- "http://192.168.1.12:7804/IQauxfn/inst/02-ExampleModels/ModelLibraryPK/model_2cpt_linsat_iv.txt"
fl <- file.path(tempdir(), basename(fileURL))
download.file(fileURL, fl)

model <- IQRmodel(fl)
simIQRmodelSpec1 <- createSimIQRmodelSpec(model, simtime = c(0,1), parameters =  c(CL = 10), dosingTable = IQRdosing(0,1,1))
do.call(sim_IQRmodel, simIQRmodelSpec1)
}
\seealso{
Other Simulation: 
\code{\link{create_GDF_fromDesignSpec}()},
\code{\link{getDefaultEventTable}()},
\code{\link{populateGDFWithSimValues_consistent}()},
\code{\link{sim_IQRmodelConsistent}()}
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
\concept{Simulation}
