% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{define_tLLOQ}
\alias{define_tLLOQ}
\title{define_tLLOQ
Defines the time when parasite clearance or growth crosses the LLOQ}
\usage{
define_tLLOQ(xx, LLOQ)
}
\arguments{
\item{xx}{data frame with parasitemia data}

\item{LLOQ}{LLOQ for parasite detection (log scale)}
}
\value{
input data frame amended by row indicating the time at which LLOQ is reached
and Phase column indicating whether this is defining the lower or upper bound for the MIC.
(MICmin for re-growth phase and MICmax for clearance phase)
}
\description{
Determines the time when the LLOQ is reached based on linear segments
fitted to clearance phase (output from \link{fit_LinearSegments}) or re-growth
phase (output from \link{fit_LinearModel}). Needed to determine the time range within
which the MIC is reached assuming that it needs to in the time
when parasites are below LLOQ
}
\details{
Input needs to contain Phase column indicating whether data is from
parasite clearance phase (Phase = "KILL") or from re-growth (Phase = "RECRUD").
Segment knots are defined by Xknot and Yknot columns
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
