% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{display_ComboTreatments}
\alias{display_ComboTreatments}
\title{display_ComboTreatments}
\usage{
display_ComboTreatments(
  data,
  stratify = "STUDY",
  fileplot = NULL,
  filetable = NULL,
  CompoundList = NULL,
  FLAGreturnObject = FALSE
)
}
\arguments{
\item{stratify}{Default: 'STUDY'}

\item{fileplot}{Default: NULL}

\item{filetable}{Default: NULL}

\item{CompoundList}{Default: NULL}

\item{FLAGreturnObject}{Default: FALSE}
}
\value{

}
\description{
display_ComboTreatments
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
