% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{evaluate_APR}
\alias{evaluate_APR}
\title{Evaluate APR}
\usage{
evaluate_APR(
  dataSim,
  LLOQ,
  timeCOL = "TIME",
  paraCOL = "PL",
  weightCOL = "WT0",
  curethreshCOL = "CureThreshold",
  aprTime = 28,
  defaultWT = 55,
  extrapTol = 48,
  Plog = TRUE,
  FLAGtruncateTime = TRUE,
  FLAGextrapolateTime = FALSE,
  FLAGverbose = FALSE
)
}
\arguments{
\item{dataSim}{data.table produced by `simulate_virtualtrials()`, specifically `$simPKPD`. `WT0` and `curethreshCOL` should also be included in this data.table if desired}

\item{LLOQ}{LLOQ parasitemia in the same scale as in `paraCOL` (i.e. linear or ln - numeric).}

\item{weightCOL}{optional column name of `dataSim` containing body weight (Default: `WT0`).}

\item{curethreshCOL}{optional column name of `dataSim` containing cure threshold (Default: "Cure_Thresh_col")}

\item{aprTime}{day that denotes final day of follow-up - metrics will be calculated based on this day}

\item{defaultWT}{default weight (kg) to be assumed if neither cure threshold or patient weight are provided (Default: 55)}

\item{extrapTol}{Numeric indicating the maximum allowed tolerance in hours for the interpolation/extrapolation in case
of last measurement not exactly at `aprTime`*24hr (Default: 48hr).}

\item{Plog}{Indicate if parasitemia is in the log or linear scale (Default: `TRUE` which means it is logged).}

\item{FLAGtruncateTime}{A logical indicating if the PL measurements should be truncated at `aprTime`*24hr,
for subjects with measurements later than `aprTime`*24hr.
If `TRUE`, all measurements after `aprTime`*24hr+`extrapTol` are removed (Default: `TRUE`).}

\item{FLAGextrapolateTime}{A logical indicating if the PL measurements should be extrapolated up to `aprTime`*24hr,
for subjects which measurements end before `aprTime`*24hr. If `TRUE` the last measured value is repeated up to
`aprTime`*24hr (Default: `FALSE`).}

\item{FLAGverbose}{Logical (Default: `FALSE`).}
}
\value{
A data.frame of a single row and columns USUBJID and binary (0 or 1) columns as follows:
\itemize{
\item USUBJID
\item TrialID
\item DOSE
\item PCloq: Time to first drop below LLOQ
\item Tcure: Time at which cure threshold is reached
\item P2BL : Presence of parasitaemia higher than baseline at 48 hours
\item P3BL : Presence of parasitaemia higher than 25% of baseline at 72 hours
\item P3 :  Presence of parasitaemia higher than LLOQ at 72 hours
\item ETF_CSon: Parasitaemia on 48h higher than baseline or parasitaemia > LLOQ on day 3 or parasitaemia > 25% baseline on day 3
\item ETF_CSoff : Parasitaemia on 48h higher than baseline or parasitaemia >25% baseline on 48h
\item simIPFabs : Absolute presence of parasites >LLOQ between day 4 and 7, irrespective of any other criteria being met
\item simLPFabs : Absolute presence of parasites >LLOQ between day 7 and aprTime, irrespective of any other criteria being met
\item LCFabs_CSon : Absolute presence of parasites >LLOQ between day 4 and aprTime, irrespective of any other criteria being met
\item LCF_CSon : Presence of parasitaemia between day 4 and aprTime, with respect to ETF_CSon
\item LPF_CSon : Presence of parasitaemia between day 7 and aprTime, with respect to ETF_CSon and LCF_CSon
\item LPF_CSoff : Presence of parasitaemia between day 7 and aprTime, with respect to ETF_CSoff
\item APR_CSon : Absence of parasitaemia on aprTime, if the patient did not meet criteria of ETF_CSon, LCF_CSon or LPF_CSon
\item APR_CSoff : Absence of parasitaemia on aprTime, if the patient did not meet criteria of ETF_CSoff or LPF_CSoff
\item APRinf_CSon : Attainment of the cure threshold within aprTime, in the absence of ETF_CSon
\item APRinf_CSoff : Attainment of the cure threshold within aprTime, in the absence of ETF_CSoff
}
}
\description{
Evaluate APR at Day `aprTime` given an individual time-course of parasitemia measurements,
baseline parasitemia, LLOQ parasitemia and weight at TIME 0. Two APRs are return, an optimistic one (CSoff - clinical symptoms off),
which assumes that no patient has fever, and a pessimistic (CSon - clinical symptoms on) one, which assumes that all patients have fever.
Note that times up to 7 days are hard-coded in and the function will crash (currently with no error message)
if shorter trials are evaluated
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Sam Jones (MMV), Venelin Mitov (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
