% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{evaluate_APR28}
\alias{evaluate_APR28}
\title{Evaluate APR at Day 28}
\usage{
evaluate_APR28(
  dataSim,
  LLOQ,
  timeCOL = "TIME",
  paraCOL = "PL",
  weightCOL = "WT0",
  curethreshCOL = "CureThreshold",
  defaultWT = 55,
  extrapTol = 48,
  Plog = TRUE,
  FLAGtruncateTime = TRUE,
  FLAGextrapolateTime = FALSE,
  FLAGverbose = FALSE
)
}
\arguments{
\item{dataSim}{data.table produced by \code{simulate_virtualtrials()}, specifically \verb{$simPKPD}. \code{WT0} and \code{curethreshCOL} should also be included in this data.table if desired}

\item{LLOQ}{LLOQ parasitemia in the same scale as in \code{paraCOL} (i.e. linear or ln - numeric).}

\item{weightCOL}{optional column name of \code{dataSim} containing body weight (Default: \code{WT0}).}

\item{curethreshCOL}{optional column name of \code{dataSim} containing cure threshold (Default: "Cure_Thresh_col")}

\item{defaultWT}{default weight (kg) to be assumed if neither cure threshold or patient weight are provided (Default: 55)}

\item{extrapTol}{Numeric indicating the maximum allowed tolerance in hours for the interpolation/extrapolation in case
of last measurement not exactly at 28*24hr (Default: 48hr).}

\item{Plog}{Indicate if parasitemia is in the log or linear scale (Default: \code{TRUE} which means it is logged).}

\item{FLAGtruncateTime}{A logical indicating if the PL measurements should be truncated at 28\emph{24hr,
for subjects with measurements later than 28}24hr.
If \code{TRUE}, all measurements after 28*24hr+\code{extrapTol} are removed (Default: \code{TRUE}).}

\item{FLAGextrapolateTime}{A logical indicating if the PL measurements should be extrapolated up to 28\emph{24hr,
for subjects which measurements end before 28}24hr. If \code{TRUE} the last measured value is repeated up to
28*24hr (Default: \code{FALSE}).}

\item{FLAGverbose}{Logical (Default: \code{FALSE}).}

\item{aprTime}{day that denotes final day of follow-up - metrics will be calculated based on this day}
}
\value{
A data.frame of a single row and columns USUBJID and binary (0 or 1) columns as follows:
\itemize{
\item USUBJID
\item TrialID
\item DOSE
\item PCloq: Time to first drop below LLOQ
\item Tcure: Time at which cure threshold is reached
\item P2BL : Presence of parasitaemia higher than baseline at 48 hours
\item P3BL : Presence of parasitaemia higher than 25\% of baseline at 72 hours
\item P3 :  Presence of parasitaemia higher than LLOQ at 72 hours
\item ETF_CSon : Parasitaemia on 48h higher than baseline or parasitaemia > LLOQ on day 3 or parasitaemia > 25\% baseline on day 3
\item ETF_CSoff : Parasitaemia on 48h higher than baseline or parasitaemia >25\% baseline on 48h
\item simIPFabs : Absolute presence of parasites >LLOQ between day 4 and 7, irrespective of any other criteria being met
\item simLPF28abs : Absolute presence of parasites >LLOQ between day 7 and 28, irrespective of any other criteria being met
\item LCF28abs_CSon : Absolute presence of parasites >LLOQ between day 4 and 28, irrespective of any other criteria being met
\item LCF28_CSon : Presence of parasitaemia between day 4 and 28, with respect to ETF_CSon
\item LPF28_CSon : Presence of parasitaemia between day 7 and 28, with respect to ETF_CSon and LCF28_CSon
\item LPF28_CSoff : Presence of parasitaemia between day 7 and 28, with respect to ETF_CSoff
\item APR28_CSon : Absence of parasitaemia on day 28, if the patient did not meet criteria of ETF_CSon, LCF28_CSon or LPF28_CSon
\item APR28_CSoff : Absence of parasitaemia on day 28, if the patient did not meet criteria of ETF_CSoff or LPF28_CSoff
}
}
\description{
Evaluate APR at Day 28 given an individual time-course of parasitemia measurements,
baseline parasitemia, LLOQ parasitemia and weight at TIME 0. Two APRs are return, an Optimistic (CSoff - clinical symptoms off) one,
which assumes that no patient has fever, and a pessmistic (CSon - clinical symptons on) one, which assumes that all patients have fever.
Note that times up to 7 days are hard-coded in and the function will crash (currently with no error message)
if shorter trials are evaluated
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Sam Jones (MMV), Venelin Mitov (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
