% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-Chemoprotection.R
\name{evaluate_BreakthroughEvent}
\alias{evaluate_BreakthroughEvent}
\title{Evaluate breakthrough event,
given an individual time-course of parasitemia measurements and LLOQ parasitemia.}
\usage{
evaluate_BreakthroughEvent(
  dataSim,
  LLOQ,
  timeCOL = "TIME",
  paraCOL = "PL",
  Plog = TRUE,
  FLAGinterpolateTime = FALSE
)
}
\arguments{
\item{dataSim}{data.table produced by `simulate_virtualtrials()` or `sim_IQRmodel()`, specifically `$simPKPD`.}

\item{LLOQ}{LLOQ parasitemia in the same scale as in `paraCOL` (i.e. linear or ln - numeric).}

\item{timeCOL}{column name of `dataSim` containing time records (Default: `TIME`).}

\item{paraCOL}{column name of `dataSim` containing parasitemia records (Default: `PL`).}

\item{Plog}{Indicate if parasitemia is in the log or linear scale (Default: `TRUE` which means it is logged).}

\item{FLAGinterpolateTime}{A logical indicating if the PL measurements should be interpolated within the simulation period.}
}
\value{
A data.frame with columns as follows:
\itemize{
\item USUBJID
\item TrialID
\item NAME: name of the event evaluated "Breakthrough"
\item TIME: Time to first exceed above LLOQ
\item VALUE: TRUE or FALSE statement regarding breakthrough event
}
}
\description{
Evaluate whether a breakthrough event has been observed and estimate the time of first occurrence.
If no event has been observed, either because it did not occur, or if the
measurement data for the subject is missing, the time of the last measurement
is reported to indicate that the subject is not anymore in the population after
that moment in time.
}
\details{
A breakthrough event is observed if PL exceeds LLOQ. We report time as the first time above LLOQ.
Trivially, a breakthrough event IS NOT observed if PL did not exceed LLOQ. We report time as the
time of last measurement.
}
\examples{

}
\seealso{
Other Chemoprotection: 
\code{\link{generate_ChemoModel_FluxLimiter}()},
\code{\link{generate_ChemoModel_UpwindFirstOrder}()}
}
\author{
Catalina Barcelo (MMV, \email{barceloc@mmv.org})
}
\concept{Chemoprotection}
