% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-PredictHumanDoses.R
\name{evaluateDoseCriterion_PRRtot}
\alias{evaluateDoseCriterion_PRRtot}
\title{Calculate PRRtot for an individual parasitemia profile}
\usage{
evaluateDoseCriterion_PRRtot(sim_results, parameters)
}
\arguments{
\item{sim_results}{a data.frame returned by \code{\link[IQRtools:sim_IQRmodel]{IQRtools::sim_IQRmodel()}}. It is assumed that
this data.frame has a column PL corresponding to log-transformed simulated parasitemia.}

\item{parameters}{currently not used.}
}
\value{
a numeric value.
}
\description{
Calculate PRRtot for an individual parasitemia profile
}
\details{
This function can be passed as argument to \code{\link[=predictDose_Generic]{predictDose_Generic()}}.
}
\author{
Venelin Mitov, IntiQuan
}
