% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-PredictHumanDoses.R
\name{evaluateDoseCriterion_TimeAboveMIC}
\alias{evaluateDoseCriterion_TimeAboveMIC}
\title{Calculate the time above the minimum inhibitory concentration (MIC) for an individual parasitemia profile}
\usage{
evaluateDoseCriterion_TimeAboveMIC(sim_results, parameters)
}
\arguments{
\item{sim_results}{a data.frame returned by \code{\link{IQRtools::sim_IQRmodel}}. It is assumed that
this data.frame has a column PL corresponding to log-transformed simulated parasitemia.}

\item{parameters}{a named numeric vector.}
}
\value{
a numeric value.
}
\description{
Calculate the time above the minimum inhibitory concentration (MIC) for an individual parasitemia profile
}
\details{
This function can be passed as argument to \code{\link{predictDose_Generic}}.
}
\seealso{
\code{\link{getKeysEMAX}}
}
\author{
Venelin Mitov, IntiQuan
}
