% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-PredictHumanDoses.R
\name{evaluateDoseCriterion_TimeAboveMPC90}
\alias{evaluateDoseCriterion_TimeAboveMPC90}
\title{Calculate the time above the minimum parasiticidal concentration (MPC90) for an individual parasitemia profile}
\usage{
evaluateDoseCriterion_TimeAboveMPC90(sim_results, parameters)
}
\arguments{
\item{sim_results}{a data.frame returned by \code{\link[IQRtools:sim_IQRmodel]{IQRtools::sim_IQRmodel()}}. It is assumed that
this data.frame has a column PL corresponding to log-transformed simulated parasitemia.}

\item{parameters}{a named numeric vector.}
}
\value{
a numeric value.
}
\description{
Calculate the time above the minimum parasiticidal concentration (MPC90) for an individual parasitemia profile
}
\details{
This function can be passed as argument to \code{\link[=predictDose_Generic]{predictDose_Generic()}}.
}
\seealso{
\code{\link[=getKeysEMAX]{getKeysEMAX()}}
}
\author{
Mohammed H. Cherkaoui, MMV
}
