% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{fit_LinearModel}
\alias{fit_LinearModel}
\title{fit_LinearModel
Fits line to parasitemia data on log scale}
\usage{
fit_LinearModel(x, FLAGlogtransY = TRUE)
}
\arguments{
\item{x}{data frame with parasitemia records}

\item{FLAGlogtransY}{Flag whether to need to be log-transformed
(TRUE (default) in case data are provided on normal scale)}
}
\value{
data frame with locations of the knots (Xknot and Yknot), number of knots (Nknots),
the sum of squares of the fit (SSresid) and fit information (fitinfo)
}
\description{
Fits a line using \code{lm} to parasitemia data (TIME and VALUE columns).
}
\details{
This function is used in the model-free assessment of PD data by describing the
PD data by linear segments. The linear segments to describe the PD data
are characterized by knots which are the endpoint of the single segments.
This function is used to fit single line to growth data.
The output has the same format as fit_LinearSegments and gives the number
of knots, here 2, and the position of the knots, the sum of squares
for the fit and some status of the fit.
Knots are given by Xknot (time scale) and Yknot (logtransformed PD scale).
Uses only non-censored data.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
