% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{fit_LinearSegments}
\alias{fit_LinearSegments}
\title{fit_LinearSegments
Fits line to parasitemia data on log scale}
\usage{
fit_LinearSegments(
  data,
  constraint = "none",
  Nsegments = 3,
  FLAGlogtransY = TRUE
)
}
\arguments{
\item{data}{data frame with parasitemia records}

\item{constraint}{character (string) specifying the kind of constraint
(recommended: "none" or "decrease"). Decrease will only allow negatve slopes for segments.}

\item{Nsegments}{maximum number of knots, defaults to 3}

\item{FLAGlogtransY}{Flag whether to need to be log-transformed
(TRUE (default) in case data are provided on normal scale)}
}
\value{
data frame with locations of the knots (Xknot and Yknot), number of knots (Nknots),
        the sum of squares of the fit (SSresid) and fit information (fitinfo)
}
\description{
Fits linear segments to parasitemia data (TIME and VALUE columns)
using `cobs` function from the cobs package.
}
\details{
This function is used in the model-free assessment of PD data by describing the
PD data by linear segments. The linear segments to describe the PD data
are characterized by knots which are the endpoint of the single segments.
This function is used to data in the clearance phase.
The output has the same format as fit_LinearModel and gives the number
of knots, the position of the knots, the sum of squares
for the fit, and some status of the fit.
If there is only one observation, no fit is done. If there are two observations,
the line through these points is taken. More than two observations are fitted
using the cobs function (constrained linear splines). If cobs fit fails a line
is fitted trough first and last timepoint.
Knots are given by Xknot (time scale) and Yknot (logtransformed PD scale).
Uses only non-censored data.
}
\seealso{
[cobs::cobs]

Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
