% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{formula_SCIDrbp}
\alias{formula_SCIDrbp}
\title{Calculate ratio blood plasma in SCID}
\usage{
formula_SCIDrbp(
  HuRbp,
  HuPPB,
  HuH = 0.45,
  MouseRbp,
  MousePPB,
  MouseH = 0.45,
  scidH = 0.8,
  huRBCinSCID = 0.6
)
}
\arguments{
\item{HuRbp}{Human ratio blood plasma}

\item{HuPPB}{Human plasma protein binding}

\item{HuH}{Human hematocrit}

\item{MouseRbp}{Mouse ratio blood plasma}

\item{MousePPB}{Mouse plasma protein binding}

\item{MouseH}{Mouse hematocrit}

\item{scidH}{SCID hematocrit}

\item{huRBCinSCID}{Human red blood cells in SCID}
}
\value{
numeric, SCID ratio blood plasma
}
\description{
Calculate ratio blood plasma in SCID
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{Key Parameters}
