% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{formulaPKpar_MacroToPrimary}
\alias{formulaPKpar_MacroToPrimary}
\title{formulaPKpar_MacroToPrimary}
\usage{
formulaPKpar_MacroToPrimary(
  PKparameters_Macro,
  PKmacroName = c(Fabs0 = "Fabs0", Tk0 = "Tk0", Fabs1 = "Fabs1", ka = "ka", Tlag1 =
    "Tlag1", Alpha = "Alpha", Beta = "Beta", Gamma = "Gamma", A = "A", B = "B", C = "C"),
  PKprimaryName = c(CL = "CL", Vc = "Vc", Q1 = "Q1", Vp1 = "Vp1", Q2 = "Q2", Vp2 =
    "Vp2")
)
}
\arguments{
\item{PKmacroName}{Default: c(Fabs0 = "Fabs0", Tk0 = "Tk0", Fabs1 = "Fabs1", ka = "ka", Tlag1 = "Tlag1",
Alpha = "Alpha", Beta = "Beta", Gamma = "Gamma", A = "A",
B = "B", C = "C")}

\item{PKprimaryName}{Default: c(CL = "CL", Vc = "Vc", Q1 = "Q1", Vp1 = "Vp1", Q2 = "Q2", Vp2 = "Vp2")}
}
\value{

}
\description{
formulaPKpar_MacroToPrimary
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR28}()},
\code{\link{evaluate_APR}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
