% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{formulaPKpar_PrimaryToMacro}
\alias{formulaPKpar_PrimaryToMacro}
\title{formulaPKpar_PrimaryToMacro}
\usage{
formulaPKpar_PrimaryToMacro(
  PKparameters_Primary,
  PKprimaryName = c(Fabs0 = "Fabs0", Tk0 = "Tk0", Fabs1 = "Fabs1", ka = "ka", Tlag1 =
    "Tlag1", CL = "CL", Vc = "Vc", Q1 = "Q1", Vp1 = "Vp1", Q2 = "Q2", Vp2 = "Vp2"),
  PKmacroName = c(Alpha = "Alpha", Beta = "Beta", Gamma = "Gamma", A = "A", B = "B", C =
    "C")
)
}
\arguments{
\item{PKprimaryName}{Default: c(Fabs0 = "Fabs0", Tk0 = "Tk0", Fabs1 = "Fabs1", ka = "ka", Tlag1 = "Tlag1",
CL = "CL", Vc = "Vc", Q1 = "Q1", Vp1 = "Vp1", Q2 = "Q2",
Vp2 = "Vp2")}

\item{PKmacroName}{Default: c(Alpha = "Alpha", Beta = "Beta", Gamma = "Gamma", A = "A", B = "B",
C = "C")}
}
\description{
formulaPKpar_PrimaryToMacro
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
