% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{forwardSelection}
\alias{forwardSelection}
\title{forwardSelection}
\usage{
forwardSelection(
  model,
  dosing,
  data,
  modelSpec,
  covariateToTest,
  COVcentering = NULL,
  projectPath = "PKmodels",
  alpha = 0.01,
  tool = "MONOLIX",
  toolVersion = NULL,
  ncores = 1,
  Nparallel = 1,
  setting = NULL
)
}
\arguments{
\item{model}{An IQRmodel object containing the default model which covariates
will be added to in the forward selection process}

\item{dosing}{IQRdosing object to be used for estimation, see\code{ ?IQRnlmeEst} and \code{?IQRdosing}}

\item{data}{data frame or character string defining file path to data, see \code{?IQRnlmeEst}}

\item{modelSpec}{A list containing the IQR model spec, see \code{?IQRnlmeEst}}

\item{covariateToTest}{A list containing covariates to be tested in the forward
selection process, can be a named list with parameter = covariate. If the
element of the list has no name, it is assumed that each element will be
composed of c("Parameter","Covariate")}

\item{COVcentering}{Named vector defining centering value for covariates i.e. COVcentering = c(WT0=55)}

\item{projectPath}{Path to the project folder, default: "PKmodels", covariate models will be written to disk here}

\item{alpha}{Default: 0.01}

\item{tool}{Default: 'MONOLIX'}

\item{toolVersion}{Default: \code{NULL}}

\item{ncores}{Default: 1}

\item{Nparallel}{Default: 1}

\item{setting}{Default: \code{NULL}}
}
\value{
list of paths to IQRnlmeProject folders selected by backwards elimination process
}
\description{
Perform forward selection of covariates in a model.
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
