% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ActivityMonitoring.R
\name{generate_ActivityContent}
\alias{generate_ActivityContent}
\title{generate_ActivityContent}
\usage{
generate_ActivityContent(
  ActivityList,
  colADD = c("ProjectName", "CompoundSummary", "ActivityName", "ActivityType",
    "ActivityPath", "ModellerName", "Status", "StartDate", "LastRunDay", "FinalDate",
    "StudyType", "Species", "SubjectType", "ParasiteType", "InfectionType",
    "IndicationType", "ActivityResultsSummary", "reviewDate", "revierName",
    "workflowName", "workflowVersion", "Rversion", "IQRversion", "tool", "toolVersion",
    "ActivityDescription"),
  filename = NULL
)
}
\arguments{
\item{colADD}{Default: c("ProjectName", "CompoundSummary", "ActivityName", "ActivityType",
"ActivityPath", "ModellerName", "Status", "StartDate", "LastRunDay",
"FinalDate", "StudyType", "Species", "SubjectType", "ParasiteType",
"InfectionType", "IndicationType", "ActivityResultsSummary",
"reviewDate", "revierName", "workflowName", "workflowVersion",
"Rversion", "IQRversion", "tool", "toolVersion", "ActivityDescription")}

\item{filename}{Default: NULL}
}
\description{
generate_ActivityContent
}
\seealso{
Other Activity Monitoring: 
\code{\link{check_ActivityInfo}()},
\code{\link{check_ActivityInfo_PiNK}()},
\code{\link{check_ActivityInfo_Sdrive}()},
\code{\link{check_ActivityResults}()},
\code{\link{create_RprofileMMV}()},
\code{\link{generate_ActivityResults}()},
\code{\link{generate_DuplicatedActivityReport}()},
\code{\link{generate_MissingActivityReport}()},
\code{\link{generate_OverviewFile}()},
\code{\link{get_ActivityInfo}()},
\code{\link{get_ActivityList}()},
\code{\link{list_ActivityType}()},
\code{\link{list_GenericType}()},
\code{\link{set_MMVsettings}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Activity Monitoring}
