% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-Chemoprotection.R
\name{generate_bite_time_poisson}
\alias{generate_bite_time_poisson}
\title{Generate Bite Times Using a Poisson Process}
\usage{
generate_bite_time_poisson(force_of_infection, Tend)
}
\arguments{
\item{force_of_infection}{Numeric. Should represent an assumption about the number of infectious bites
that occur over a given time period. This can derived, for example, from estimated annual entomological
inoculation rates.This number should represent a yearly count : the provided parameter will be 
in the form x / 365 /24 to provide a daily rate}

\item{Tend}{Numeric. The observation period (in hours) during which bite events will be simulated.
If being used as part of a simulation, should likely reflect the simulation's maximum run time}
}
\value{
A numeric vector of bite times (in hours) within the specified observation period.
        Each element represents the cumulative time (in hours) at which a bite event occurs.
}
\description{
This function generates bite event times over a specified observation period using 
a Poisson process, based on a mean bite rate (events per day).
}
\details{
The function calculates a rate parameter for a Poisson process that represents 
the mean number of bite events per hour, derived from a yearly figure of infectious bites (`force_of_infection`). 
Using this rate parameter, it generates a cumulative sum of exponentially-distributed
inter-arrival times for bite events.

If the generated sequence of bite events does not span the entire observation period (`Tend`), 
additional events are appended to ensure sufficient coverage, stopping when the maximum 
bite event time exceeds the observation period.

The function returns only those events that occur within the observation period.
}
\examples{
# Generate bite times with an average of 10 bites per year over a 24-hour period
generate_bite_time_poisson(force_of_infection = 10, Tend = 24)

# Generate bite times with an average of 5 bites per year over a 48-hour period
generate_bite_time_poisson(force_of_infection = 5, Tend = 48)

}
\seealso{
Other Chemoprotection: 
\code{\link{add_bites_to_trial_files}()},
\code{\link{evaluate_BreakthroughEvent}()},
\code{\link{generate_ChemoModel_FluxLimiter}()},
\code{\link{generate_ChemoModel_UpwindFirstOrder}()}
}
\author{
Sam Jones (MMV, \email{joness@mmv.org})
}
\concept{Chemoprotection}
