% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-Chemoprotection.R
\name{generate_ChemoModel_UpwindFirstOrder}
\alias{generate_ChemoModel_UpwindFirstOrder}
\title{Generate Chemoprevention Model (Upwind First Order scheme)}
\usage{
generate_ChemoModel_UpwindFirstOrder(
  nCpt,
  Finv = 1,
  VBlood = 5000,
  LiverDuration = 6,
  LiverGrowth = list(GR = "kGRliver", parameters = list(kGRliver = list(value = 0.07,
    notes = "Net parasite growth rate in liver [1/hr]")), extraStates = NULL),
  LiverActivity = list(Kill =
    "EMAXLiver*(Cc+yps)^hLiver/((Cc+yps)^hLiver+EC50Liver^hLiver)", parameters =
    list(EMAXLiver = list(value = 0.3, notes = "Maximum killing rate in liver [1/hr]"),
    EC50Liver = list(value = 1, notes =
    "Blood concentration achieving 50percent of maximum effect [ug/mL] in liver"), hLiver
    = list(value = 3, notes = "Hill coefficient in liver [.]")), extraStates = NULL),
  BloodGrowth = list(GR = "kGRblood", parameters = list(kGRblood = list(value = 0.07,
    notes = "Net parasite growth rate in blood [1/hr]")), extraStates = NULL),
  BloodActivity = list(Kill =
    "EMAXBlood*(Cc+yps)^hBlood/((Cc+yps)^hBlood+EC50Blood^hBlood)", parameters =
    list(EMAXBlood = list(value = 0.3, notes = "Maximum killing rate in blodd [1/hr]"),
    EC50Blood = list(value = 1, notes =
    "Concentration achieving 50percent of maximum effect [ug/mL] in blood"), hBlood =
    list(value = 3, notes = "Hill coefficient in blood")), extraStates = NULL),
  PKmodel = IQRmodel(file.path(get_MMVmalariaPath(subdir = "inst"),
    "modelLibrary/PKmodels/model_3cpt_linear_abs1.txt")),
  Output = list(OUTPUT1 = "PBlood"),
  outputFile
)
}
\arguments{
\item{nCpt}{Integer representing the number of liver compartments to be generated in the model}

\item{Finv}{Fraction of parasites invading the liver per infectious bite (Default: 1)}

\item{VBlood}{Blood volume (mL) (Default: 5000)}

\item{LiverDuration}{Duration of liver stage (days) (Default: 6)}

\item{LiverGrowth}{List of GR (referring to the PD model, e.g. exponential growth), parameters and extraStates if applicable}

\item{LiverActivity}{List of Kill (referring to the PD model, e.g. Emax equation), parameters and extraStates if applicable}

\item{BloodGrowth}{List of GR (referring to the PD model, e.g. exponential growth), parameters and extraStates if applicable}

\item{BloodActivity}{List of Kill (referring to the PD model, e.g. Emax equation), parameters and extraStates if applicable}

\item{PKmodel}{"IQRmodel" or list of states, parameters and variables}

\item{Output}{List of outputs to be included in the model script (Default: \verb{"OUTPUT1 = "PBlood""})}

\item{outputFile}{Pathway where to save the resulting model file}
}
\value{
IQRmodel script to be used for estimation/simulation activities
}
\description{
Generates a chemoprevention model script using a Upwind First Order scheme to characterize the liver stage
}
\examples{
# Example 1: Emax model , PKmodel from library
 outputFile <- tempfile()
 generate_ChemoModel_UpwindFirstOrder(nCpt = 4,
                                      Finv = 1,
                                      VBlood = 5000, # in mL
                                      LiverDuration = 6,   # In Days,
                                      LiverGrowth   = list(GR          = "kGRliver",
                                                           parameters  = list(kGRliver = list(value = 0.07, notes = "Net parasite growth rate in liver (1/hr)")),
                                                           extraStates = NULL),
                                      LiverActivity = list(Kill = "EMAXLiver*(Cc+yps)^hLiver/((Cc+yps)^hLiver+EC50Liver^hLiver)",
                                                           parameters = list(EMAXLiver = list(value = 0.3, notes = "Maximum killing rate in liver (1/hr)"),
                                                                             EC50Liver = list(value = 1, notes = "Blood concentration achieving 50percent of maximum effect (ug/mL) in liver"),
                                                                             hLiver    = list(value = 3, notes = "Hill coefficient in liver")),
                                                           extraStates = NULL),
                                      BloodGrowth   = list(GR          = "kGRblood",
                                                           parameters  = list(kGRblood = list(value = 0.07, notes = "Net parasite growth rate in blood (1/hr)")),
                                                           extraStates = NULL),
                                      BloodActivity = list(Kill = "EMAXBlood*(Cc+yps)^hBlood/((Cc+yps)^hBlood+EC50Blood^hBlood)",
                                                           parameters = list(EMAXBlood = list(value = 0.3, notes = "Maximum killing rate in blodd (1/hr)"),
                                                                             EC50Blood = list(value = 1, notes = "Concentration achieving 50percent of maximum effect (ug/mL) in blood"),
                                                                             hBlood    = list(value = 3, notes = "Hill coefficient in blood")),
                                                           extraStates = NULL),

                                      PKmodel = IQRmodel(file.path(get_MMVmalariaPath(subdir="inst"),"modelLibrary/PKmodels/model_3cpt_linear_abs1.txt")),
                                      Output = list(OUTPUT1 = "PBlood"),
                                      outputFile)

# Example 2: Effect compartment model in blood, PKmodel manually defined
 outputFile <- tempfile()
 generate_ChemoModel_UpwindFirstOrder(nCpt = 4,
                                      Finv = 1,
                                      VBlood = 5000, # in mL
                                      LiverDuration = 6,   # In Days,
                                      LiverGrowth   = list(GR          = "kGRliver",
                                                           parameters  = list(kGRliver = list(value = 0.07, notes = "Net parasite growth rate in liver (1/hr)")),
                                                           extraStates = NULL),
                                      LiverActivity = list(Kill = "EMAXBlood*(Cc+yps)^hBlood/((Cc+yps)^hBlood+EC50Blood^hBlood)",
                                                           parameters = list(EMAXLiver = list(value = 0.3, notes = "Maximum killing rate in liver (1/hr)"),
                                                                             EC50Liver = list(value = 1, notes = "Blood concentration achieving 50percent of maximum effect (ug/mL) in liver"),
                                                                             hLiver    = list(value = 3, notes = "Hill coefficient in liver")),
                                                           extraStates = NULL),
                                      BloodGrowth   = list(GR          = "kGRblood",
                                                           parameters  = list(kGRblood = list(value = 0.07, notes = "Net parasite growth rate in blood (1/hr)")),
                                                           extraStates = NULL),
                                      BloodActivity = list(Kill = "EMAXLiver*(Ce+yps)^hLiver/((Ce+yps)^hLiver+EC50Liver^hLiver)",
                                                           parameters = list(EMAXBlood = list(value = 0.3, notes = "Maximum killing rate in blodd (1/hr)"),
                                                                             EC50Blood = list(value = 1, notes = "Concentration achieving 50percent of maximum effect (ug/mL) in blood"),
                                                                             hBlood    = list(value = 3, notes = "Hill coefficient in blood")),
                                                           extraStates = list(Ce = list(ODE = "d/dt(Ce) = ke*(Cc - Ce)", IC = "0"))),

                                      PKmodel = list(states = list(Ad = list(ODE = "-ka*Ad + Fabs1*INPUT1", IC = "0"),
                                                                   Ac = list(ODE = "ka*Ad - Q1/Vc*Ac + Q1/Vp1*Ap1 - Q2/Vc*Ac + Q2/Vp2*Ap2 - CL/Vc*Ac", IC = "0"),
                                                                   Ap1 = list(ODE = "Q1/Vc*Ac - Q1/Vp1*Ap1", IC = "0"),
                                                                   Ap2 = list(ODE = "Q2/Vc*Ac - Q2/Vp2*Ap2", IC = "0")
                                                                   ),
                                                     parameters = list(Fabs1 = list(value = 1, notes="Relative bioavailability (-)"),
                                                                       ka    = list(value = 2, notes="Absorption rate parameter (1/hour)"),
                                                                       CL    = list(value = 3, notes="Apparent clearance (L/hour)"),
                                                                       Vc    = list(value = 32, notes="Apparent central volume (L)"),
                                                                       Q1    = list(value = 1, notes="Apparent intercompartmental clearance to first peripheral compartment (L/hour)"),
                                                                       Vp1   = list(value = 10, notes="Apparent first peripheral volume (L)"),
                                                                       Q2    = list(value = 1, notes="Apparent intercompartmental clearance to second peripheral compartment (L/hour)"),
                                                                       Vp2   = list(value = 10, notes="Apparent second peripheral volume (L)"),
                                                                       Tlag1 = list(value = 0, notes="Absorption lag time (hours)")
                                                                       ),
                                                    variables  = list(Cc = list(formula = "Ac/Vc"))),
                                      Output = list(OUTPUT1 = "PBlood"),
                                      outputFile)

}
\seealso{
Other Chemoprotection: 
\code{\link{add_bites_to_trial_files}()},
\code{\link{evaluate_BreakthroughEvent}()},
\code{\link{generate_ChemoModel_FluxLimiter}()},
\code{\link{generate_bite_time_poisson}()}
}
\author{
Catalina Barcelo (MMV, \email{barceloc@mmv.org})
}
\concept{Chemoprotection}
