% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{generate_EstimateTableForReport}
\alias{generate_EstimateTableForReport}
\title{generate_EstimateTableForReport}
\usage{
generate_EstimateTableForReport(
  projectPath,
  tableOutputPath = NULL,
  FLAGLaTeX = TRUE,
  setting = NULL
)
}
\arguments{
\item{projectPath}{IQR pr GPF project path}

\item{tableOutputPath}{optional string denoting location to write table, will attempt to default to provided projectPath if NULL}

\item{FLAGLaTeX}{Flag to convert parameters name into LaTeX format (Default: \code{TRUE})}

\item{setting}{Optional settings for \code{table_strToLaTeX}; \code{ListStrToLatex}, \code{TitleRow}, \code{report} and \code{FLAGcompliance}}
}
\value{
List containing the table as data.frame, title and footer
}
\description{
Generate a table of estimate from IQR or GPF output in a summarized format
}
\seealso{
Other Reporting: 
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_PKmodelHypos}()},
\code{\link{reportText_convertListToText}()},
\code{\link{table_PKsummaryCompare}()},
\code{\link{table_PKsummaryResults}()},
\code{\link{table_covariateSummaryResults}()},
\code{\link{table_strToLaTeX}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Reporting}
