% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{generate_GPFfromOldIQRproject}
\alias{generate_GPFfromOldIQRproject}
\title{Generate GPF Object from an old IQR project}
\usage{
generate_GPFfromOldIQRproject(
  projectPath,
  filename = NULL,
  referenceCovariate = NULL
)
}
\arguments{
\item{projectPath}{IQR project path}

\item{filename}{Name of the XLS file to save the GPF model is not \code{NULL} (Default: \code{NULL})}

\item{referenceCovariate}{For the oldest IQR project the referenced covariate can be entered manually. It should be a vector or data frame with the name of the covariate and their reference value: e.g. \code{referenceCovariate = c("WT0"=70, "FORM"=1)} (Default: \code{NULL})}
}
\value{
A GPF object
}
\description{
IQRtools evolved since its first implementation, and it is not possible to
convert old IQR project to the GPF format using the default IQRtools functions.
Therefore, this MMVmalaria function help to generate a GPF object from the summary
result table generated by IQRtools. However, the covariance matrix is lost in the process.
As well, covariate reference might not be retrieve for the oldest project.
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
