% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{generate_GPFfromParametersTable}
\alias{generate_GPFfromParametersTable}
\title{Generate GPF Object from a parameter table}
\usage{
generate_GPFfromParametersTable(
  parametersTable,
  uncertaintyCorrelation = NULL,
  filename = NULL,
  referenceCovariate = NULL,
  modelInfo = NULL
)
}
\arguments{
\item{parametersTable}{IQR/IQM result table}

\item{uncertaintyCorrelation}{IQR/IQM uncertainty correlation results table}

\item{filename}{Name of the XLS file to save the GPF model is not \code{NULL} (Default: \code{NULL})}

\item{referenceCovariate}{For the oldest IQR project the referenced covariate can be entered manually. It should be a vector or data frame with the name of the covariate and their reference value: e.g. \code{referenceCovariate = c("WT0"=70, "FORM"=1)} (Default: \code{NULL})}

\item{modelInfo}{Model information extracted from `project.est` or `project.IQRnlmeEst` of the IQRproject}
}
\value{
A GPF object
}
\description{
Generate GPF Object from a parameter table
}
\details{
Function to be used in `generate_GPFfromOldIQRproject` to geneate a GPF object.
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
