% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-NLMEmodeling.R
\name{generate_NLMEalgorithmSettings}
\alias{generate_NLMEalgorithmSettings}
\title{generate_NLMEalgorithmSettings}
\usage{
generate_NLMEalgorithmSettings(
  multiTestN = 1,
  multiTestSD = 0.5,
  FLAGanalytic = TRUE,
  keepProjectFolder = FALSE,
  algOpt.SEED = 123456,
  algOpt.K1 = 500,
  algOpt.K2 = 200,
  algOpt.NRCHAINS = 5,
  algOpt.NONMEM.METHOD = "SAEM",
  algOpt.NONMEM.MAXEVAL = 9999,
  algOpt.NONMEM.SIGDIGITS = 3,
  algOpt.NONMEM.PRINT = 1,
  algOpt.NONMEM.COVSTEP_MATRIX = "S",
  algOpt.NONMEM.ADVAN7 = TRUE,
  algOpt.NONMEM.N1 = 1000,
  algOpt.NONMEM.TOL = 6,
  algOpt.NONMEM.SIGL = NULL,
  algOpt.NONMEM.M4 = FALSE,
  algOpt.NONMEM.FOCEIOFV = FALSE,
  algOpt.NONMEM.IMPORTANCESAMPLING = TRUE,
  algOpt.NONMEM.IMP_ITERATIONS = 10,
  algOpt.NONMEM.ITS = TRUE,
  algOpt.NONMEM.ITS_ITERATIONS = 10,
  algOpt.NONMEM.WRES = NULL,
  algOpt.NONMEM.PRED = NULL,
  algOpt.NONMEM.RES = NULL,
  algOpt.MONOLIX.individualParameters = "conditionalMode",
  algOpt.MONOLIX.indivMCMClength = 50,
  algOpt.MONOLIX.indivNsim = 10,
  algOpt.MONOLIX.indivRatio = 0.05,
  algOpt.MONOLIX.logLikelihood = "Linearization",
  algOpt.MONOLIX.fim = "Linearization",
  algOpt.MONOLIX.variability = "Decreasing",
  algOpt.MONOLIX.startTime = NULL,
  algOpt.MONOLIX.STIFF = TRUE,
  algOpt.NLMIXR.method = "SAEM",
  algOpt.NLMIXR.control = NULL
)
}
\arguments{
\item{multiTestN}{Default: 1}

\item{multiTestSD}{Default: 0.5}

\item{FLAGanalytic}{Default: \code{TRUE}}

\item{keepProjectFolder}{Default: \code{FALSE}}

\item{algOpt.SEED}{Default: 123456}

\item{algOpt.K1}{Default: 500}

\item{algOpt.K2}{Default: 200}

\item{algOpt.NRCHAINS}{Default: 5}

\item{algOpt.NONMEM.METHOD}{Default: \code{"SAEM"}}

\item{algOpt.NONMEM.MAXEVAL}{Default: 9999}

\item{algOpt.NONMEM.SIGDIGITS}{Default: 3}

\item{algOpt.NONMEM.PRINT}{Default: 1}

\item{algOpt.NONMEM.COVSTEP_MATRIX}{Default: \code{"S"}}

\item{algOpt.NONMEM.ADVAN7}{Default: \code{TRUE}}

\item{algOpt.NONMEM.N1}{Default: 1000}

\item{algOpt.NONMEM.TOL}{Default: 6}

\item{algOpt.NONMEM.SIGL}{Default: \code{NULL}}

\item{algOpt.NONMEM.M4}{Default: \code{FALSE}}

\item{algOpt.NONMEM.FOCEIOFV}{Default: \code{FALSE}}

\item{algOpt.NONMEM.IMPORTANCESAMPLING}{Default: \code{TRUE}}

\item{algOpt.NONMEM.IMP_ITERATIONS}{Default: 10}

\item{algOpt.NONMEM.ITS}{Default: \code{TRUE}}

\item{algOpt.NONMEM.ITS_ITERATIONS}{Default: 10}

\item{algOpt.MONOLIX.individualParameters}{Default: \code{"conditionalMode"}}

\item{algOpt.MONOLIX.logLikelihood}{Default: \code{"Linearization"}}

\item{algOpt.MONOLIX.fim}{Default: \code{"Linearization"}}

\item{algOpt.MONOLIX.variability}{Default: \code{"Decreasing"}}

\item{algOpt.MONOLIX.startTime}{Default: \code{NULL}}

\item{algOpt.MONOLIX.STIFF}{Default: \code{TRUE}}

\item{algOpt.NLMIXR.method}{Default: \code{"SAEM"}}

\item{algOpt.NLMIXR.control}{Default: \code{NULL}}
}
\value{

}
\description{
generate_NLMEalgorithmSettings
}
\seealso{
Other NLME Modeling: 
\code{\link{generate_InitialParametersIQRnlme}()},
\code{\link{modelSpec_MMVmalariaProject}()},
\code{\link{testPKmodelsIQR}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{NLME Modeling}
