% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ActivityMonitoring.R
\name{generate_OverviewFile}
\alias{generate_OverviewFile}
\title{generate_OverviewFile}
\usage{
generate_OverviewFile(
  outputOverview = "../../../../../Content/00-ActivityList",
  ActivityName = NULL,
  ProjectName = NULL,
  ActivityPath = NULL,
  NewOverviewFolder = FALSE,
  FLAGforceOverview = FALSE
)
}
\arguments{
\item{outputOverview}{Default: '../../../../../Content/00-ActivityList'}

\item{ActivityName}{Default: NULL}

\item{ProjectName}{Default: NULL}

\item{ActivityPath}{Default: NULL}

\item{NewOverviewFolder}{Default: FALSE}

\item{FLAGforceOverview}{Default: FALSE}
}
\description{
generate_OverviewFile
}
\seealso{
Other Activity Monitoring: 
\code{\link{check_ActivityInfo}()},
\code{\link{check_ActivityInfo_PiNK}()},
\code{\link{check_ActivityInfo_Sdrive}()},
\code{\link{check_ActivityResults}()},
\code{\link{create_RprofileMMV}()},
\code{\link{generate_ActivityContent}()},
\code{\link{generate_ActivityResults}()},
\code{\link{generate_DuplicatedActivityReport}()},
\code{\link{generate_MissingActivityReport}()},
\code{\link{get_ActivityInfo}()},
\code{\link{get_ActivityList}()},
\code{\link{list_ActivityType}()},
\code{\link{list_GenericType}()},
\code{\link{set_MMVsettings}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Activity Monitoring}
