% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-SYSFITmodeling.R
\name{generate_SYSalgorithmSettings}
\alias{generate_SYSalgorithmSettings}
\title{generate_SYSalgorithmSettings}
\usage{
generate_SYSalgorithmSettings(
  SIMOPT.method = "lsodes",
  SIMOPT.atol = 1e-06,
  SIMOPT.rtol = 1e-06,
  SIMOPT.hmin = 0,
  SIMOPT.hmax = NULL,
  SIMOPT.hini = 0,
  SIMOPT.maxsteps = 5000,
  SIMOPT.nauxtimes = 0,
  opt.nfits = 10,
  opt.sd = 1,
  opt.rinit = 1,
  opt.rmax = 10,
  opt.iterlim = 100,
  opt.prior_sigma = 10,
  opt.parlower = NULL,
  opt.parupper = NULL,
  algOpt.SEED = 123456,
  FLAGprofileLL = FALSE,
  FLAGkeepFits = FALSE
)
}
\arguments{
\item{SIMOPT.method}{Default: 'lsodes'}

\item{SIMOPT.atol}{Default: 1e-06}

\item{SIMOPT.rtol}{Default: 1e-06}

\item{SIMOPT.hmin}{Default: 0}

\item{SIMOPT.hmax}{Default: NULL}

\item{SIMOPT.hini}{Default: 0}

\item{SIMOPT.maxsteps}{Default: 5000}

\item{SIMOPT.nauxtimes}{Default: 0}

\item{opt.nfits}{Default: 10}

\item{opt.sd}{Default: 1}

\item{opt.rinit}{Default: 1}

\item{opt.rmax}{Default: 10}

\item{opt.iterlim}{Default: 100}

\item{opt.prior_sigma}{Default: 10}

\item{opt.parlower}{Default: NULL}

\item{opt.parupper}{Default: NULL}

\item{algOpt.SEED}{Default: 123456}

\item{FLAGprofileLL}{Default: FALSE}

\item{FLAGkeepFits}{Default: FALSE}
}
\description{
generate_SYSalgorithmSettings
}
\seealso{
Other SYSFITmodeling: 
\code{\link{generate_InitialParametersIQRsys}()},
\code{\link{table_BestFitEstimation}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{SYSFITmodeling}
