% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{generateTrialSim}
\alias{generateTrialSim}
\title{generateTrialSim}
\usage{
generateTrialSim(modelFolder, data, NTRIALS, covariateDose = NULL)
}
\arguments{
\item{modelFolder}{String denoting path to IQRnlmeProject or IQRnlmeProjectMulti; if Multi the first element will be used}

\item{data}{IQRdataGeneral object used to generate dosing info for VPC production}

\item{NTRIALS}{Number of trials to simulate}

\item{covariateDose}{Default: \code{NULL}}
}
\value{
data.frame containing simulated clinical trials
}
\description{
Simulate a number of clinical trials from provided IQRnlmeProject
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Aline Fuchs (MMV)
}
\concept{Simulations}
