% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ActivityMonitoring.R
\name{get_ActivityList}
\alias{get_ActivityList}
\title{get_ActivityList}
\usage{
get_ActivityList(folder_List = "../00-ActivityList")
}
\arguments{
\item{folder_List}{Default: '../00-ActivityList'}
}
\description{
get_ActivityList
}
\seealso{
Other Activity Monitoring: 
\code{\link{check_ActivityInfo}()},
\code{\link{check_ActivityInfo_PiNK}()},
\code{\link{check_ActivityInfo_Sdrive}()},
\code{\link{check_ActivityResults}()},
\code{\link{create_RprofileMMV}()},
\code{\link{generate_ActivityContent}()},
\code{\link{generate_ActivityResults}()},
\code{\link{generate_DuplicatedActivityReport}()},
\code{\link{generate_MissingActivityReport}()},
\code{\link{generate_OverviewFile}()},
\code{\link{get_ActivityInfo}()},
\code{\link{list_ActivityType}()},
\code{\link{list_GenericType}()},
\code{\link{set_MMVsettings}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Activity Monitoring}
