% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{get_CLmaxTlag}
\alias{get_CLmaxTlag}
\title{get_CLmaxTlag
Determines maximum clearance and lag time of parasitemia}
\usage{
get_CLmaxTlag(x)
}
\arguments{
\item{x}{data frame with one or two segments characterizing parasite clearance (output of \link{fit_LinearSegments})}
}
\value{
input data frame amended by CLmax and Tlag columns
}
\description{
Derives maximum rate of clearance and lag time before it is achieved
based on segments fitted to PD data in clearance phase.
}
\details{
The input is a data frame with knots of the (at most 2) segments for an individual as
returned by fit_LinearSegments.
If only one segment present, lag time is zero and maximum clearance the slope of this segment.
If there are two segments, maximum clearance is the steeper slope of the two.
Tlag is the time of the middle knot if the first slope is smaller than 20\% of the second and 0 otherwise.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
