% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{get_DataClearance}
\alias{get_DataClearance}
\title{get_DataClearance
Gets PD subset from the parasite clearance phase}
\usage{
get_DataClearance(x)
}
\arguments{
\item{x}{data frame with parasitemia records for an individual}
}
\value{
subsetted input data frame
}
\description{
Subsets parasitemia data to initial clearance phase
after dosing, i.e., until minimum in parasite count
or first occurence of LLOQ data point
}
\details{
Expects that given data starts as drug administration
(no pre-dose growth data to be present).
Expects TIME, VALUE, and LLOQ columns.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
