% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{get_DataGrowth}
\alias{get_DataGrowth}
\title{get_DataGrowth
Gets PD subset from the parasite recrudescence phase}
\usage{
get_DataGrowth(x)
}
\arguments{
\item{x}{data frame with parasitemia records for an individual}
}
\value{
subsetted input data frame
}
\description{
Subsets parasitemia data to re-growth or recrudescence phase
, i.e., after minimum in parasite count
or last occurence of LLOQ data point
}
\details{
Expects TIME, VALUE, and LLOQ columns.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{simulate_indivPK}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
