% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{get_fileNameExtension}
\alias{get_fileNameExtension}
\title{get_fileNameExtension
Get the extension of 'filename'}
\usage{
get_fileNameExtension(filename)
}
\arguments{
\item{filename}{Path to a file}
}
\value{
TRUE/FALSE
}
\description{
get_fileNameExtension
Get the extension of 'filename'
}
\seealso{
Other General Function: 
\code{\link{is.fileMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV), \email{cherkaouim@mmv.org}
}
\concept{General Function}
