% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{get_Growth}
\alias{get_Growth}
\title{get_Growth
Determines growth rate from line through growth data}
\usage{
get_Growth(x)
}
\arguments{
\item{x}{data frame with two knots characterizing growth (output of \link{fit_LinearModel})}
}
\value{
input data frame amended by GR column with growth rate value
}
\description{
Determines growth rate based on line fitted to PD data in growth phase.
As the line is on logscale the growth rate is the slope of the line.
}
\details{
The input is a data frame with knots of the line for an individual as
returned by fit_LinearModel
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{simulate_indivPK}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
