% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{get_IQRdistribution}
\alias{get_IQRdistribution}
\title{Get Distribution}
\usage{
get_IQRdistribution(Distribution)
}
\arguments{
\item{Distribution}{Character vector containing the transform formula from an IQR model}
}
\value{
Return a charaacter vector containing the distribution index; i.e. "N" for Normal, "L" for Log-Normal and "G" for Logit
}
\description{
Get Distribution of list of parameters from an IQR model
}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{adjust_Dose}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{create_PKPDsimtime}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{getDoselevel}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{get_MMVmalariaPath}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{saveActivityInfo}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
