% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{get_IXGDFtoRemove}
\alias{get_IXGDFtoRemove}
\title{Get Row ID to be removed}
\usage{
get_IXGDFtoRemove(
  dataGeneral,
  obNAME = "log(Parasitemia Total)",
  removeType = "ToLastOb"
)
}
\arguments{
\item{dataGeneral}{A general dataset to look into}

\item{obNAME}{Default: 'log(Parasitemia Total)'}

\item{removeType}{Default: 'ToLastOb'}
}
\description{
Get Row ID to be removed
}
\details{
_IXGDFtoRemove
}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{adjust_Dose}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{convert_Unit}()},
\code{\link{create_PKPDsimtime}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{getDoselevel}()},
\code{\link{get_IQRdistribution}()},
\code{\link{get_MMVmalariaPath}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{saveActivityInfo}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
