% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{get_LagTime}
\alias{get_LagTime}
\title{get_LagTime}
\usage{
get_LagTime(
  dataSim,
  GR = NULL,
  EMAX = NULL,
  effCOL = "Eff",
  Parasitemia = "PL",
  Plog = TRUE,
  lagTimeTYPE = "MIC"
)
}
\arguments{
\item{GR}{Default: `NULL`}

\item{EMAX}{Default: `NULL`}

\item{effCOL}{Default: 'Effect'}

\item{Plog}{Default: `TRUE`}
}
\value{

}
\description{
get_LagTime
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR28}()},
\code{\link{evaluate_APR}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
