% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{get_linepar}
\alias{get_linepar}
\title{Slope and intercept of line}
\usage{
get_linepar(twoknots)
}
\arguments{
\item{twoknots}{data frame with Xknot and Yknot colums and two rows}
}
\value{
A name vector with intercept and slope
}
\description{
\code{get_linepar} Calculates the slope and intercept of a line defined by two points
}
\details{
Given the coordinates of two points, the slope and the intercept
for the line through these points is determined.

Each row of the data frame defines one point.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Model-free PD assessment by linear segments}
