% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{get_MICextrapolPK}
\alias{get_MICextrapolPK}
\title{get_MICextrapolPK
Determine MIC based on linearily inter/extrapolated PK}
\usage{
get_MICextrapolPK(
  tMICindiv,
  dataPK,
  plottitle = "MIC determined by linear extrapolation of parasitemia and PK data"
)
}
\arguments{
\item{tMICindiv}{data frame with individual times of MIC (Xknot column)}

\item{dataPK}{data frame with PK data (expects columns USUBJID, TIME, VALUE, LLOQ, CENS, TIMEUNIT
as provided by general dataset format in MMVmalaria or IQRtools)}

\item{plottitle}{character (string) to be printed as figure title}
}
\value{
list with data frame of individual MIC values and a plot (ggplot object) illustrating the MIC determination
}
\description{
The MIC is determined as the concentration extrapolated
linearily based on the PK data at a given time of MIC
which was determined based on linear PD data interpolation.
}
\details{
Linear segments are fitted to the individual PK data by \link{fit_LinSegments}.
The MIC is taken as the value of the respective segment at the given time of MIC in 'tMICindiv'
If the time of MIC is later than the observed data, i.e., later than the time described by the segment,
the last segment is extrapolated.
Concentration data is expected to be given in ug/mL.
Note that only single dose data is handled appropriately.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{simulate_indivPK}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
