% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{get_MICsimPK}
\alias{get_MICsimPK}
\title{get_MICsimPK
Determine MIC based on simulated PK}
\usage{
get_MICsimPK(
  dataDOS,
  model,
  tMICindiv,
  dataPK,
  plottitle =
    "MIC determined by linear extrapolation of parasitemia data and PK simulation"
)
}
\arguments{
\item{dataDOS}{data frame with individual dosing information (defined by TIME, AMT and RATE)
and individual PK parameters.
Expects columns USUBJID, TRT, TRTNAME, and STUDY to identify and annotate individuals}

\item{model}{IQRmodel of the PK model to use for simulation}

\item{tMICindiv}{data frame with timepoints of MIC (Xknot) for individuals (USUBJID)}

\item{dataPK}{data frame with PK data (expects columns USUBJID, TIME, VALUE, LLOQ, CENS, TIMEUNIT
as provided by general dataset format in MMVmalaria or IQRtools)}

\item{plottitle}{character (string) to be printed as figure title}
}
\value{
list with data frame of individual MIC values and a plot (ggplot object) illustrating the MIC determination
}
\description{
The MIC is determined as the concentration simulated
at a given time of MIC which was determined based on
linear PD data interpolation.
}
\details{
Individual PK timexourses are simulated by given model and individual PK parameters.
The MIC is taken as the simulated value given the time of MIC in 'tMICindiv'
Concentration data is expected to be given in ug/mL.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Model-free PD assessment by linear segments}
