% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{get_MMVmalariaPath}
\alias{get_MMVmalariaPath}
\title{Retrieve Package Path}
\usage{
get_MMVmalariaPath(subdir = NULL)
}
\arguments{
\item{subdir}{A subdirectory within the MMVmalaria package directory. If NULL, the function returns the main package path.}
}
\value{
A character string representing the path to the MMVmalaria package directory or a specified subdirectory.
}
\description{
This function retrieves the path to the MMVmalaria package directory
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
