% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{get_PRRtot}
\alias{get_PRRtot}
\title{get_PRRtot}
\usage{
get_PRRtot(dataSim, paraCOL = "PL", Plog = TRUE, MethodMin = "CubicSpline")
}
\arguments{
\item{dataSim}{Simulation data generated by \code{sim_IQRmodel()} or equivalent, containing at least a column named TIME and a column defined by \code{paraCOL}}

\item{paraCOL}{Character string denoting name of parasitaemia column in \code{dataSim}.Default: "PL"}

\item{Plog}{Logical, is parasitaemia in \code{dataSim} in log-scale? Default: \code{TRUE}}

\item{MethodMin}{Default: 'CubicSpline'}
}
\value{
Estimated log10 total PRR
}
\description{
Estimate total parasite reduction ratio from provided simulation data
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
