% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{getApparentKeys}
\alias{getApparentKeys}
\title{getApparentKeys}
\usage{
getApparentKeys(
  dataSim,
  concCOL = "Cc",
  effCOL = "Eff",
  paraCOL = "PL",
  convConc = 1,
  Plog = TRUE
)
}
\arguments{
\item{dataSim}{Simulation data generated by \code{sim_IQRmodel()} or equivalent}

\item{concCOL}{Character string denoting name of concentration column in \code{dataSim. }Default: 'Cc'}

\item{effCOL}{Character string denoting name of eff column in \code{dataSim}Default: 'Eff'}

\item{convConc}{Numeric: scalar for concentration data. Default: 1}

\item{Plog}{Logical, is parasitaemia in \code{dataSim} in log-scale? Default: \code{TRUE}}
}
\value{
Apparent MIC, MPC90, PRR24, PRR48 and PPR72
}
\description{
Estimate apparent MIC, MPC90, PRR24, PRR48 and PPR72 from provided simulation data
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
