% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{getCorrParNames}
\alias{getCorrParNames}
\title{getCorrParNames}
\usage{
getCorrParNames(corrString)
}
\arguments{
\item{corrString}{}
}
\value{
character vector with parameter names
}
\description{
Extract parameter names from IIV correlation name
Convention of IIV correlation estimate name:
'corr_(par1,par2)'
}
