% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-NLMEmodeling.R
\name{getCovRefValue}
\alias{getCovRefValue}
\title{Helper function to extract continuous covariate reference value
from FORMULA entry of MMVmalariaCSV model estimate table}
\usage{
getCovRefValue(formula)
}
\arguments{
\item{formula}{Character string of the covariate equation used}
}
\value{
Reference covariate value (as character)
}
\description{
Looks basically for numerical value in string. If none or more than one
is found NA is return and a warning issued.
}
\examples{
getCovRefValue("X=X_Pop*(WT0/65)^Beta")
}
