% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-IQRaux.R
\name{getDefaultEventTable}
\alias{getDefaultEventTable}
\title{Get a default IQReventTable}
\usage{
getDefaultEventTable(model, dosingTable = NULL, parameters = NULL)
}
\arguments{
\item{model}{\code{\link[IQRtools:IQRmodel]{IQRtools::IQRmodel()}}}

\item{dosingTable}{NULL or \code{\link[IQRtools:IQRdosing]{IQRtools::IQRdosing()}}}

\item{parameters}{Named vector to override parameter values}
}
\value{
\code{\link[IQRtools:IQReventTable]{IQRtools::IQReventTable()}}
}
\description{
Get a default IQReventTable
}
\details{
The two results are equivalent: \cr\preformatted{sim_IQRmodel(m)
sim_IQRmodel(m, eventTable = getDefaultEventTable(m))
}

Same with dosing: \cr\preformatted{sim_IQRmodel(m, dosingTable = IQRdosing(c(0,1), 1, 1))
sim_IQRmodel(m, eventTable = getDefaultEventTable(m, IQRdosing(c(0,1), 1, 1)))
}
}
\examples{
library(IQRtools)
fileURL <- "http://192.168.1.12:7804/IQauxfn/inst/02-ExampleModels/ModelLibraryPK/model_2cpt_linsat_iv.txt"
fl <- file.path(tempdir(), basename(fileURL))
download.file(fileURL, fl)
m <- IQRmodel(fl)
d <- IQRdosing(c(0,1),1,1)
eT <- getDefaultEventTable(m, dosingTable = d)

simres1 <- sim_IQRmodel(m, dosingTable = d)
simres2 <- sim_IQRmodel(m, eventTable = eT)

p <- c(CL = 100)
eT <- getDefaultEventTable(m, parameters = p)

simres1 <- sim_IQRmodel(m, parameters = p)
simres2 <- sim_IQRmodel(m, eventTable = eT)
}
\seealso{
Other Simulation: 
\code{\link{createSimIQRmodelSpec}()},
\code{\link{create_GDF_fromDesignSpec}()},
\code{\link{populateGDFWithSimValues_consistent}()},
\code{\link{sim_IQRmodelConsistent}()}
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
\concept{Simulation}
\concept{eventTable}
