% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-ModelComparison.R
\name{getEstimateTable}
\alias{getEstimateTable}
\title{getEstimateTable}
\usage{
getEstimateTable(modelFolder)
}
\value{

}
\description{
getEstimateTable
}
\seealso{
Other ModelComparison: 
\code{\link{table_ModelEstimates}()}
}
\author{
Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{ModelComparison}
