% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/98-Experimental.R
\name{getPopIndPred}
\alias{getPopIndPred}
\title{getPopIndPred}
\usage{
getPopIndPred(
  modelFolder,
  timeGrid = "obsTimes",
  abs0inputs = NULL,
  abs0Tk0param = NULL
)
}
\arguments{
\item{timeGrid}{Default: 'obsTimes'}

\item{abs0inputs}{Default: NULL}

\item{abs0Tk0param}{Default: NULL}
}
\value{

}
\description{
getPopIndPred
}
\author{
Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Experimental}
