% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{getPopParameters_MMVmalariaProject}
\alias{getPopParameters_MMVmalariaProject}
\title{getPopParameters_MMVmalariaProject}
\usage{
getPopParameters_MMVmalariaProject(
  projectPath,
  IndCovariates = NULL,
  verbose = TRUE
)
}
\arguments{
\item{projectPath}{A character string with the path to an IQRsysProject or IQRnlmeProject folder, or a GPF file.}

\item{IndCovariates}{data.frame describing patient or other subject covariate and regressor data, see \code{IQRtools::sampleIndParamValues}. Default: \code{NULL}}

\item{verbose}{Default: TRUE}
}
\value{
List of parameter values, see \code{IQRtools::sampleIndParamValues}
}
\description{
Get population parameters from a MMVmalaria Project
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
