% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{getPopParameters_MMVmalariaXLS}
\alias{getPopParameters_MMVmalariaXLS}
\title{getPopParameters_MMVmalariaXLS}
\usage{
getPopParameters_MMVmalariaXLS(filename, IndCovariates = NULL)
}
\arguments{
\item{filename}{A character string with the path to a CSV file with all parameters.}

\item{IndCovariates}{data.frame describing patient or other subject covariate and regressor data, see \code{IQRtools::sampleIndParamValues}. Default: \code{NULL}}
}
\value{
List of parameter values, see \code{IQRtools::sampleIndParamValues}
}
\description{
Get population parameters from a GPF/XLS file
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
