% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{getTimeAboveMIC}
\alias{getTimeAboveMIC}
\title{getTimeAboveMIC}
\usage{
getTimeAboveMIC(
  dataSim,
  MIC = NULL,
  timeCOL = "TIME",
  concCOL = "Cc",
  convConc = 1
)
}
\arguments{
\item{MIC}{Default: `NULL`}

\item{convConc}{Default: 1}
}
\value{

}
\description{
getTimeAboveMIC
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR28}()},
\code{\link{evaluate_APR}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
