% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{getTimeKRaboveGR}
\alias{getTimeKRaboveGR}
\title{Get Time When Kill Rate is Above Growth Rate (tKRGR)}
\usage{
getTimeKRaboveGR(dataSim, timeCOL = "TIME", killCOL = "KillBlood", GR = 0.07)
}
\arguments{
\item{dataSim}{A data frame or data.table containing the simulation data.}

\item{timeCOL}{A string specifying the name of the time column in \code{dataSim}. Default is \code{"TIME"}.}

\item{killCOL}{A string specifying the name of the kill rate column in \code{dataSim}. Default is \code{"KillBlood"}.}

\item{GR}{A numeric value representing the growth rate threshold. Default is 0.07. Expects a single value, but if growth rate
is a column in \code{dataSim}, arguments such as \code{dataSim$GR[1]} or \code{unique(dataSim$GR)} are suitable.}
}
\value{
A data frame with one row and one column:
\describe{
  \item{\code{tKRGR}}{Numeric value representing the total time during which the kill rate exceeds the growth rate.}
}
}
\description{
Calculates the total time during which the kill rate exceeds the growth rate in a simulation dataset.
}
\details{
The function computes the total time (\code{tKRGR}) during which the kill rate exceeds the growth rate in the provided simulation data.

It handles cases where the kill rate crosses the growth rate threshold between time points by applying corrections using linear interpolation. This ensures a more accurate estimation of \code{tKRGR}, especially when time steps are large.
}
\examples{
\dontrun{
# Example simulation data
dataSim <- data.frame(
  TIME = seq(0, 24, by = 1),
  KillBlood = runif(25, min = 0, max = 1),
  GR = 0.5
)

# Calculate tKRGR
resultKRGR <- getTimeKRaboveGR(
  dataSim = dataSim,
  timeCOL = "TIME",
  killCOL = "KillBlood",
  GR = 0.5
)

print(resultKRGR)
}

}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Sam Jones (MMV)
}
\concept{Key Parameters}
