% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{getTimeRecrudescence}
\alias{getTimeRecrudescence}
\title{getTimeRecrudescence}
\usage{
getTimeRecrudescence(
  x,
  GR,
  LLOQ,
  valCol = "DV",
  log = TRUE,
  Time2RecDef = "MMV"
)
}
\arguments{
\item{x}{data.frame containing \code{"TIME"} and \code{valCol} where \code{valCol} is a measurement of parasitaemia}

\item{GR}{Numeric, value of growth rate.}

\item{LLOQ}{Numeric: LLOQ of parasitaemia measurements.}

\item{valCol}{A string specifying the name of the column containing parasitaemia measurements. Default is \code{"TIME"}. Default: 'DV'}

\item{log}{Logical, is parasitaemia in log-scale? Default: \code{TRUE}}

\item{Time2RecDef}{String cenoting methodology with which to determine recrudescence: One of \code{MMV} or \code{TAD} MMV = once min Para or LOQ is achieved. TAD = return to Para BL value: Default: 'MMV'}
}
\value{
data.frame containing time of recrudesence
}
\description{
Estimate time of recrudescence
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
