% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{import_SCIDpkpdData}
\alias{import_SCIDpkpdData}
\title{import_SCIDpkpdData}
\usage{
import_SCIDpkpdData(
  dataFile,
  compound,
  PKsheet = NULL,
  PKrange = NULL,
  PKmapping,
  PKlloq = 1,
  PKlloqIdentifier = c("BLQ", "<LLOQ", "< LLOQ"),
  PKfactor = 0.001,
  PDsheet = NULL,
  PDrange = NULL,
  PDmapping,
  PDlloq = 0.01,
  PDlloqIdentifier = c("BLQ", "<0.01", "< 0,01", "< 0.01"),
  PDname = "Parasitemia Total",
  study = NULL,
  DayOfFirstDrugAdmin = NULL,
  centerNumber = -1,
  centerName = "",
  visitNumber = -1,
  nrdoses = NULL,
  intervaldose = NULL,
  route = NULL,
  DoseTimeSheet = "DosingTimeTable",
  DoseTimeRange = NULL,
  DoseTimeMapping,
  DoseSheet = "IndividualMetadataDrug1",
  DoseRange = NULL,
  DoseMapping = NULL,
  positiveQC,
  importVersion = centerName
)
}
\arguments{
\item{dataFile}{Path to the Excel file containing the data.}

\item{compound}{Name of compound administered in the study.}

\item{PKsheet}{Default: NULL}

\item{PKrange}{Default: NULL}

\item{PKmapping}{Character vector mapping PKsheet headers (left) to desired name in converted data format (right) (eg c("StudyID"= "STUDY", "Concentration" = "VALUE"))}

\item{PKlloq}{Default: 1}

\item{PKlloqIdentifier}{Default: c("BLQ", "<LLOQ", "< LLOQ")}

\item{PKfactor}{Default: 0.001}

\item{PDsheet}{Default: NULL}

\item{PDrange}{Default: NULL}

\item{PDmapping}{Character vector mapping PKsheet headers (left) to desired name in converted data format (right) (eg c("StudyID"= "STUDY", "Concentration" = "VALUE"))}

\item{PDlloq}{Default: 0.01}

\item{PDlloqIdentifier}{Default: c("BLQ", "<0.01", "< 0,01", "< 0.01")}

\item{PDname}{Default: 'Parasitemia Total'}

\item{study}{Default: NULL}

\item{DayOfFirstDrugAdmin}{Default: NULL}

\item{centerNumber}{Default: -1}

\item{centerName}{Default: ''}

\item{visitNumber}{Default: -1}

\item{nrdoses}{Default: NULL}

\item{intervaldose}{Default: NULL}

\item{route}{Default: NULL}

\item{DoseTimeSheet}{Default: 'DosingTimeTable'}

\item{DoseTimeRange}{Default: NULL}

\item{DoseTimeMapping}{Character vector mapping DoseTimesheet headers (left) to desired name in converted final dataset (right) (eg c("StudyID"="STUDY", "DosingTime"="NT"))}

\item{DoseSheet}{Default: 'IndividualMetadataDrug1'}

\item{DoseRange}{Default: NULL}

\item{DoseMapping}{Default: NULL}

\item{positiveQC}{Character vector identifying the name of positive control (must match upper or lower case as in original dataset)}

\item{importVersion}{Default: centerName}
}
\value{
Imported SCID data
}
\description{
Imports SCID PK and PD data from an Excel file.
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Catalina Barcelo (MMV), Mohammed H. Cherkaoui (MMV), \email{cherkaouim@mmv.org}
}
\concept{Data Preparation}
