% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{import_SCIDpkpdData_GSK}
\alias{import_SCIDpkpdData_GSK}
\title{import_SCIDpkpdData_GSK}
\usage{
import_SCIDpkpdData_GSK(
  dataFile,
  compound = NULL,
  PKsheet = "BloodLevels",
  PKrange = NULL,
  PKmapping,
  PKlloq = 1,
  PKlloqIdentifier = c("BLQ", "<LLOQ", "< LLOQ"),
  PKfactor = 0.001,
  PDsheet = "Parasitemia",
  PDrange = NULL,
  PDmapping,
  PDlloq = 0.01,
  PDlloqIdentifier = c("BLQ", "<0.01", "< 0,01", "< 0.01"),
  PDname = "Parasitemia Total",
  study = NULL,
  DayOfFirstDrugAdmin = 3,
  centerNumber = -1,
  centerName = "",
  visitNumber = -1,
  nrdoses = NULL,
  intervaldose = "qd",
  positiveQC = NULL
)
}
\arguments{
\item{dataFile}{Path to the Excel file containing the data.}

\item{compound}{Default: NULL}

\item{PKsheet}{Default: 'BloodLevels'}

\item{PKrange}{Default: NULL}

\item{PKmapping}{Character vector mapping PKsheet headers (left) to desired name in converted data format (right) (eg c("StudyID"= "STUDY", "Concentration" = "VALUE"))}

\item{PKlloq}{Default: 1}

\item{PKlloqIdentifier}{Default: c("BLQ", "<LLOQ", "< LLOQ")}

\item{PKfactor}{Default: 0.001}

\item{PDsheet}{Default: 'Parasitemia'}

\item{PDrange}{Default: NULL}

\item{PDmapping}{Character vector mapping PKsheet headers (left) to desired name in converted data format (right) (eg c("StudyID"= "STUDY", "Concentration" = "VALUE"))}

\item{PDlloq}{Default: 0.01}

\item{PDlloqIdentifier}{Default: c("BLQ", "<0.01", "< 0,01", "< 0.01")}

\item{PDname}{Default: 'Parasitemia Total'}

\item{study}{Default: NULL}

\item{DayOfFirstDrugAdmin}{Default: 3}

\item{centerNumber}{Default: -1}

\item{centerName}{Default: ''}

\item{visitNumber}{Default: -1}

\item{nrdoses}{Default: NULL}

\item{intervaldose}{Default: 'qd'}

\item{positiveQC}{Character vector identifying the name of positive control (must match upper or lower case as in original dataset)}
}
\value{
Imported SCID data
}
\description{
Imports SCID PK and PD data from an Excel file.
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()}
}
\author{
catalinbarcelo (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Data Preparation}
