% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{import_SCIDpkpdData_oldTAD}
\alias{import_SCIDpkpdData_oldTAD}
\title{import_SCIDpkpdData_oldTAD}
\usage{
import_SCIDpkpdData_oldTAD(
  dataFile,
  compound = NULL,
  PKsheet = "BloodLevels",
  PKrange = NULL,
  PKmapping,
  PKlloq = 1,
  PKlloqIdentifier = c("BLQ", "<LLOQ", "< LLOQ"),
  PKfactor = 0.001,
  PDsheet = "Parasitemia",
  PDrange = NULL,
  PDmapping,
  PDlloq = 0.01,
  PDlloqIdentifier = c("BLQ", "<0.01", "< 0,01", "< 0.01"),
  PDname = "Parasitemia Total",
  study = NULL,
  DayOfFirstDrugAdmin = 3,
  centerNumber = -1,
  centerName = "",
  visitNumber = -1,
  nrdoses = NULL,
  intervaldose = "qd",
  positiveQC = NULL
)
}
\arguments{
\item{compound}{Default: NULL}

\item{PKsheet}{Default: 'BloodLevels'}

\item{PKrange}{Default: NULL}

\item{PKlloq}{Default: 1}

\item{PKlloqIdentifier}{Default: c("BLQ", "<LLOQ", "< LLOQ")}

\item{PKfactor}{Default: 0.001}

\item{PDsheet}{Default: 'Parasitemia'}

\item{PDrange}{Default: NULL}

\item{PDlloq}{Default: 0.01}

\item{PDlloqIdentifier}{Default: c("BLQ", "<0.01", "< 0,01", "< 0.01")}

\item{PDname}{Default: 'Parasitemia Total'}

\item{study}{Default: NULL}

\item{DayOfFirstDrugAdmin}{Default: 3}

\item{centerNumber}{Default: -1}

\item{centerName}{Default: ''}

\item{visitNumber}{Default: -1}

\item{nrdoses}{Default: NULL}

\item{intervaldose}{Default: 'qd'}

\item{positiveQC}{Default: NULL}
}
\description{
import_SCIDpkpdData_oldTAD
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{list_Center}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()},
\code{\link{summarizeReplicates}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
catalinbarcelo (MMV)
}
\concept{Data Preparation}
