% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{import_SCIDpkpdDataCombo_GSK}
\alias{import_SCIDpkpdDataCombo_GSK}
\title{import_SCIDpkpdDataCombo_GSK}
\usage{
import_SCIDpkpdDataCombo_GSK(
  dataFile,
  Compound1 = NULL,
  Compound2 = NULL,
  PKsheets = c("Blood levels 1", "Blood levels 2"),
  PKranges = NULL,
  PKmapping,
  PKlloq = c(1, 1),
  PKlloqIdentifier = "< LLOQ",
  PKfactor = c(0.001, 0.001),
  PKfact = PKfactor,
  PDsheet = "Parasitemia",
  PDrange = NULL,
  PDmapping,
  PDlloq = 0.01,
  PDlloqIdentifier = c("BLQ", "<0.01"),
  PDname = "Parasitemia Total",
  study = NULL,
  studyID = NULL,
  DayOfFirstDrugAdmin = 3,
  centerNumber = -1,
  centerName = "",
  visitNumber = -1
)
}
\arguments{
\item{Compound1}{Default: NULL}

\item{Compound2}{Default: NULL}

\item{PKsheets}{Default: c("Blood levels 1", "Blood levels 2")}

\item{PKranges}{Default: NULL}

\item{PKlloq}{Default: c(1, 1)}

\item{PKlloqIdentifier}{Default: '< LLOQ'}

\item{PKfactor}{Default: c(0.001, 0.001)}

\item{PKfact}{Default: PKfactor}

\item{PDsheet}{Default: 'Parasitemia'}

\item{PDrange}{Default: NULL}

\item{PDlloq}{Default: 0.01}

\item{PDlloqIdentifier}{Default: c("BLQ", "<0.01")}

\item{PDname}{Default: 'Parasitemia Total'}

\item{study}{Default: NULL}

\item{studyID}{Default: NULL}

\item{DayOfFirstDrugAdmin}{Default: 3}

\item{centerNumber}{Default: -1}

\item{centerName}{Default: ''}

\item{visitNumber}{Default: -1}
}
\value{

}
\description{
import_SCIDpkpdDataCombo_GSK
}
\seealso{
Other Data Preparation: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{censorGametocyte}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{convert_catCovToTxt}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{list_Center}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()},
\code{\link{summarizeReplicates}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Data Preparation}
