% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{is.fileMMV}
\alias{is.fileMMV}
\title{is.fileMMV
Checks if 'filename' is a file}
\usage{
is.fileMMV(filename)
}
\arguments{
\item{filename}{Path to a file}
}
\value{
TRUE/FALSE
}
\description{
is.fileMMV
Checks if 'filename' is a file
}
\seealso{
\link[base:files]{base::file.exists}, \link[base:files2]{base::dir.exists}

Other General Function: 
\code{\link{get_fileNameExtension}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{General Function}
