% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-GPF.R
\name{is_GPF}
\alias{is_GPF}
\title{Check if an object inherits from S3 class GPF}
\usage{
is_GPF(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
a logical
}
\description{
Check if an object inherits from S3 class GPF
}
\seealso{
Other GPF: 
\code{\link{removeParametersWithNAValue_GPF}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{GPF}
