% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_repo.R
\name{listfunctions_MMVmalaria}
\alias{listfunctions_MMVmalaria}
\title{Build a dependency tree of function sources and targets from a package}
\usage{
listfunctions_MMVmalaria(package = "MMVmalaria", onlyExported = FALSE)
}
\arguments{
\item{package}{package name}
}
\value{
Data frame of all functions found in the package. The row "n(from)"
indicates the number of other functions where the actual function is used. The
row "n(to)" indicates the number of other functions used within the actual function.

The output carries several attributes which are needed in \code{\link{get_targets}} and
\code{\link{get_sources}}.
}
\description{
Build a dependency tree of function sources and targets from a package
}
\examples{
# Get all functions from MMVmalaria package
functions <- listfunctions_MMVmalaria()

# Copy a list of all functions to the clipboard to paste into Excel (or Google Docs)
write.table(
  data.frame(Function = names(functions)),
  file = "clipboard",
  sep = "\t",
  row.names = FALSE
)
}
