% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{load_AfricanPediatricMalariaPopulation2to5}
\alias{load_AfricanPediatricMalariaPopulation2to5}
\title{load_AfricanPediatricMalariaPopulation2to5}
\usage{
load_AfricanPediatricMalariaPopulation2to5()
}
\description{
load Malaria population demographic data only for African children 2 to 5 yrs old
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Data Preparation}
