% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{load_PediatricParameters}
\alias{load_PediatricParameters}
\title{Load Pediatric Parameters}
\usage{
load_PediatricParameters()
}
\description{
Load default pediatric parameters such as dose adjustement (\code{Fpediatric}) and
volume of blood per kg, per body weight band.
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{load_MalariaPopulation}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Data Preparation}
