% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-Simulate_VirtualTrials.R
\name{loadTrialFile}
\alias{loadTrialFile}
\title{Helper function: Load a named trial file-list for a given trial ScenID, ExpID, DoseID, TrialID}
\usage{
loadTrialFile(trialFilename)
}
\description{
Helper function: Load a named trial file-list for a given trial ScenID, ExpID, DoseID, TrialID
}
