% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_repo.R
\name{loadversion_MMVmalaria}
\alias{loadversion_MMVmalaria}
\title{Load specific version of MMVmalaria package}
\usage{
loadversion_MMVmalaria(version)
}
\arguments{
\item{version}{version number as charager, e.g., "v0.1.2".}
}
\description{
Installs the requested version from the \strong{local} MMVmalaria
git repository and loads the package.
Please make sure that your local repository
is synchronized with the github repository to be sure that
all version numbers are locally available.
}
\details{
The function requires that the user has created MMVmalaria
options via\preformatted{options(MMVmalaria = list(
  user.name = "Daniel Kaschek, IntiQuan",
  user.email = "daniel.kaschek@intiquan.com",
  local.repo = "C:/PROJTOOLS/MMVmalaria"
))
}

The general procedure to create new versions of the MMVmalaria package is:
\enumerate{
\item Modify the DESCRIPTION file and set the new version number
\item Commit all changes\preformatted{git add .
git commit -m "commit message"
}
\item Add tag to your commit with the new version number, e.g.\preformatted{git tag v0.1.2
}
\item Synchronize with github repository\preformatted{git pull
git push
git push origin --tags
}
}
}
\concept{Package Administration}
