% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{logLinTrapzMMV}
\alias{logLinTrapzMMV}
\title{log-Linear Trapez Integration Method}
\usage{
logLinTrapzMMV(x, y, logCrieria = 0.99)
}
\arguments{
\item{x}{Numerical vector such as y=f(x)}

\item{y}{Numerical vector such as y=f(x)}

\item{logCrieria}{Criteria to use the log-lin rule (Default: 0.99)}
}
\description{
log-Linear Trapez Integration Method
}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{add_FuncToEnv}()},
\code{\link{adjust_Dose}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{cbindMMV}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{combMMV}()},
\code{\link{convert_Unit}()},
\code{\link{create_PKPDsimtime}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{getDoselevel}()},
\code{\link{getMKLthreadsMMV}()},
\code{\link{get_IQRdistribution}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{get_MMVmalariaPath}()},
\code{\link{libraryMMV}()},
\code{\link{load_RData}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rbindMMV}()},
\code{\link{rectintMMV}()},
\code{\link{reverse_List}()},
\code{\link{saveActivityInfo}()},
\code{\link{saveMMV}()},
\code{\link{setMKLthreadsMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
