% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-GraphicalFunctions.R
\name{MMVggplot}
\alias{MMVggplot}
\title{ggplot functionality implementing MMV style}
\usage{
MMVggplot(
  ...,
  style = c("MMV", "MMVvpc", "IQR")[1],
  ActivityPath = NULL,
  Caption = NULL
)
}
\arguments{
\item{...}{Typical ggplot input arguments}

\item{style}{Style to choose between \code{"MMV"}, \code{"MMVvpc"} or \code{"IQR"} (Default: \code{"MMV"}).}

\item{ActivityPath}{Path of the current activity (Default: \code{NULL})}

\item{Caption}{Caption to add to the plot (Default: \code{NULL})}
}
\value{
A ggplot object with MMV standars.
}
\description{
ggplot function similar to `ggplot2::ggplot` implementing MMV style.
}
\seealso{
Other GraphicalFunctions: 
\code{\link{transform_IQRggplotToMMVggplot}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{GraphicalFunctions}
