% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMVmalaria-package.R
\docType{package}
\name{MMVmalaria-package}
\alias{MMVmalaria}
\alias{MMVmalaria-package}
\title{MMVmalaria: Collection of Malaria Functions, Models, Data and Workflows}
\description{
MMVmalaria supports and increases the efficiency, quality, and compliance of modelling activity in malaria by providing functions, models, data and workflows.
}
\keyword{internal}
