% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-NLMEmodeling.R
\name{modelSpec_MMVmalariaProject}
\alias{modelSpec_MMVmalariaProject}
\title{Get ModelSpec from a MMVmalaria Project}
\usage{
modelSpec_MMVmalariaProject(projectPath, fixParameters = FALSE)
}
\arguments{
\item{projectPath}{A character string with the path to an \code{IQRsysProject} or \code{IQRnlmeProject} folder, or a GPF file.}

\item{fixParameters}{Flag whether to fix the estimated parameters}
}
\value{
Model specification list
}
\description{
Determines model specification based on parameter estimation results of a MMVmalaria
project stored in \code{projectPath}. Estimated parameter are defined as initial guess.
}
\seealso{
Other NLME Modeling: 
\code{\link{generate_InitialParametersIQRnlme}()},
\code{\link{generate_NLMEalgorithmSettings}()},
\code{\link{testPKmodelsIQR}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{NLME Modeling}
