% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{plot_keyPDparametersMMV}
\alias{plot_keyPDparametersMMV}
\title{plot_keyPDparametersMMV}
\usage{
plot_keyPDparametersMMV(object, outputFolder, ActivityPath = NULL)
}
\arguments{
\item{ActivityPath}{Default: `NULL`}
}
\description{
plot_keyPDparametersMMV
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
