% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{plot_LinSegments}
\alias{plot_LinSegments}
\title{plot_LinSegments
Plots linear segments fitted to clearance and recrudescence phase of PD data}
\usage{
plot_LinSegments(x)
}
\arguments{
\item{x}{list with information for an individual with the following fields:
data: PD data (linear scale)
fitcl: segments (knots) fitted to clearance phase ().
fitgr: segments (knots) fitted to growth phase.}
}
\value{
ggplot object of plot
}
\description{
Parasitemia data is overlaid with linear segments for the clearance phase
and the (re-)growth phase separately.
}
\details{
fitcl is the output of fit_LinearSegments and fitgr is the output of fit_LinearModel
Parasitemia data is assumed to be on linear scale and log-transformed before plotting.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan)
}
\concept{Model-free PD assessment by linear segments}
