% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{plot_PKdataMMV}
\alias{plot_PKdataMMV}
\title{plot_PKdataMMV}
\usage{
plot_PKdataMMV(
  dataGen,
  filePath,
  PKname = NULL,
  scale_y = "log",
  facet_scales = "fixed"
)
}
\arguments{
\item{dataGen}{IQRdataGeneral object}

\item{filePath}{Character string with path to folder for saving plot}

\item{PKname}{Default: NULL}

\item{scale_y}{Default: 'log'}

\item{facet_scales}{Default: 'fixed'}
}
\value{
None, but saves plots to filePath
}
\description{
Plot PK data from an IQRdataGeneral object
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
