% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{plot_PKPDdataMMV}
\alias{plot_PKPDdataMMV}
\title{plot_PKPDdataMMV}
\usage{
plot_PKPDdataMMV(
  dataGen,
  filePath,
  PDname = "Parasitemia",
  PKname = NULL,
  HuErys = NULL,
  doseUnit = NULL,
  ActivityPath = NULL
)
}
\arguments{
\item{PDname}{Default: 'Parasitemia'}

\item{PKname}{Default: NULL}

\item{HuErys}{Default: NULL}

\item{doseUnit}{Default: NULL}

\item{ActivityPath}{Default: NULL}
}
\description{
plot_PKPDdataMMV
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
