% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{plot_PKPDdataMMVhuCh_Ratio_Gam}
\alias{plot_PKPDdataMMVhuCh_Ratio_Gam}
\title{Plot PKPD data available from a Human Challenge study in "standard" MMV format, suitable for exploration and reporting.
plots ratio of Individual parasite data over gametocyte counts, stratified by treatment.}
\usage{
plot_PKPDdataMMVhuCh_Ratio_Gam(
  dataGen,
  filePath,
  PDname = "Parasitemia",
  Gametos = "Parasitemia Gametocytes",
  rescuename = "Riamet"
)
}
\arguments{
\item{dataGen}{: Object containing data to plot in \code{IQRDataGeneral} format}

\item{filePath}{Character string with path to folder for saving plot}

\item{PDname}{: Name of PD observations in the NAME column of IQRDataGeneral. Default: 'Parasitemia'}

\item{Gametos}{: Name of Gametocyte observations in the NAME column of IQRDataGeneral. Default: 'Parasitemia Gametocytes'}

\item{rescuename}{: Name of observations in the NAME column that correspond to Rescue medication doses.}
}
\value{
No return; Figures saved as .png to destination specified in \code{filePath}
}
\description{
Plot PKPD data available from a Human Challenge study in "standard" MMV format, suitable for exploration and reporting.
plots ratio of Individual parasite data over gametocyte counts, stratified by treatment.
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV), Karsten Kuritz #' (IntiQuan)
}
\concept{DataExploration}
