% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{plot_PKPDhuChCombo}
\alias{plot_PKPDhuChCombo}
\title{plot_PKPDhuChCombo}
\usage{
plot_PKPDhuChCombo(
  dataGen,
  filePath,
  PDname = "Parasitemia",
  Gametos = "Parasitemia Gametocytes",
  PKname = NULL,
  TRTNAME = "TRTNAME",
  ByUSUBJID = TRUE,
  ActivityPath = NULL
)
}
\arguments{
\item{PDname}{Default: 'Parasitemia'}

\item{Gametos}{Default: 'Parasitemia Gametocytes'}

\item{PKname}{Default: NULL}

\item{TRTNAME}{Column name of treatment group to use}

\item{ByUSUBJID}{If the plot should be split by subject (Default: TRUE)}

\item{ActivityPath}{Path of the current activity (Default: NULL)}
}
\value{

}
\description{
plot_PKPDhuChCombo
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
