% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{plot_PKPDscidCombo}
\alias{plot_PKPDscidCombo}
\title{plot_PKPDscidCombo}
\usage{
plot_PKPDscidCombo(
  dataGen,
  filePath,
  PDname = "Parasitemia",
  PKname = NULL,
  HuErys = NULL,
  CompoundList = NULL,
  ActivityPath = NULL
)
}
\arguments{
\item{dataGen}{IQRdataGeneral object}

\item{filePath}{Character string with path to folder for saving plot}

\item{PDname}{Default: 'Parasitemia'}

\item{PKname}{Default: NULL}

\item{HuErys}{Default: NULL}

\item{CompoundList}{Default: NULL}

\item{ActivityPath}{Path of the current activity (Default: NULL)}
}
\value{
List of plots
}
\description{
Plot PK and PD data from an IQRdataGeneral object
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
