% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{plot_tMICextrapolation}
\alias{plot_tMICextrapolation}
\title{plot_tMICextrapolation
Visualize determination of time of MIC by linear segments}
\usage{
plot_tMICextrapolation(
  EXTRA,
  FIT,
  data,
  xlabel = "x",
  ylabel = "y",
  title = "Determination of time at MIC by linear extrapolation",
  FLAGlogtransData = TRUE,
  ActivityPath = NULL
)
}
\arguments{
\item{EXTRA}{Data frame defining the  that define the time of MIC
(output of \link{define_tMIClinearSegments}).
Phase column need to be defined (see Details).}

\item{FIT}{Data frame defining linear segments fitted
to clearance and re-growth data
(combined output of \link{fit_LinearModel} and \link{fit_LinearSegments}).
Phase column need to be defined (see Details).}

\item{data}{Data frame with parasitemia data (expects NT, DV, TIME and CENS columns).}

\item{xlabel}{Character (string) to be printed on x axis.}

\item{ylabel}{Character (string) to be printed on y axis.}

\item{title}{Character (string) to be printed as figure title.}

\item{FLAGlogtransData}{Flag whether data needs to be logtransformed before plotting.}
}
\value{
ggplot object
}
\description{
Plots parasitemia data overlaid with linear segments fitted to
clearance and re-growth phase and the interpolation of the
time of MIC.
}
\details{
Segments are expected to be defined on logtransformed parasitemia data.
Data may need to be transformed if given on linear scale by setting
FLAGlogtransData accordingly.
Phases are differentiated by 'Phase' column with KILL for clearance phase,
RECRUD for re-growth phase, and "MIC" for the interpolation between them.
Individuals are plotted to separate panels.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolationRange}()},
\code{\link{simulate_indivPK}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Model-free PD assessment by linear segments}
