% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{plot_tMICextrapolationRange}
\alias{plot_tMICextrapolationRange}
\title{plot_tMICextrapolationRange
Visualize determination of possible time range for MIC}
\usage{
plot_tMICextrapolationRange(
  EXTRA,
  FIT,
  data,
  xlabel = "x",
  ylabel = "y",
  title = "Determination of time at MIC min/max by linear extrapolation",
  FLAGlogtransData = TRUE,
  ActivityPath = NULL
)
}
\arguments{
\item{EXTRA}{Data frame defining intersect that define the time of MIC
(combined output of \link{define_tLLOQ} applied to clearance and regrowth data).
Phase column need to be defined (see Details).}

\item{FIT}{Data frame defining linear segments fitted
to clearance and re-growth data
(combined output of \link{fit_LinearModel} and \link{fit_LinearSegments}).
Phase column need to be defined (see Details).}

\item{data}{Data frame with parasitemia data (expects NT, DV, TIME and CENS columns).}

\item{xlabel}{Character (string) to be printed on x axis.}

\item{ylabel}{Character (string) to be printed on y axis.}

\item{title}{Character (string) to be printed as figure title.}

\item{FLAGlogtransData}{Flag whether data needs to be logtransformed before plotting.}
}
\value{
ggplot object
}
\description{
Plots parasitemia data overlaid with linear segments fitted to
clearance and re-growth phase and the extrapolation to the times
at which the LLOQ is crosses while killing or re-growing of parasites
}
\details{
Segments are expected to be defined on logtransformed parasitemia data.
Data may need to be transformed if given on linear scale by setting
FLAGlogtransData accordingly.
Phases are differentiated by 'Phase' column with KILL for clearance phase,
RECRUD for re-growth phase, and "MICmax" for the extrapolation of
the clearance phase to the LLOQ that is an upper limit for the MIC and
"MICmin" for the extrapolation of the re-growth phase to the LLOQ that
is a lower limit for the MIC.
Individuals are plotted to separate panels.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{simulate_indivPK}()}
}
\author{
Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Model-free PD assessment by linear segments}
