% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{plotSummaryComparisonContinuous}
\alias{plotSummaryComparisonContinuous}
\title{Plotting Comparison of Summarized Data for Continuous Variables}
\usage{
plotSummaryComparisonContinuous(
  dataSim,
  dataObs = NULL,
  CIlevel = 90,
  compCategory = "TRTNAME",
  variableCol = "Variable"
)
}
\arguments{
\item{dataSim}{Data.frame containing summarized simulation data.}

\item{dataObs}{Optional data.frame containing summarized observation data.}

\item{CIlevel}{Numeric value representing the confidence interval used for summarizing simulated percentiles (default is 90).}

\item{compCategory}{String specifying the column to be used for comparison (default is "TRTNAME").}

\item{variableCol}{String specifying the column name to be used as variables (default is "Variable").}
}
\value{
A list of ggplot objects comparing summarized continuous variables from simulation and observation.
}
\description{
Generates a list of ggplot objects where summarized simulations and
observations are compared visually. Summarized input is obtained from
applying \code{summaryByTrial} and \code{summarAcrossTrials} to data from VPCs.
Comparison is done per \code{TRTNAME} by default. A list of ggplot objects is returned with one plot per unique variable.
}
\details{
This function creates a visual comparison of summarized data for continuous variables. It generates a ggplot object for each unique variable in the dataset.
The summarized input for simulations is obtained using the \code{summaryByTrial} and \code{summarAcrossTrials} functions. The comparison is performed per \code{TRTNAME} by default, but this can be customized
using the \code{compCategory} parameter.
}
\examples{
\dontrun{
# Example usage:
# prepare categories for dataSim and dataObs
dataObs <- dataSim <- expand.grid(
  TRTNAME =  c("Treatment A", "Treatment B"), 
  Variable =  c("Var1", "Var2"), 
  Metric = c("Metric1", "Metric2", "Metric3"))

# add values for dataSim
dataSim$`CI Median` <- rep(c(10,20,30), each = 4) + rnorm(12)
dataSim$`CI Low` <- dataSim$`CI Median` -3  - runif(12)
dataSim$`CI High` <- dataSim$`CI Median` +  3 + runif(12)

# add values for dataObs
dataObs$Value <- rep(c(10,20,30), each = 4) + rnorm(12)

# plot
ggplots <- plotSummaryComparisonContinuous(dataSim, dataObs)

# Display the first plot
}

}
\seealso{
\code{\link[=plotSummaryComparisonBinary]{plotSummaryComparisonBinary()}}

Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Karsten Kuritz (IntiQuan)
}
\concept{Model Assessment}
