% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{plotVPC_MMVdataVPC}
\alias{plotVPC_MMVdataVPC}
\title{MMV wrapper for plotVPC_IQRdataVPC}
\usage{
plotVPC_MMVdataVPC(
  dataVPC,
  stratifyBy = NULL,
  filename = NULL,
  FLAGlogY = FALSE,
  FLAGuseTAD = FALSE,
  FLAGdataPlotOnly = FALSE,
  FLAGnoDataPoints = FALSE,
  FLAGpc = FALSE,
  FLAGplotBins = FALSE,
  FLAGplotN = FALSE,
  title = NULL,
  subtitle = NULL,
  BIN.column = NULL,
  BIN.breaks = NULL,
  BIN.groupsize = NULL,
  BIN.lambda = 1,
  BIN.resolution = 0.1,
  CIlevel = 95,
  percentiles = c(5, 95),
  alphaDataPoints = 0.4,
  alphaPredInter = 0.4,
  ActivityPath = NULL,
  Caption = NULL,
  setting = NULL
)
}
\arguments{
\item{stratifyBy}{Default: \code{NULL}}

\item{filename}{Default: \code{NULL}}

\item{FLAGlogY}{Default: \code{FALSE}}

\item{FLAGuseTAD}{Default: \code{FALSE}}

\item{FLAGdataPlotOnly}{Default: \code{FALSE}}

\item{FLAGnoDataPoints}{Default: \code{FALSE}}

\item{FLAGpc}{Default: \code{FALSE}}

\item{FLAGplotBins}{Default: \code{FALSE}}

\item{FLAGplotN}{Default: \code{FALSE}}

\item{title}{Default: \code{NULL}}

\item{subtitle}{Default: \code{NULL}}

\item{BIN.column}{Default: \code{NULL}}

\item{BIN.breaks}{Default: \code{NULL}}

\item{BIN.groupsize}{Default: \code{NULL}}

\item{BIN.lambda}{Default: \code{NULL}}

\item{BIN.resolution}{Default: \code{NULL}}

\item{CIlevel}{Default: \code{NULL}}

\item{percentiles}{Default: \code{NULL}}

\item{alphaDataPoints}{Default: \code{NULL}}

\item{alphaPredInter}{Default: \code{NULL}}

\item{ActivityPath}{Default: \code{NULL}}

\item{Caption}{Default: \code{NULL}}

\item{setting}{Default: \code{NULL}}
}
\description{
Upgrades the corresponding function in IQRtools. Check its documentation for a description of the arguments.
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
