% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{predict_HumanChallenge}
\alias{predict_HumanChallenge}
\title{predict_HumanChallenge}
\usage{
predict_HumanChallenge(
  PKeventTable,
  PDmodelFolder,
  data,
  GRhuman = NULL,
  model = NULL,
  CureThreshold = log(1/5000)
)
}
\arguments{
\item{PKeventTable}{IQReventTable object specifying PK model, see \code{IQRtools::IQReventTable}}

\item{PDmodelFolder}{Path to IQRnlme project containing structural model and pre-clinical PD parameters}

\item{data}{IQRdataGeneral of human challenge data}

\item{GRhuman}{Numeric value of growth human: If NULL, will use "GR" from \code{PDmodelFolder}. Default: \code{NULL}}

\item{model}{Path to IQRmodel.txt, if NULL, will use "model.txt" from \code{PDmodelFolder}. Default: \code{NULL}}

\item{CureThreshold}{Numeric, value of parasitaemia to treat as cure threshold. Default: log(1/5000)}
}
\value{
List including simulation data and figure output
}
\description{
Produce a simulated prediction of, and visual output for, a human challenge study
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
