% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-PredictHumanDoses.R
\name{predictDose_PRRtot}
\alias{predictDose_PRRtot}
\title{Predict doses for a target PRRtot for a number of sampled virtual trial populations}
\usage{
predictDose_PRRtot(
  targetPRRtot,
  doseInterval,
  modelFile,
  projectPath,
  simtime,
  Tk0 = NULL,
  nbrDoses = 1,
  timeInterval = 24,
  args_IQRdosing = list(TIME = simtime[1], ADM = 1, AMT = NA_real_, ADDL = nbrDoses -
    1, II = timeInterval),
  N_Trial,
  N_SubjperTrial,
  covariates = NULL,
  DOSEcovariate = NULL,
  Fpediatric = NULL,
  regressorExtra = character(0),
  FLAG_SAMPLE = 1,
  percentiles = c(5, 50, 95),
  CIlevel = 90,
  addArgs_sim_IQRmodel = NULL,
  addArgs_uniroot = list(tol = 1),
  fun_EvaluateCriterion = "evaluateDoseCriterion_PRRtot",
  Nparallel = 1,
  seed = 1234567,
  FLAG_BROWSER1 = FALSE
)
}
\arguments{
\item{targetPRRtot}{a scalar specifying the target PRRtot.}

\item{doseInterval}{a numeric vector of length 2 denoting the dosing interval to search within.}

\item{modelFile}{Path to a PKPD model to be used for the simulations or a parsed IQRmodel object.}

\item{projectPath}{Path to a PK/PD \code{IQRprojectNLME} folder or a path to a GPF.xlsx file or a GPF object.}

\item{simtime}{a sorted numeric vector denoting the simulation times (passed to \code{\link{IQRtools::sim_IQRmodel}}).}

\item{Tk0}{A named character vector (default: NULL). Example : c(INPUT1 = "Tk0", INPUT2 = "Tk0_2"), meaning that
the IQRdosing object for each simulation would be adjusted to have TINF = set to the model parameter Tk0 for ADM = 1.}

\item{nbrDoses}{an integer (default 1), denoting the number of doses to simulate}

\item{timeInterval}{a numeric (default 24), denoting the time interval between doses (in the model time units).}

\item{args_IQRdosing}{a named list (default NULL) with arguments to be passed to \code{\link{IQRtools::IQRdosing}}. For flexibility, this
can be used to overwrite some of the otherwise set by default arguments to IQRdosing, such as \code{TIME}, \code{ADM}, \code{ADDL}, \code{II}
(advanced use only). Default: \code{list(TIME = simtime[1], ADM = 1, AMT = NA_real_, ADDL = nbrDoses - 1, II = timeInterval)}.}

\item{N_Trial}{Number of trials to be simulated.}

\item{N_SubjperTrial}{Number of subjects per trial.}

\item{covariates}{A dataframe with covariate values, and optionally regressors (see also arugment regressorExtra).}

\item{DOSEcovariate}{Name of covariates describing dose. Example: \code{c("INPUT1"="DOSELEVEL")}. Default: NULL.}

\item{Fpediatric}{Name of covariate describing pediatric dose. Example: \code{"Fpediatric"}. Default: NULL.}

\item{regressorExtra}{A character vector indicating the names of regressors to take from the covariate data,
instead of sampling them as model parameters. Extra regressors not present in \code{projectPath} but present in \code{modelFile}.}

\item{FLAG_SAMPLE}{An integer (by default 1) specifying how model parameters are sampled for each virtual trial population.
Possible values:\tabular{lcccl}{
   FLAG_SAMPLE \tab Sample from uncertainty distribution \tab Sample covariate data and apply formulas \tab Sample from IIV distribution \tab Notes \cr
   0 \tab No \tab Yes \tab Yes \tab Use point estimates as pop. parameter values \cr
   1 (default) \tab Yes \tab Yes \tab Yes \tab  \cr
   2 \tab Yes \tab No \tab No \tab Forces N_SubjperTrial = 1 \cr
   3 \tab No \tab Yes \tab No \tab Use point estimates as pop. parameter values; \cr
   4 \tab Yes \tab Yes \tab No \tab  \cr
}}

\item{percentiles}{Percentiles to estimate in each trial (Default: \code{c(5,50,95)}).}

\item{CIlevel}{Confidence interval to estimate for each percentile (Default: 90),
based on the simulated trials}

\item{addArgs_sim_IQRmodel}{a named list (default NULL) with additional arguments to be passed to \code{\link{IQRtools::sim_IQRmodel}}.}

\item{addArgs_uniroot}{a named list (default \code{list(tol = .1)}) with additional arguments to be passed to \code{\link{stats::uniroot}}.}

\item{fun_EvaluateCriterion}{a function object or a character string denoting the name of a function accessible
from the calling (parent) environment. This function should have the following synopsis:
\code{function(sim_results, parameters) { ... }}, where \code{sim_results} is a data.frame resulting from a
previous call to \code{\link{IQRtools::sim_IQRmodel}} and \code{parameters} is a data.frame with a single row
and numeric columns named as the model parameters; the function should return a numeric value. If
\code{fun_EvaluateCriterion} is a character string, the function is searched for via
\code{mget(fun_EvaluateCriterion, mode = "function", ifnotfound = list(NULL), inherits = TRUE)[[fun_EvaluateCriterion]]}.
An error will be thrown if no such function is found.}

\item{Nparallel}{Number of cores to use for parallel simulations (Default: 1).}

\item{seed}{An integer specifying a seed for the random generator (Default: 1234567).}

\item{FLAG_BROWSER1}{A logical (default: \code{FALSE}), indicating whether a browser should be launched for each individual simulation (debug purpose only).}
}
\value{
Data frame with coumns Variable, Metric, CI Low, CI Median, CI High, CI Level. The Variable column equals PRRtot and
the Metric values correspond to the different values in the percentiles argument. The other columns correspond to the
predicted doses for each percentile in Metric.
}
\description{
Predict doses for a target PRRtot for a number of sampled virtual trial populations
}
\author{
Venelin Mitov, IntiQuan
}
