% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Reduce character to the non-repeating entries}
\usage{
remove_duplicates(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
character of the same length as x.
}
\description{
Entries which are the same as the preceding one are are replaced by empty character.
The inverse of insert_duplicates().
}
