% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{remove_OneCovariateToModelSpec}
\alias{remove_OneCovariateToModelSpec}
\title{remove_OneCovariateToModelSpec}
\usage{
remove_OneCovariateToModelSpec(
  modelSpec,
  covariateModelToRemove,
  COVcentering = NULL
)
}
\arguments{
\item{modelSpec}{modelSpec object in format generated by \code{modelSpec_IQRest}}

\item{COVcentering}{Named vector defining centering value for covariates i.e. COVcentering = c(WT0=55)}

\item{covariateModelToAdd}{covariateModel and values to remove, see \code{modelSpec_IQRest}}
}
\value{
A list containing the IQR model spec, see \code{?modelSpec_IQRest}
}
\description{
Remove a single covariate model to a modelSpec object.
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
