% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{reportText_convertListToText}
\alias{reportText_convertListToText}
\title{reportText_convertListToText}
\usage{
reportText_convertListToText(list, sepWord)
}
\arguments{
\item{list}{list to convert}

\item{sepWord}{Character string to separate elements of list}
}
\value{
Combined character string
}
\description{
Convert list input to text string
}
\seealso{
Other Reporting: 
\code{\link{generate_EstimateTableForReport}()},
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_PKmodelHypos}()},
\code{\link{table_PKsummaryCompare}()},
\code{\link{table_PKsummaryResults}()},
\code{\link{table_covariateSummaryResults}()},
\code{\link{table_strToLaTeX}()}
}
\author{
Unknown
}
\concept{Reporting}
