% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{reportText_PKmodelHypos}
\alias{reportText_PKmodelHypos}
\title{reportText_PKmodelHypos}
\usage{
reportText_PKmodelHypos(PKmodelTest)
}
\arguments{
\item{PKmodelTest}{list containing description of PK models to test}
}
\value{
Character string of PK models to test
}
\description{
Generate report text from list of PK models to test
}
\seealso{
Other Reporting: 
\code{\link{generate_EstimateTableForReport}()},
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_convertListToText}()},
\code{\link{table_PKsummaryCompare}()},
\code{\link{table_PKsummaryResults}()},
\code{\link{table_covariateSummaryResults}()},
\code{\link{table_strToLaTeX}()}
}
\author{
Unknown
}
\concept{Reporting}
