% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{reverse_List}
\alias{reverse_List}
\title{Reverse a List}
\usage{
reverse_List(list, simplify = FALSE)
}
\arguments{
\item{list}{List to reverse}

\item{simplify}{Logical or character string; should the result be simplified
to a vector, matrix or higher dimensional array if possible? For \code{sapply}
it must be named and not abbreviated. The default value, \code{TRUE}, returns
a vector or matrix if appropriate, whereas if \code{simplify = "array"} the result
may be an array of "rank" (\code{=length(dim(.))}) one higher than the result of
\code{FUN(X[[i]])} (Default: \code{FALSE})}
}
\value{
Reversed list
}
\description{
Reverse a list, where all elements of the list have the same length and
similar object:
If the input is
     \code{a = list(list(Obj1 = data.frame(X = c(1,2), Y = c(1,4))),
                    list(Obj1 = data.frame(X = c(3,4), Y = c(9,16))))}
the output will be
     \code{list(Obj1 = list(data.frame(X = c(1,2,3,4), Y = c(1,4,9,16)),
                            data.frame(X = c(3,4), Y = c(1,4))))}
}
\examples{
a = list(list(Obj1 = data.frame(X = c(1,2), Y = c(1,4))),
         list(Obj1 = data.frame(X = c(3,4), Y = c(9,16))))
reverse_List(a)

}
\seealso{
Other General Functions: 
\code{\link{IQRtableToDataFrame}()},
\code{\link{add_FuncToEnv}()},
\code{\link{adjust_Dose}()},
\code{\link{aux_CommonSubPath}()},
\code{\link{aux_addNamesToPercentiles}()},
\code{\link{aux_constructCIpercentiles}()},
\code{\link{aux_createUSUBJID}()},
\code{\link{aux_formatErrorName}()},
\code{\link{aux_removeEscapeChar}()},
\code{\link{cbindMMV}()},
\code{\link{clopperPearsonMMV}()},
\code{\link{combMMV}()},
\code{\link{convert_Unit}()},
\code{\link{create_PKPDsimtime}()},
\code{\link{file.copyMMV}()},
\code{\link{find_MinMMV}()},
\code{\link{getDoselevel}()},
\code{\link{getMKLthreadsMMV}()},
\code{\link{get_ActivityPath}()},
\code{\link{get_IQRdistribution}()},
\code{\link{get_IXGDFtoRemove}()},
\code{\link{get_MMVmalariaPath}()},
\code{\link{libraryMMV}()},
\code{\link{load_RData}()},
\code{\link{logLinTrapzMMV}()},
\code{\link{newtonRaphson.Function}()},
\code{\link{newtonRaphson.Vector}()},
\code{\link{rbindMMV}()},
\code{\link{rectintMMV}()},
\code{\link{saveActivityInfo}()},
\code{\link{saveMMV}()},
\code{\link{setMKLthreadsMMV}()},
\code{\link{simpsonMMV}()},
\code{\link{thisFile}()},
\code{\link{trapzMMV}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
