% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{sample_CustomParameters}
\alias{sample_CustomParameters}
\title{sample_CustomParameters}
\usage{
sample_CustomParameters(
  parameters.POP,
  parameters.IIV,
  parameters.IIVcor = NULL,
  parameters.DIST,
  Nsamples
)
}
\arguments{
\item{parameters.POP}{Named vector containing population parameter values}

\item{parameters.IIV}{Named vector containing IIV values}

\item{parameters.IIVcor}{Named vector containing correlations between IIVs. Will be parsed by \code{getCorrParNames}. Default: NULL}

\item{parameters.DIST}{Vector containing 'N', 'L' or 'G', representing normal, log-normal and Logit distribution, respectively.}

\item{Nsamples}{Numeric integer denoting number of individuals to sample}
}
\value{
data.frame wth \code{Nsamples} rows and one column per parameter, providing sampled parameter values.
}
\description{
Sample individual parameter values from provided custom information
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
