% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{sample_MMVmalariaProject}
\alias{sample_MMVmalariaProject}
\title{sample_MMVmalariaProject}
\usage{
sample_MMVmalariaProject(
  projectPath,
  Nsamples = NULL,
  FLAG_SAMPLE = 0,
  covariates = NULL,
  FLAGid = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{projectPath}{A character string with the path to an IQRsysProject or IQRnlmeProject folder, or a GPF file.}

\item{Nsamples}{Numeric integer denoting number of individuals to sample, default: NULL}

\item{FLAG_SAMPLE}{Numeric, value for sample flag, see \code{?IQRtools::sampledIndParamValues}. Default: 0}

\item{covariates}{data.frame describing patient or other subject covariate and regressor data, see \code{IQRtools::sampleIndParamValues}. Default: \code{NULL}}

\item{FLAGid}{Default: FALSE}

\item{verbose}{Default: TRUE}
}
\value{
Named list containing parameter values, see \code{?IQRtools::sampledIndParamValues}
}
\description{
Sample individual parameter values from an MMVmalaria project
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
