% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-MMVnlmeParamSamplingXLS.R
\name{sampleIDs_MMVmalariaXLS}
\alias{sampleIDs_MMVmalariaXLS}
\title{Sample or subset records from a matrix or a data.frame}
\usage{
sampleIDs_MMVmalariaXLS(data, Nsamples)
}
\arguments{
\item{data}{A data.frame describing patient or other subject covariate and regressor data. If \code{nrow(data)==Nsamples}, then all rows in data are returned (in their original order) in the resulting data.frame; if \code{nrow(data)>Nsamples}, sampling without replacement from the rows in data is done; if \code{nrow(data)<Nsamples} then sampling with replacement is done.}

\item{Nsamples}{Integer denoting the number of individuals in a population. Note: in the case of \code{FLAG_SAMPLE=2}, this value is used to specify the number of populations (see \code{Npop}). Default: \code{if(is.IQRnlmeParamSpec(obj)) obj$Nsamples else 1L}.}
}
\value{
a data.frame. If nrow(data)==Nsamples, then all rows in data
are included (in their original order); if nrow(data)>Nsamples, sampling
without replacement from the rows in data is done; if nrow(data)<Nsamples
then sampling with replacement is done. The returned data.frame has a
column ID set to the integers 1:Nsamples. If the original data object
did have a column ID, this column is copied as a column ID0 in the returned
data.frame.
}
\description{
Sample or subset records from a matrix or a data.frame
}
\examples{
\dontrun{
#See the examples for sampleIndParamValues_MMVmalariaXLS:
?sampleIndParamValues_MMVmalariaXLS
}
}
\seealso{
Other functions in the NLME parameter sampling API: 
\code{\link{calcIndParamValues_MMVmalariaXLS}()},
\code{\link{calcTypicalIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleParamFromUncertainty_MMVmalariaXLS}()},
\code{\link{sampleRandomEffects_MMVmalariaXLS}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{functions in the NLME parameter sampling API}
