% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-MMVnlmeParamSamplingXLS.R
\name{sampleIndParamValues_MMVmalariaXLS}
\alias{sampleIndParamValues_MMVmalariaXLS}
\title{Sample individual parameter values, based on a NLME parameter
specification}
\usage{
sampleIndParamValues_MMVmalariaXLS(
  spec,
  data = NULL,
  Nsamples = 1,
  Npop = if (FLAG_SAMPLE == 2) Nsamples else 1L,
  FLAG_SAMPLE = 0
)
}
\arguments{
\item{spec}{A filename (character string) denoting the path to a GPF excel file, or a GPF object, or an IQRnlmeParamSpec object.}

\item{data}{a data.frame describing patient or other subject covariate and regressor data.
If nrow(data)==Nsamples, then all rows in data
are returned (in their original order) in the resulting data.frame; if
nrow(data)>Nsamples, sampling without replacement from the rows in data
is done; if nrow(data)<Nsamples then sampling with replacement is done.}

\item{Nsamples}{integer denoting the number of individuals in a
population. Note: in the case of \code{FLAG_SAMPLE=2}, this value is used to specify
the number of populations (see \code{Npop}). Default:
\code{if(is.IQRnlmeParamSpec(obj)) obj$Nsamples else 1L}.}

\item{Npop}{integer denoting the number of populations to sample. Default:
\code{if(is.IQRnlmeParamSpec(obj)) obj$Npop else if(FLAG_SAMPLE==2) Nsamples else 1L}.}

\item{FLAG_SAMPLE}{Flag indicating the type of sampling:
\tabular{ll}{
FLAG_SAMPLE       \tab Meaning \cr
===========       \tab =================================================== \cr
0                 \tab Use point estimates of population parameters (do not consider uncertainty) and sample Nsample
individual patients based on these. Covariates considered if defined by user and used in model.
Please note: population parameters do not take covariates into account!\cr
1                 \tab Sample single set of population parameters from uncertainty distribution and sample Nsample
individual patient parameters based on these. Covariates considered if defined by user and
used in model. Please note: population parameters do not take covariates into account!\cr
2                 \tab Sample Nsample sets of population parameters from uncertainty distribution.
Do not sample from variability distribution and do not take into account covariates (even
if user specified).\cr
3                 \tab Use point estimates of population parameters (do not consider uncertainty)
Return Nsamples sets of population parameters with covariates taken into account.\cr
4                 \tab Sample single set of population parameters from uncertainty distribution
Return Nsamples sets of population parameters with covariates taken into account.\cr
5                 \tab Use point estimates of population parameters (do not consider uncertainty) and sample Nsample
individual patients based on these. Covariates considered if defined by user and used in model.
Population parameters, typical individual parameters (no IIV, but considering covariates),
and individual parameters are returned.\cr
6                 \tab Sample single set of population parameters from uncertainty distribution and sample Nsample
individual patient parameters based on these. Covariates considered if defined by user and used in model.
Sampled population parameters, typical individual parameters (no IIV, but considering covariates),
and individual parameters are returned.\cr
7                 \tab Same as FLAG_SAMPLE=0 but obtaining the empirical random effect covariance matrix from the post-hoc ETAs.
Only used for IQRsysProject. If IQRnlmeProject then fallback to flag 0.\cr
8                 \tab Same as FLAG_SAMPLE=1 but obtaining the empirical random effect covariance matrix from the post-hoc ETAs.\cr
Only used for IQRsysProject. If IQRnlmeProject then fallback to flag 1.\cr
}}
}
\description{
Sample individual parameter values, based on a NLME parameter
specification
}
\details{
If \code{spec} is an IQRnlmeParamSpec object and \code{Npop} is specified,
the specified argument Npop overwrites the \code{spec$Npop}.
Default: \code{if(is.IQRnlmeParamSpec(spec)) spec$Npop else 1L}.
}
\examples{
\dontrun{
set.seed(1)
covrtData <- data.frame(
  ID = seq_len(4),
  WT0 = rlnorm(4, 3, sdlog = 0.2),
  SEX = sample(c(0, 1), 4, replace = TRUE))

set.seed(1)
listResultsSampling <- sampleIndParamValues_MMVmalariaXLS(
  fileXLS,
  Nsamples = 10,
  FLAG_SAMPLE = 1,
  data = covrtData)

# create a IQRnlmeParamSpec object for future use (not suitable for reading by humans):
spec <- specifyParamSampling(fileXLS, Nsamples = 10, FLAG_SAMPLE = 1)
is.IQRnlmeParamSpec(spec)

set.seed(1)

popParamValues<- sampleParamFromUncertainty_MMVmalariaXLS(spec)
# should be all 0s
listResultsSampling$popParamValues - popParamValues

sampledData <- sampleIDs_MMVmalariaXLS(covrtData, spec$Nsamples)
# should be all 0s
listResultsSampling$sampledData - sampledData

typicalIndParamValues <- calcTypicalIndParamValues_MMVmalariaXLS(spec, popParamValues, sampledData)
# should be all 0s
listResultsSampling$typicalIndParamValues - typicalIndParamValues

randomEffects <- sampleRandomEffects_MMVmalariaXLS(spec, popParamValues)
# should be all 0s
listResultsSampling$randomEffects - randomEffects

indParamValues <- calcIndParamValues_MMVmalariaXLS(spec, typicalIndParamValues, randomEffects)
# should be all 0s
listResultsSampling$indParamValues - indParamValues
}

}
\seealso{
Other functions in the NLME parameter sampling API: 
\code{\link{calcIndParamValues_MMVmalariaXLS}()},
\code{\link{calcTypicalIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleIDs_MMVmalariaXLS}()},
\code{\link{sampleParamFromUncertainty_MMVmalariaXLS}()},
\code{\link{sampleRandomEffects_MMVmalariaXLS}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{functions in the NLME parameter sampling API}
