% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-MMVnlmeParamSamplingXLS.R
\name{sampleParamFromUncertainty_MMVmalariaXLS}
\alias{sampleParamFromUncertainty_MMVmalariaXLS}
\title{Sample NLME parameters from their uncertainty distribution}
\usage{
sampleParamFromUncertainty_MMVmalariaXLS(
  spec,
  Npop = if (is_IQRnlmeParamSpec(spec)) spec$Npop else 1L
)
}
\arguments{
\item{spec}{A filename (character string) denoting the path to a GPF excel file, or a GPF object, or an IQRnlmeParamSpec object.}

\item{Npop}{integer denoting the number of populations to sample. Default:
\code{if(is.IQRnlmeParamSpec(obj)) obj$Npop else if(FLAG_SAMPLE==2) Nsamples else 1L}.}
}
\value{
a data.frame of \code{Npop} rows with columns \code{ID.POP} and parameter
value columns as follows:
\describe{
\item{X,Y...: }{reference population values for model parameters X,Y...}
\item{omega(X),omega(Y)...: }{IIV population values for model parameters X,Y,...}
\item{corr(X,Y),...: }{IIV correlation population values for model parameter pairs.}
\item{beta_X(COVRT1),beta_Y(COVRT2),...: }{Covariate coefficient population values.}
\item{error_PROP1,...: }{Residual error population parameters.}
}
}
\description{
Sample NLME parameters from their uncertainty distribution
}
\examples{
\dontrun{
#See the examples for sampleIndParamValues_MMVmalariaXLS:
?sampleIndParamValues_MMVmalariaXLS
}
}
\seealso{
Other functions in the NLME parameter sampling API: 
\code{\link{calcIndParamValues_MMVmalariaXLS}()},
\code{\link{calcTypicalIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleIDs_MMVmalariaXLS}()},
\code{\link{sampleIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleRandomEffects_MMVmalariaXLS}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{functions in the NLME parameter sampling API}
