% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-MMVnlmeParamSamplingXLS.R
\name{sampleRandomEffects_MMVmalariaXLS}
\alias{sampleRandomEffects_MMVmalariaXLS}
\title{Sample NLME random effects given population parameter values
describing the inter-individual variability distribution}
\usage{
sampleRandomEffects_MMVmalariaXLS(
  spec,
  iivPopParamValues,
  Nsamples = if (is_IQRnlmeParamSpec(spec)) spec$Nsamples else 1
)
}
\arguments{
\item{spec}{A filename (character string) denoting the path to a GPF excel file, or a GPF object, or an IQRnlmeParamSpec object.}

\item{iivPopParamValues}{a data.frame with population parameter values
describing the IIV distribution. The number of rows in this data.frame
defines the number of populations for which random effects are to be
sampled.}
}
\value{
a data.frame of \code{Nsamples*nrow(iivPopParamValues)} rows and
columns \code{ID.POP}, \code{ID} and a column for each model parameter
defined in \code{spec}.
}
\description{
Sample NLME random effects given population parameter values
describing the inter-individual variability distribution
}
\examples{
\dontrun{
#See the examples for sampleIndParamValues_MMVmalariaXLS:
?sampleIndParamValues_MMVmalariaXLS
}
}
\seealso{
Other functions in the NLME parameter sampling API: 
\code{\link{calcIndParamValues_MMVmalariaXLS}()},
\code{\link{calcTypicalIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleIDs_MMVmalariaXLS}()},
\code{\link{sampleIndParamValues_MMVmalariaXLS}()},
\code{\link{sampleParamFromUncertainty_MMVmalariaXLS}()}
}
\author{
Venelin Mitov (IntiQuan)
}
\concept{functions in the NLME parameter sampling API}
