% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ActivityMonitoring.R
\name{saveActivityInfo}
\alias{saveActivityInfo}
\title{saveActivityInfo
Save Activity Information}
\usage{
saveActivityInfo(list, ActivityInfoFile = "../01-Data/S001-ActivityInfo")
}
\arguments{
\item{list}{A character vector containing the names of objects to be saved.}

\item{ActivityInfoFile}{The name of the file where the activity information will be saved.}
}
\description{
\link{saveActivityInfo} save a 'list' activity information into an R object
to the specified file 'ActivityInfoFile'.
}
\seealso{
\link[base:save]{base::save}, \link[MMVbase:saveMMV]{MMVbase::saveMMV}

Other General Functions: 
\code{\link{get_MMVmalariaPath}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
