% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-Simulate_VirtualTrials.R
\name{saveToTrialFile}
\alias{saveToTrialFile}
\title{Helper function: Save a data.frame or a named list to a trial file}
\usage{
saveToTrialFile(
  outputFolder,
  data,
  listMemberName,
  FLAGreuseStoredTrialSummaryFiles
)
}
\description{
Helper function: Save a data.frame or a named list to a trial file
}
