% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{scatterConcACPR}
\alias{scatterConcACPR}
\title{Scatterplot concentrationand ACPR}
\usage{
scatterConcACPR(data, conc, concDay = "Day 7", type = "Crude", acprDay = 28)
}
\arguments{
\item{data}{general data set}

\item{conc}{NAME of the concentration to plot}

\item{concDay}{VISNAME of the timepoint to take the concentrations from}

\item{type}{'Crude' or 'PCR-Adjusted'}

\item{acprDay}{ACPR assessment day, can only be 28, 42, or 63}
}
\description{
Scatterplot concentrationand ACPR
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Anne Kuemmel (IntiQuan)
}
\concept{DataExploration}
