% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ActivityMonitoring.R
\name{set_MMVsettings}
\alias{set_MMVsettings}
\title{set_MMVsettings}
\usage{
set_MMVsettings(
  IQRversion = NULL,
  ModifyRprofile = TRUE,
  UnloadMORpackage = TRUE
)
}
\arguments{
\item{IQRversion}{Default: NULL}

\item{ModifyRprofile}{Default: TRUE}

\item{UnloadMORpackage}{Default: TRUE}
}
\description{
set_MMVsettings
}
\seealso{
Other Activity Monitoring: 
\code{\link{check_ActivityInfo}()},
\code{\link{check_ActivityInfo_PiNK}()},
\code{\link{check_ActivityInfo_Sdrive}()},
\code{\link{check_ActivityResults}()},
\code{\link{create_RprofileMMV}()},
\code{\link{generate_ActivityContent}()},
\code{\link{generate_ActivityResults}()},
\code{\link{generate_DuplicatedActivityReport}()},
\code{\link{generate_MissingActivityReport}()},
\code{\link{generate_OverviewFile}()},
\code{\link{get_ActivityInfo}()},
\code{\link{get_ActivityList}()},
\code{\link{list_ActivityType}()},
\code{\link{list_GenericType}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Activity Monitoring}
