% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-IQRaux.R
\name{sim_IQRmodelConsistent}
\alias{sim_IQRmodelConsistent}
\title{sim_IQRmodel with strict output format}
\usage{
sim_IQRmodelConsistent(simIQRmodelSpec)
}
\arguments{
\item{simIQRmodelSpec}{Argument list for sim_IQRmodel, created by \code{\link[=createSimIQRmodelSpec]{createSimIQRmodelSpec()}}}
}
\value{
\code{IQRsimres} with consistent format, see description
}
\description{
This function needs \code{\link[=createSimIQRmodelSpec]{createSimIQRmodelSpec()}} and \code{\link[=getDefaultEventTable]{getDefaultEventTable()}} and \code{\link[=verifyArg]{verifyArg()}}
}
\details{
\code{\link[IQRtools:sim_IQRmodel]{IQRtools::sim_IQRmodel()}} output format depends on arguments of the sim_IQRmodel call.
\itemize{
\item TIME is very context dependent, eventfor explicit \code{simtime} values passed as argument.
\item ID is not always present: If the eventTable contains one ID only, or sim_IQRmodel is called without eventTable, the output an ID column does not contain
}

By contrast, \strong{this} function has consistent output, suitable for programming.
\itemize{
\item TIME corresponds exactly to the input given in \code{simtime}
\item It always contains an ID column
\item It always has a simNames attribute
}
}
\examples{
library(IQRtools)
library(ggplot2)
fileURL <- "http://192.168.1.12:7804/IQauxfn/inst/02-ExampleModels/ModelLibraryPK/model_2cpt_linsat_iv.txt"
fl <- file.path(tempdir(), basename(fileURL))
download.file(fileURL, fl)

model <- IQRmodel(fl)
simIQRmodelSpec1 <- createSimIQRmodelSpec(model, simtime = c(0), parameters =  c(CL = 10), dosingTable = IQRdosing(0,1,1))
simres1 <- do.call(sim_IQRmodel, simIQRmodelSpec1)
simres1Consistent <- sim_IQRmodelConsistent(simIQRmodelSpec = simIQRmodelSpec1)
}
\seealso{
Other Simulation: 
\code{\link{createSimIQRmodelSpec}()},
\code{\link{create_GDF_fromDesignSpec}()},
\code{\link{getDefaultEventTable}()},
\code{\link{populateGDFWithSimValues_consistent}()}
}
\author{
Daniel Lill (daniel.lill@intiquan.com)
}
\concept{Simulation}
