% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{simTimeRecrudescenceCombo}
\alias{simTimeRecrudescenceCombo}
\title{simTimeRecrudescenceCombo
Simulate combo model for a grid of dose-combinations}
\usage{
simTimeRecrudescenceCombo(
  model,
  parameters,
  doses1,
  doses2,
  n1,
  n2,
  LLOQ,
  simlength = 24 * 32,
  setting = NULL
)
}
\arguments{
\item{model}{IQRmodel}

\item{parameters}{Named vector of parameters}

\item{doses1, doses2}{Vectors of doses. Each combination of both will be simulated}

\item{n1, n2}{Number doses per drug, administration is assumed to be every 24h.}

\item{LLOQ}{LLOQ of model OUTPUT1}

\item{simlength}{Default: 24 * 32. maximum simtime}

\item{setting}{Default: \code{NULL}. Named list of arguments going to \code{\link[IQRtools]{sim_IQRmodel}}}
}
\value{
data.frame with colums c("ID", "Time2Recrud", "n1", "n2", "Dose1", "Dose2")
}
\description{
This function is helpful for isobole plots.
Simulate combo model for all combinations of doses of drug 1 and 2.
Then compute time to recrudescence for
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Anne K?mmel (IntiQuan), Mohammed Cherkaoui (MMV)
}
\concept{Simulations}
