% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{simulate_ComboMouse2Human}
\alias{simulate_ComboMouse2Human}
\title{simulate_ComboMouse2Human}
\usage{
simulate_ComboMouse2Human(
  TreatmentGroups,
  PKmodelFolders,
  PDmodelFolders,
  InteractionModelFolder,
  ComboModel,
  PbaseSample = function(x) {
     sample_Distribution(Mean = 10^6.72, SD = 0.1, Dist =
    "L", nSample = x)
 },
  GRsample = function(x) {
     sample_Distribution(Mean = 0.048, SD = 0.1, Dist = "L",
    nSample = x)
 },
  doseCovariate = list(drug1 = NULL, drug2 = NULL),
  covariates = NULL,
  evalDay = 28,
  simtime = MMVbase::create_PKPDsimtime(24 * evalDay),
  nTrial = 10,
  nSubj = 10,
  LLOQ.PD,
  cureThreshold,
  replace = TRUE,
  setting = NULL
)
}
\arguments{
\item{TreatmentGroups}{Named list of treatment groups - each element should
be length-2 list i.e. drug1, drug2, containing at least dose and time
(for example, drug1 = list(dose=1, time=c(0,1,2,3)))}

\item{PKmodelFolders}{List of length 2, corresponding to each drug. Each
element must contain either a path to an IQRnlme project containing PK
parameters or a function to generate these parameters}

\item{PDmodelFolders}{List of length 2, corresponding to each drug. Each
element must contain either a path to an IQRnlme project containing PD
parameters or a function to generate these parameters}

\item{InteractionModelFolder}{String containing path to IQRnlme project with
DDI model or a function to generate DDI parameters from}

\item{ComboModel}{Either an IQRmodel object or a path to an IQRmodel.txt}

\item{PbaseSample}{Function from which to sample baseline parasitaemia.
Default: function(x) {
sample_Distribution(Mean = 10^6.72, SD = 0.1, Dist = "L",
nSample = x)
}}

\item{GRsample}{Function from which to sample growth rate
Default: function(x) {
sample_Distribution(Mean = 0.048, SD = 0.1, Dist = "L", nSample = x)
}}

\item{doseCovariate}{list with 2 elements with dose covariate for each drug,
assuming dose is a covariate, i.e. list(drug1="DOSELEVEL1", drug2="DOSELEVEL2")
Default: list(drug1 = NULL, drug2 = NULL)}

\item{covariates}{Named list of covariate vectors or sampling functions}

\item{evalDay}{Numeric, day at which to evaluate cure. Default: 28}

\item{simtime}{Numeric vector of simulation time in hours. Default: MMVbase::create_PKPDsimtime(24 * evalDay)}

\item{nTrial}{Integer, number of virtual trials to simulate. Default: 10}

\item{nSubj}{Integer, number of individuals to simulate per trial. Default: 10}

\item{LLOQ.PD}{Numeric, lower limit of quantification of parasitaemia}

\item{cureThreshold}{Numeric, value of parasitaemia to treat as cure threshold. Default: log(1/5000)}

\item{replace}{Logical, sampling replacement flag. Default: TRUE}

\item{setting}{Optional named list of settings for solver. Default: \code{NULL}}
}
\value{
Named list including simulation data, event table, covariate table,
time course summary, ACPR summary and PRR summary
}
\description{
Simulate combo PK/PD models using provided PK and PD models and an interaction model.
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
