% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-FitLines.R
\name{simulate_indivPK}
\alias{simulate_indivPK}
\title{simulate_indivPK
Simulation of PK for individual}
\usage{
simulate_indivPK(x, model = model)
}
\arguments{
\item{x}{data frame with dosing records (defined by TIME, AMT and RATE), PK parameters and the time of MIC (Xknot) for one individual}

\item{model}{IQRmodel for the PK}
}
\value{
data frame with simulated PK timecourse (TIME and )
}
\description{
Concentrations are simulated with the given dosing, individual PK parameters
at 100 steps between minTIME and maxTIME and the time of MIC defined in the input data frame.
The simulation at the time of MIC is flagged in FLAGmic.
}
\seealso{
Other Model-free PD assessment by linear segments: 
\code{\link{define_tLLOQ}()},
\code{\link{define_tMIClinearSegments}()},
\code{\link{fit_LinearModel}()},
\code{\link{fit_LinearSegments}()},
\code{\link{get_CLmaxTlag}()},
\code{\link{get_DataClearance}()},
\code{\link{get_DataGrowth}()},
\code{\link{get_Growth}()},
\code{\link{get_MICextrapolPK}()},
\code{\link{get_MICsimPK}()},
\code{\link{get_linepar}()},
\code{\link{plot_LinSegments}()},
\code{\link{plot_tMICextrapolation}()},
\code{\link{plot_tMICextrapolationRange}()}
}
\author{
Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Model-free PD assessment by linear segments}
