% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-Simulate_VirtualTrials.R
\name{simulate_OneVirtualTrial}
\alias{simulate_OneVirtualTrial}
\title{Simulate one virtual trial based on a input .rds file}
\usage{
simulate_OneVirtualTrial(trialFileName)
}
\arguments{
\item{trialFileName}{A character string denoting a file path to a trial .rds file. By convention, the general settings for the simulation
are loaded from a file setting.rds found in the parent directory of trialFileName.}
}
\value{
a named list
}
\description{
Simulate one virtual trial based on a input .rds file
}
