% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-Simulate_VirtualTrials.R
\name{simulate_VirtualTrialsInFolder}
\alias{simulate_VirtualTrialsInFolder}
\title{Run simulations for all virtual trial rds files stored in a directory}
\usage{
simulate_VirtualTrialsInFolder(
  simFolder,
  Nparallel = parallel::detectCores(logical = FALSE),
  setting = NULL
)
}
\arguments{
\item{simFolder}{A character string denoting a folder path where the virtual trial .rds files are stored.}

\item{Nparallel}{An integer denoting the number of parallel processes to use for the simulations. Default is
the number of physical cores detected on the machine.}

\item{setting}{A named list of settings as returned by validateSetting_SimulateVirtualTrials(). If NULL (default),}
}
\value{
a named list with elements:
\itemize{
\item summaryPKPD.ByTrial: a data.table summarizing PKPD results by trial
\item summaryClinEnd.ByIndiv: a data.table summarizing clinical end-points by individual
\item summaryClinEnd.ByTrial: a data.table summarizing clinical end-points by trial
\item simPKPD: a data.table with simulated PKPD results for all individuals
}
}
\description{
Run simulations for all virtual trial rds files stored in a directory
}
\examples{
\dontrun{
# Assuming that virtual trial .rds files are stored in the folder "simulations/"
results <- simulate_VirtualTrialsInFolder(simFolder = "simulations/", Nparallel = 4)
}
}
