% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summarize_KeyPDparametersFromSimulations}
\alias{summarize_KeyPDparametersFromSimulations}
\title{Summarize a selection of Key PD parameters from simulated data}
\usage{
summarize_KeyPDparametersFromSimulations(
  simPKPD,
  timeCOL = "TIME",
  concCOL = "Cc",
  paraCOL = "PL",
  Plog = TRUE,
  percentiles = c(5, 50, 95),
  CIlevel = 90,
  usubjidCOL = c("IndivID", "USUBJID"),
  trialCOL = c("TrialID"),
  scenCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses")
)
}
\arguments{
\item{simPKPD}{Data-table output of \code{simulate_virtualtrials()} (or similar format).}

\item{timeCOL}{Name of column containing values for simulation time (Default: \code{"TIME"}).}

\item{concCOL}{Name of column containing values for drug concentration (Default: \code{"Cc"}).}

\item{paraCOL}{Name of column containing values for parasitaemia (Default:  \code{"PL"}).}

\item{Plog}{Is paraistaemia provided in Log scale?  (Default: \code{TRUE}).}

\item{percentiles}{Percentiles of outcome variable(s) to be calculated if appropriate (Default: \code{c(5,50,90)}).}

\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}

\item{usubjidCOL}{Name(s) of column(s) of \code{simPKPD} uniquely describing individuals (Default: \code{c("IndivID","USUBJID")}).}

\item{trialCOL}{Name(s) of column(s) of \code{simPKPD} uniquely describing trial (Default: \code{c("TrialID")}).}

\item{scenCOL}{Name(s) of column(s) of \code{simPKPD} uniquely describing scenario (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses")})}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Data-table containing values of population PK parameters from simulated trials at specified percentiles
}\if{html}{\out{</div>}}
}
\description{
\code{"MIC"},\code{"MPC90"}, \code{"PRR24"}, \code{"PRR48"}, \code{"PRR72"}, \code{"PRRtot"},
\code{"tMIC"} and \code{"tMPC90"} are estimated for each individual and summarized.
The predictive interval (defined in with the variable 'percentiles') is returned with the confidence
interval as given by \code{CIlevel}.
}
\seealso{
Other Summarize Functions: 
\code{\link{summarizeTrial_ChemoBreakthrough}()},
\code{\link{summarizeTrial_ChemoSurvival}()},
\code{\link{summarizeTrial_ChemoSurvival_ComboPD}()},
\code{\link{summarizeTrial_ChemoSurvival_MonoPD}()},
\code{\link{summarize_KeyPKparametersFromSimulations}()},
\code{\link{summarize_PKPDprofilesFromSimulations}()},
\code{\link{summarize_PRRtotFromSimulations}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Summarize Functions}
