% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summarize_KeyPKparametersFromSimulations}
\alias{summarize_KeyPKparametersFromSimulations}
\title{Summarize a selection of Key PK parameters from simulated data}
\usage{
summarize_KeyPKparametersFromSimulations(
  simPKPD,
  timeCOL = "TIME",
  concCOL = "Cc",
  percentiles = c(5, 50, 95),
  CIlevel = 90,
  usubjidCOL = c("IndivID", "USUBJID"),
  trialCOL = c("TrialID"),
  scenCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses")
)
}
\arguments{
\item{simPKPD}{Data-table output of `simulate_virtualtrials()` (or similar format).}

\item{timeCOL}{Name of column containing values for simulation time (Default: \code{"TIME"}).}

\item{concCOL}{Name of column containing values for drug concentration (Default: \code{"Cc"}).}

\item{percentiles}{Percentiles of outcome variable(s) to be calculated if appropriate (Default: \code{c(5,50,90)}).}

\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}

\item{usubjidCOL}{Name(s) of column(s) of `simPKPD` uniquely describing individuals (Default: \code{c("IndivID","USUBJID")}).}

\item{trialCOL}{Name(s) of column(s) of `simPKPD` uniquely describing trial (Default: \code{c("TrialID")}).}

\item{scenCOL}{Name(s) of column(s) of `simPKPD` uniquely describing scenario (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses")})}
}
\value{
Data-table containing values of population PK parameters from simulated trials at specified percentiles
}
\description{
\code{"AUCinf"},\code{"Cmax"} and \code{"Tmax"} are estimated for each individual and summarized.
The predictive interval (defined in with the variable 'percentiles') is returned with the confidence
interval as given by `CIlevel`.
}
\seealso{
Other Summarize Functions: 
\code{\link{summarizeTrial_ChemoBreakthrough}()},
\code{\link{summarizeTrial_ChemoSurvival}()},
\code{\link{summarizeTrial_ChemoSurvival_ComboPD}()},
\code{\link{summarizeTrial_ChemoSurvival_MonoPD}()},
\code{\link{summarize_KeyPDparametersFromSimulations}()},
\code{\link{summarize_PIandCIgeneric}()},
\code{\link{summarize_PKPDprofilesFromSimulations}()},
\code{\link{summarize_PRRtotFromSimulations}()},
\code{\link{summaryAcrossTrials}()},
\code{\link{summaryByTrial}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Summarize Functions}
