% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summarize_PIandCIgeneric}
\alias{summarize_PIandCIgeneric}
\title{Generic function to estimate PI and CI}
\usage{
summarize_PIandCIgeneric(
  dataInput,
  varCOL,
  percentiles = c(5, 50, 90),
  CIlevel = 90,
  usubjidCOL = c("IndivID", "USUBJID"),
  trialCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses", "TrialID", "TIME"),
  scenCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses", "TIME")
)
}
\arguments{
\item{dataInput}{data.table containing outcome variable(s) and selected other variables for each individual.}

\item{varCOL}{List containing metrics (percentiles, mean and/or geomean) to be calculated for each outcome variable - i.e. \code{list(Cmax=c("Percentile","Mean"),APR="Mean")}.}

\item{percentiles}{Percentiles of outcome variable(s) to be calculated if appropriate (Default: \code{c(5,50,90)}).}

\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}

\item{trialCOL}{Name(s) of column(s) of `dataInput` uniquely describing trial (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses","TrialID","TIME")}).}

\item{scenCOL}{Name(s) of column(s) of `dataInput` uniquely describing scenario (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses","TIME")})}

\item{usubJIDCOL}{Name(s) of column(s) of `dataInput` uniquely describing individuals (Default: \code{c("IndivID","USUBJID")}).}
}
\value{

}
\description{
Generic function to estimate the confidence interval of a metric (mean, percentiles, and/or geoman)
of various given variables, when multiple trials have been simulated. It calls `summaryByTrial`
and `summmaryAcrossTrials` in a row. If no metric is provided, percentile is used by default
for numeric variable, while mean is used for logical variable.
}
\seealso{
Other Summarize Functions: 
\code{\link{summarizeTrial_ChemoBreakthrough}()},
\code{\link{summarizeTrial_ChemoSurvival}()},
\code{\link{summarize_KeyPDparametersFromSimulations}()},
\code{\link{summarize_KeyPKparametersFromSimulations}()},
\code{\link{summarize_PKPDprofilesFromSimulations}()},
\code{\link{summarize_PRRtotFromSimulations}()},
\code{\link{summaryAcrossTrials}()},
\code{\link{summaryByTrial}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Summarize Functions}
