% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summarize_PRRtotFromSimulations}
\alias{summarize_PRRtotFromSimulations}
\title{Summarizes the total parasite reduction ratio}
\usage{
summarize_PRRtotFromSimulations(
  simPKPD,
  timeCOL = "TIME",
  paraCOL = "PL",
  Plog = TRUE,
  percentiles = c(5, 50, 95),
  CIlevel = 90,
  usubjidCOL = c("IndivID", "USUBJID"),
  trialCOL = c("TrialID"),
  scenCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses")
)
}
\arguments{
\item{simPKPD}{Data-table output of `simulate_virtualtrials()` (or similar format).}

\item{timeCOL}{Name of column containing values for simulation time (Default: \code{"TIME"}).}

\item{paraCOL}{Name of column containing values for parasitaemia (Default:  \code{"PL"}).}

\item{Plog}{Is paraistaemia provided in Log scale?  (Default: `TRUE`).}

\item{percentiles}{Percentiles of outcome variable(s) to be calculated if appropriate (Default: \code{c(5,50,90)}).}

\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}

\item{usubjidCOL}{Name(s) of column(s) of `simPKPD` uniquely describing individuals (Default: \code{c("IndivID","USUBJID")}).}

\item{trialCOL}{Name(s) of column(s) of `simPKPD` uniquely describing trial (Default: \code{c("TrialID")}).}

\item{scenCOL}{Name(s) of column(s) of `simPKPD` uniquely describing scenario (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses")})}
}
\value{

}
\description{
Summarizes the total parasite reduction ratio
}
\seealso{
Other Summarize Functions: 
\code{\link{summarizeTrial_ChemoBreakthrough}()},
\code{\link{summarizeTrial_ChemoSurvival}()},
\code{\link{summarize_KeyPDparametersFromSimulations}()},
\code{\link{summarize_KeyPKparametersFromSimulations}()},
\code{\link{summarize_PIandCIgeneric}()},
\code{\link{summarize_PKPDprofilesFromSimulations}()},
\code{\link{summaryAcrossTrials}()},
\code{\link{summaryByTrial}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Summarize Functions}
