% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summarizeTrial_APRLongitudinal}
\alias{summarizeTrial_APRLongitudinal}
\title{Summarize PKPD simulation results for a virtual trial of one experimental setting and one dose}
\usage{
summarizeTrial_APRLongitudinal(
  simPKPD,
  trialFileName,
  varPKPD = c("Ccx1", "Ccx2", "Effx1", "Effx2", "Kkillx1", "Kkillx2", "Kkill", "PL"),
  percentiles = c(5, 50, 95),
  samplePDtime = seq(0, 28 * 24, 24),
  LLOQ.PD = log(10000),
  aprTime = 28,
  extrapTol = 48,
  FLAGextrapolateTime = TRUE
)
}
\arguments{
\item{simPKPD}{A data.frame representing the merge (join) for data.frame returned by \code{\link[IQRtools:sim_IQRmodel]{IQRtools::sim_IQRmodel()}}, and
individual and experimental covariates. This argument is set automatically by the calling function.}

\item{trialFileName}{A character string denoting the trial .rds file on the disk. This file can be loaded to get information about the
virtual subjects, such as individual and experimental covariates.}

\item{varPKPD}{a character vector indicating time varying PKPD variables of the simulated PKPD model. All of these
should be names of columns in \code{simPKPD}.}

\item{percentiles}{a numeric vector indicating percentiles (default: \code{c(5, 50, 95)}) for summarizing
individual PKPD variables and clinical end-points.}

\item{samplePDtime}{a numeric vector (default \code{seq(0,28*24,24)}) of time-points at which PL is
needed for summarizing individual clinical endpoints. This should contain aprTime.}

\item{LLOQ.PD}{a numeric indicating log parasitemia LLOQ (default: \code{log(10000)}).}

\item{aprTime}{a numeric indicating time after beginning of treatment in days at which APR should be evaluated. Default: 28.
This argument is passed to function \code{\link[=evaluate_APR]{evaluate_APR()}}.}

\item{extrapTol, FLAGextrapolateTime}{These arguments are passed to function \code{\link[=evaluate_APR]{evaluate_APR()}}}
}
\value{
a named list with the following members:
\itemize{
\item ScenID: an integer
\item ExpID: an integer
\item DoseID: an integer
\item TrialID: an integer
\item summaryPKPD.ByTrial: a data.frame
\item summaryClinEnd.ByTrial: a data.frame of 1 row
\item summaryClinEnd.ByIndiv: a data.frame of as many rows as there are virtual subjects per trial
}
}
\description{
This function is called at runtime by \code{\link[=simulate_OneVirtualTrial]{simulate_OneVirtualTrial()}}.
}
\details{
This function is designed to be passed as argument of \code{\link[=simulate_VirtualTrials]{simulate_VirtualTrials()}}. Currently, the function
assumes that the \code{simPKPD} has a column PL denoting natural log parasitemia.
}
