% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summarizeTrial_APRLongitudinal_PK_secondary}
\alias{summarizeTrial_APRLongitudinal_PK_secondary}
\title{Summarize PKPD simulation results for a virtual trial of one experimental setting and one dose}
\usage{
summarizeTrial_APRLongitudinal_PK_secondary(
  simPKPD,
  trialFileName,
  varPKPD = c("Cc", "Eff", "Kkill", "PL"),
  percentiles = c(5, 50, 95),
  samplePDtime = seq(0, 28 * 24, 24),
  LLOQ.PD = log(10000),
  aprTime = 28,
  extrapTol = 48,
  FLAGextrapolateTime = TRUE,
  timeCOL = "TIME",
  concCOL = "Cc",
  CIlevel = 90,
  usubjidCOL = c("IndivID", "USUBJID"),
  trialCOL = c("TrialID"),
  scenCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses")
)
}
\arguments{
\item{simPKPD}{A data.frame representing the merge (join) for data.frame returned by \code{\link{IQRtools::sim_IQRmodel}}, and
individual and experimental covariates. This argument is set automatically by the calling function.}

\item{trialFileName}{A character string denoting the trial .rds file on the disk. This file can be loaded to get information about the
virtual subjects, such as individual and experimental covariates.}

\item{varPKPD}{a character vector indicating time varying PKPD variables of the simulated PKPD model. All of these
should be names of columns in \code{simPKPD}.}

\item{percentiles}{a numeric vector indicating percentiles (default: \code{c(5, 50, 95)}) for summarizing
individual PKPD variables and clinical end-points.}

\item{samplePDtime}{a numeric vector (default \code{seq(0,28*24,24)}) of time-points at which PL is
needed for summarizing individual clinical endpoints. This should contain aprTime.}

\item{LLOQ.PD}{a numeric indicating log parasitemia LLOQ (default: \code{log(10000)}).}

\item{aprTime}{a numeric indicating time after beginning of treatment in days at which APR should be evaluated. Default: 28.
This argument is passed to function \code{\link{evaluate_APR}}.}

\item{extrapTol, FLAGextrapolateTime}{These arguments are passed to function \code{\link{evaluate_APR}}}

\item{timeCOL}{Name of column containing values for simulation time (Default: \code{"TIME"}).}

\item{concCOL}{Name of column containing values for drug concentration (Default: \code{"Cc"}).}

\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}

\item{usubjidCOL}{Name(s) of column(s) of \code{simPKPD} uniquely describing individuals (Default: \code{c("IndivID","USUBJID")}).}

\item{trialCOL}{Name(s) of column(s) of \code{simPKPD} uniquely describing trial (Default: \code{c("TrialID")}).}

\item{scenCOL}{Name(s) of column(s) of \code{simPKPD} uniquely describing scenario (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses")})}
}
\value{
a named list with the following members:
\itemize{
\item ScenID: an integer
\item ExpID: an integer
\item DoseID: an integer
\item TrialID: an integer
\item summaryPKPD.ByTrial: a data.frame
\item summaryClinEnd.ByTrial: a data.frame of 1 row
\item summaryClinEnd.ByIndiv: a data.frame of as many rows as there are virtual subjects per trial
\item summaryPK.ByTrial: a data.frame of 1 row
\item summaryPK.ByIndiv: a data.frame of as many rows as there are virtual subjects per trial
}
}
\description{
This function is called at runtime by \code{\link{simulate_OneVirtualTrial}}.
}
\details{
This function is designed to be passed as argument of \code{\link{simulate_VirtualTrials}}. Currently, the function
assumes that the \code{simPKPD} has a column PL denoting natural log parasitemia.
In addition to the primary parasitological clinical endpoints, this function also summarizes secondary endpoints for PK and PD.
}
