% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summarizeTrial_Basic}
\alias{summarizeTrial_Basic}
\title{Summarize PKPD simulation results for a virtual trial of one experimental setting and one dose}
\usage{
summarizeTrial_Basic(
  simPKPD,
  trialFileName,
  varPKPD = c("Cc", "PL"),
  percentiles = c(5, 50, 95)
)
}
\arguments{
\item{simPKPD}{A data.frame representing the merge (join) for data.frame returned by \code{\link[IQRtools:sim_IQRmodel]{IQRtools::sim_IQRmodel()}}, and
individual and experimental covariates. This argument is set automatically by the calling function.}

\item{trialFileName}{A character string denoting the trial .rds file on the disk. This file can be loaded to get information about the
virtual subjects, such as individual and experimental covariates.}

\item{varPKPD}{a character vector indicating time varying PKPD variables of the simulated PKPD model. All of these
should be names of columns in \code{simPKPD}.}

\item{percentiles}{a numeric vector indicating percentiles (default: \code{c(5, 50, 95)}) for summarizing
individual PKPD variables and clinical end-points.}
}
\value{
a named list with the following members:
\itemize{
\item ScenID: an integer
\item ExpID: an integer
\item DoseID: an integer
\item TrialID: an integer
\item summaryPKPD.ByTrial: a data.frame
\item summaryClinEnd.ByTrial: NULL
\item summaryClinEnd.ByIndiv: NULL
}
}
\description{
This function is called at runtime by \code{\link[=simulate_OneVirtualTrial]{simulate_OneVirtualTrial()}}.
}
\details{
This function is designed to be passed as argument of \code{\link[=simulate_VirtualTrials]{simulate_VirtualTrials()}}. Currently, the function
assumes that the \code{simPKPD} has a column PL denoting natural log parasitemia.
}
