% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summarizeTrial_ChemoBreakthrough}
\alias{summarizeTrial_ChemoBreakthrough}
\title{Summarize PKPD simulation results for a virtual trial of one experimental setting and one dose}
\usage{
summarizeTrial_ChemoBreakthrough(
  simPKPD,
  trialFileName,
  varPKPD = c("Cc", "PLiver", "PBlood"),
  percentiles = c(5, 50, 95),
  LLOQ.PD = log(10000),
  timeCOL = "TIME",
  PbloodCOL = "PBlood",
  Plog = TRUE,
  FLAGinterpolateTime = FALSE
)
}
\arguments{
\item{simPKPD}{A data.frame representing the merge (join) for data.frame returned by \code{\link{IQRtools::sim_IQRmodel}}, and
individual and experimental covariates. This argument is set automatically by the calling function.}

\item{trialFileName}{A character string denoting the trial .rds file on the disk. This file can be loaded to get information about the
virtual subjects, such as individual and experimental covariates.}

\item{varPKPD}{a character vector indicating time varying PKPD variables of the simulated PKPD model. All of these
should be names of columns in \code{simPKPD}.}

\item{percentiles}{a numeric vector indicating percentiles (default: \code{c(5, 50, 95)}) for summarizing
individual PKPD variables and clinical end-points.}

\item{LLOQ.PD}{a numeric indicating log parasitemia LLOQ (default: \code{log(10000)}).}

\item{timeCOL}{Name of column containing values for simulation time (Default: \code{"TIME"}).}

\item{PbloodCOL}{Name of column containing values for blood stage parasitemia (Default: \code{"PBlood"}).}

\item{Plog}{Indicate if parasitemia is in the log or linear scale (Default: \code{TRUE} which means it is logged).}

\item{FLAGinterpolateTime}{A logical indicating if the PL measurements should be interpolated within the simulation period.}
}
\value{
a named list with the following members:
\itemize{
\item summaryPKPD.ByTrial: a data.frame
\item summaryClinEnd.ByIndiv: a data.frame of as many rows as there are virtual subjects per trial
}
}
\description{
This function is called at runtime by \code{\link{simulate_OneVirtualTrial}}.
}
\details{
This function is designed to be passed as argument of \code{\link{simulate_VirtualTrials}}.It performs
a summary of individual clinical endpoints using \code{\link{MMVmalaria:::evaluate_BreakthroughEvent}}. It does
not perform summaries by trial, or across trials.
}
\seealso{
Other Summarize Functions: 
\code{\link{summarizeTrial_ChemoSurvival}()},
\code{\link{summarize_KeyPDparametersFromSimulations}()},
\code{\link{summarize_KeyPKparametersFromSimulations}()},
\code{\link{summarize_PIandCIgeneric}()},
\code{\link{summarize_PKPDprofilesFromSimulations}()},
\code{\link{summarize_PRRtotFromSimulations}()},
\code{\link{summaryAcrossTrials}()},
\code{\link{summaryByTrial}()}
}
\author{
Catalina Barcelo (MMV), Sam Jones (MMV)
}
\concept{Summarize Functions}
