% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-ModelComparison.R
\name{summary_ParametersMulti}
\alias{summary_ParametersMulti}
\title{summary_ParametersMulti
Summarize parameters over multiple SysFit projects}
\usage{
summary_ParametersMulti(projfolders, sortBy = NULL)
}
\arguments{
\item{sortBy}{Either NULL (no sorting), character denoting a model metric (BIC, OBJ, AIC)
or a parameter by which to sort the output columns.}

\item{projfolder}{Character vector with the project folders}
}
\value{
IQRtable object
}
\description{
Reads the parameter results from different project folders and
creates an output table with separate columns for each model/folder.
}
\seealso{
\code{\link[=summary_ParametersDDI]{summary_ParametersDDI()}}
}
\author{
Daniel Kaschek (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
