% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{summary_PKdataMMV}
\alias{summary_PKdataMMV}
\title{summary_PKdataMMV}
\usage{
summary_PKdataMMV(dataGen, filePath = NULL, NewTRTNAME = NULL)
}
\arguments{
\item{filePath}{Default: NULL}

\item{NewTRTNAME}{Default: NULL}
}
\description{
summary_PKdataMMV
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
