% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{summary_PKPDdataCombo}
\alias{summary_PKPDdataCombo}
\title{summary_PKPDdataCombo}
\usage{
summary_PKPDdataCombo(
  dataGen,
  filePath = NULL,
  CompoundList = NULL,
  positiveQC = NULL,
  TRTNAME = "TRTNAME"
)
}
\arguments{
\item{dataGen}{IQRdataGeneral object}

\item{filePath}{Character string with path to folder for saving plot}

\item{CompoundList}{Default: NULL}

\item{positiveQC}{Default: NULL}
}
\value{
Data frame with summary of PK and PD data
}
\description{
Summarize PK and PD data by study and treatment
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
