% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{summary_PKPDdataMMV}
\alias{summary_PKPDdataMMV}
\title{summary_PKPDdataMMV
Create summary tables for pharmacokinetic data}
\usage{
summary_PKPDdataMMV(
  dataGen,
  filePath = NULL,
  NewTRTNAME = NULL,
  positiveQC = NULL
)
}
\arguments{
\item{dataGen}{IQRdataGeneral object}

\item{filePath}{Character string with path to folder for saving plot}

\item{NewTRTNAME}{Character vector denoting new treatment name. Default: NULL}

\item{positiveQC}{Character vector identifying the name of positive control (must match upper or lower case as in original dataset). Default: NULL}
}
\value{
A data frame with summary of pharmacokinetic data.
}
\description{
Summarizes pharmacokinetic data in 3 different tables: (1) number of subjects per study and treatment group, (2) nominal sampling times per observation type and (3) dosing/inoculation nominal time per treatment group and dosing/inoculation type.
}
\author{
Aline Fuchs (MMV), catalinbarcelo (MMV), Mohammed H. Cherkaoui (MMV)
}
