% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{summary_PKPDdataSimplified}
\alias{summary_PKPDdataSimplified}
\title{summary_PKPDdataSimplified}
\usage{
summary_PKPDdataSimplified(
  dataGen,
  filename = NULL,
  PKname = NULL,
  PDname = NULL,
  StudyTYPE,
  Center,
  NewTRTNAME = NULL,
  CovToTRTNAME = NULL,
  CovNewTRTNAME = NULL
)
}
\arguments{
\item{filename}{Default: NULL}

\item{PKname}{Default: NULL}

\item{PDname}{Default: NULL}

\item{NewTRTNAME}{Default: NULL}

\item{CovToTRTNAME}{Default: NULL}

\item{CovNewTRTNAME}{Default: NULL}
}
\value{

}
\description{
summary_PKPDdataSimplified
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
