% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-SummarizeFunctions.R
\name{summaryAcrossTrials}
\alias{summaryAcrossTrials}
\title{Summarize confidence intervals for provided metrics across trials.}
\usage{
summaryAcrossTrials(
  dataInput,
  CIlevel = 90,
  metricCOL = "Metric",
  variableCOL = "Variable",
  valueCOL = "Value",
  scenCOL = c("ScenID", "ExpID", "DoseID", "Dose", "nbrDoses", "TIME")
)
}
\arguments{
\item{dataInput}{data.table containing summarized outcome variable(s) summarized by trials - produced by `summaryByTrial` or similar format.}

\item{CIlevel}{Numeric containing the confidence interval in percent to be estimated (Default: 90)}

\item{metricCOL}{name of column of `dataInput` containing names of metrics on which CI is being estimated (Default: \code{"Metric"})}

\item{variableCOL}{name of column of `dataInput` containing names of variables (Default: \code{"Variable"})}

\item{valueCOL}{name of column of `dataInput` containing values of metrics for specific variables (Default: \code{"Value"})}

\item{scenCOL}{Name(s) of column(s) of `dataInput` uniquely describing scenario (Default: \code{c("ScenID","ExpID","DoseID","Dose","nbrDoses","TIME")})}
}
\value{
data table containing confidence intervals for selected outcome variable(s) and summarized metric(s)
}
\description{
Summarize confidence intervals for provided metrics across trials.
}
\seealso{
Other Summarize Functions: 
\code{\link{summarizeTrial_ChemoBreakthrough}()},
\code{\link{summarizeTrial_ChemoSurvival}()},
\code{\link{summarize_KeyPDparametersFromSimulations}()},
\code{\link{summarize_KeyPKparametersFromSimulations}()},
\code{\link{summarize_PIandCIgeneric}()},
\code{\link{summarize_PKPDprofilesFromSimulations}()},
\code{\link{summarize_PRRtotFromSimulations}()},
\code{\link{summaryByTrial}()}
}
\author{
Sam Jones (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Summarize Functions}
