% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{summarybyGroup_MMVnca}
\alias{summarybyGroup_MMVnca}
\title{symmarybyGroup_MMVnca}
\usage{
summarybyGroup_MMVnca(
  resultNCA,
  stratify = NULL,
  stats = "geomean",
  colSummary = NULL,
  SIGNIF = 4,
  report = TRUE
)
}
\arguments{
\item{stratify}{Default: NULL}

\item{stats}{Default: 'geomean'}

\item{colSummary}{Default: NULL}

\item{SIGNIF}{Default: 4}

\item{report}{Default: TRUE}
}
\description{
symmarybyGroup_MMVnca
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summary_PKdata_byStudyTRT}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
