% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{swapName_MMVnameToName}
\alias{swapName_MMVnameToName}
\title{swapName_MMVnameToName
Swaps from MMV name for compound for nick name of interest}
\usage{
swapName_MMVnameToName(data, CompoundList)
}
\arguments{
\item{data}{IQRdataGENERAL object, data frame or character vector with MMV name for compound}

\item{CompoundList}{list with both MMV and nick name for the compound}
}
\value{
\code{data} object with nick name for compound
}
\description{
Swaps the MMV name for the nick name of interest within \code{data} based on \code{CompoundList}.
If \code{data} is a IQRdataGENERAL object, MMV name for compound will also be swapped in IQRdataGENERAL object attributes.
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{list_Center}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()},
\code{\link{summarizeReplicates}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
To be defined
}
\concept{Data Preparation}
