% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{swapName_NameToMMVname}
\alias{swapName_NameToMMVname}
\title{swapName_NameToMMVname
Swaps from nick name of interest for compound for MMV name}
\usage{
swapName_NameToMMVname(data, CompoundList)
}
\arguments{
\item{data}{IQRdataGENERAL object, data frame or character vector with nick name of interest for compound}

\item{CompoundList}{list with both MMV and nick name for the compound}
}
\value{
\code{data} object with MMV name for compound
}
\description{
Swaps the nick name of interest for the MMV name within \code{data} based on \code{CompoundList}.
If \code{data} is a IQRdataGENERAL object, nick name of interest for compound will also be swapped in IQRdataGENERAL object attributes.
}
\seealso{
Other Data Preparation: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{censorGametocyte}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{convert_catCovToTxt}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{list_Center}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()},
\code{\link{summarizeReplicates}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
To be defined
}
\concept{Data Preparation}
