% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{table_covariateSummaryResults}
\alias{table_covariateSummaryResults}
\title{table_covariateSummaryResults}
\usage{
table_covariateSummaryResults(
  tableInputPath,
  tableOutputPath = NULL,
  report = TRUE,
  FUNnameModel = function(x) {
     gsub("MODEL_", "", x)
 }
)
}
\arguments{
\item{tableInputPath}{String denoting path to IQRtable .txt file saved with \code{IQRtools::IQRoutputTable()}}

\item{tableOutputPath}{Optional string denoting location to write table, will attempt to default to provided tableInputPath if NULL}

\item{report}{Logical, TRUE: add reporting tags FALSE: do not add (for screen output only)
NULL: same as TRUE if xtable provided. Default: TRUE}

\item{Function}{Default: function(x) {
gsub("MODEL_", "", x)
}}
}
\description{
Read a covariate summary table in .txt format as generated by
IQRtools, restructure it into a tidier format and save it
}
\seealso{
Other Reporting: 
\code{\link{generate_EstimateTableForReport}()},
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_PKmodelHypos}()},
\code{\link{reportText_convertListToText}()},
\code{\link{table_PKsummaryCompare}()},
\code{\link{table_PKsummaryResults}()},
\code{\link{table_strToLaTeX}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Reporting}
