% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{table_EstimatesComboNLME}
\alias{table_EstimatesComboNLME}
\title{table_EstimatesComboNLME}
\usage{
table_EstimatesComboNLME(
  x,
  filename = NULL,
  FLAGout = FALSE,
  title = NULL,
  footer = NULL
)
}
\arguments{
\item{x}{An IQRnlmeProjectMulti object containing models to compare}

\item{filename}{Character string containing path to filename for writing results.
Default: \code{NULL}}

\item{FLAGout}{Logical: Return generated table? Default: \code{FALSE}}

\item{title}{Character string defining name of table. Default: \code{NULL}}

\item{footer}{Character string defining footer for table. Default: \code{NULL}}
}
\value{
Results table if \code{FLAGout} = TRUE, otherwise table is saved to path provided
in \code{filename}
}
\description{
Convert interaction parameters from an IQRnlmeProjectMulti containing
PD interaction models into interpretable fold-change values.
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()}
}
\author{
Aline Fuchs (MMV)
}
\concept{Model Assessment}
