% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-ModelComparison.R
\name{table_ModelEstimates}
\alias{table_ModelEstimates}
\title{table_ModelEstimates}
\usage{
table_ModelEstimates(
  modelFolderList,
  filename = NULL,
  xtitle = NULL,
  xfooter = NULL,
  FLAGreturnOBJECT = FALSE,
  FLAGreport = TRUE
)
}
\arguments{
\item{filename}{Default: NULL}

\item{xtitle}{Default: NULL}

\item{xfooter}{Default: NULL}

\item{FLAGreturnOBJECT}{Default: FALSE}

\item{FLAGreport}{Default: TRUE}
}
\description{
table_ModelEstimates
}
\seealso{
Other ModelComparison: 
\code{\link{getEstimateTable}()}
}
\author{
Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{ModelComparison}
