% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{table_PKsummaryCompare}
\alias{table_PKsummaryCompare}
\title{table_PKsummaryCompare}
\usage{
table_PKsummaryCompare(
  tableInputPath1,
  tableInputPath2,
  tableOutputPath = NULL,
  renameMODEL = NULL,
  addDescription = NULL,
  Metric = "BIC",
  xTitle = NULL,
  xFooter = NULL,
  report = TRUE
)
}
\arguments{
\item{tableInputPath1}{String denoting path to IQRtable .txt file saved with \code{IQRtools::IQRoutputTable()}}

\item{tableInputPath2}{String denoting path to IQRtable .txt file saved with \code{IQRtools::IQRoutputTable()}}

\item{tableOutputPath}{Optional string denoting location to write table, will
attempt to default to provided tableInputPath1 and tableInputPath2 if NULL. (Default: \code{NULL}).}

\item{renameMODEL}{Optional function to rename combined model output (Default: \code{NULL}).}

\item{addDescription}{Optional function to add description to combined model output (Default: \code{NULL}).}

\item{Metric}{Which metric (OBJ, BIC, AIC) to compare and arrange models with. Default: 'BIC'}

\item{xTitle}{Optional string for adding title to output table (Default: \code{NULL}).}

\item{xFooter}{Optional string for adding footer to output table (Default: \code{NULL}).}

\item{report}{Logical, TRUE: add reporting tags FALSE: do not add
(for screen output only) NULL: same as TRUE if xtable provided.Default: TRUE}
}
\description{
Read two PK models in .txt format as generated by
IQRtools, combine and compare them.
}
\seealso{
Other Reporting: 
\code{\link{generate_EstimateTableForReport}()},
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_PKmodelHypos}()},
\code{\link{reportText_convertListToText}()},
\code{\link{table_PKsummaryResults}()},
\code{\link{table_covariateSummaryResults}()},
\code{\link{table_strToLaTeX}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Reporting}
