% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{table_PKsummaryResults}
\alias{table_PKsummaryResults}
\title{table_PKsummaryResults}
\usage{
table_PKsummaryResults(
  tableInputPath,
  tableOutputPath = NULL,
  colToAdd = c("Nr. of Compartments", "Clearance", "Absorption", "Time Lag",
    "Error Model"),
  Metric = "BIC",
  xTitle = NULL,
  xFooter = NULL,
  report = TRUE
)
}
\arguments{
\item{tableOutputPath}{(Default: \code{NULL}).}

\item{colToAdd}{Default: c("Nr. of Compartments", "Clearance", "Absorption", "Time Lag",
"Error Model")}

\item{Metric}{Default: 'BIC'}

\item{xTitle}{(Default: \code{NULL}).}

\item{xFooter}{(Default: \code{NULL}).}

\item{report}{Default: TRUE}
}
\value{

}
\description{
table_PKsummaryResults
}
\seealso{
Other Reporting: 
\code{\link{generate_EstimateTableForReport}()},
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_PKmodelHypos}()},
\code{\link{reportText_convertListToText}()},
\code{\link{table_PKsummaryCompare}()},
\code{\link{table_covariateSummaryResults}()},
\code{\link{table_strToLaTeX}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Reporting}
