% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{table_strToLaTeX}
\alias{table_strToLaTeX}
\title{Convert equations/parameters to LaTeX mode}
\usage{
table_strToLaTeX(
  tableInputPath,
  tableOutputPath = NULL,
  colToChange = "PARAMETER",
  ListStrToLatex = parameter_strToLaTeX(),
  TitleRow = FALSE,
  keepCol = NULL,
  report = TRUE,
  FLAGcompliance = TRUE
)
}
\arguments{
\item{tableInputPath}{Path of the IQR table to adjust}

\item{tableOutputPath}{Output path to save the updated table (Default: \code{NULL} which will save it to tableInputPath"_LaTeX").}

\item{colToChange}{Column in which to apply the changes (Default: `PARAMETER`)}

\item{ListStrToLatex}{list the character to change and their LaTeX expression; e.g. \code{list(Fabs0 = "F_{abs,0}", ka = "k_{a}", CL = "CL")} (Default: \code{MMVmalaria::parameter_strToLaTeX()})}

\item{TitleRow}{To change adjust the title of each column (Default: \code{FALSE}).}

\item{keepCol}{List of column of th etable to keep (Default: \code{NULL} which will keep all columns).}

\item{report}{Logical to choose report mode or not (Default: \code{TRUE}).}

\item{FLAGcompliance}{Logical to choose compliance mode or not (Default: \code{TRUE}).}
}
\value{
An IQR table with the specified changes
}
\description{
Convert equations/parameters to LaTeX mode
}
\details{
Parameters and/or equations within IQRtable can be automatically converted to a LaTeX equation
for nicer reporting. `colToChange` specify to which column the changes should be made, while
`ListStrToLatex` is a list providing the characters to detect and their respective LaTeX notation.
}
\seealso{
Other Reporting: 
\code{\link{generate_EstimateTableForReport}()},
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_PKmodelHypos}()},
\code{\link{reportText_convertListToText}()},
\code{\link{table_PKsummaryCompare}()},
\code{\link{table_PKsummaryResults}()},
\code{\link{table_covariateSummaryResults}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Reporting}
