% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-NLMEmodeling.R
\name{testPKmodelsIQR}
\alias{testPKmodelsIQR}
\title{testPKmodelsIQR}
\usage{
testPKmodelsIQR(
  testSetup,
  data,
  parOptions = list(),
  projectPath = "PKmodels",
  FLAGrun = TRUE,
  tool = "MONOLIX",
  toolVersion = NULL,
  ncores = 1,
  Nparallel = 1,
  NPROCESSORS = 1,
  NPROCESSORSpar = 1,
  setting = NULL
)
}
\arguments{
\item{testSetup}{A list containing the model assumptions to test, with fields
'Compartments' (numeric, 1, 2 or 3),
'Absorption' (character vector, 'zero order' or 'first order')
'Elimination' (character vector, 'linear' or 'saturable' or 'linear+saturable'),
'LagTime' (logical, TRUE or FALSE),
and 'ErrorModels (character vector, 'abs', 'rel' or 'absrel')'.
multiple options can be tested for each field i.e. list(Compartments = c(1,2,3), Absorption =c('zero order', 'first order'),....)}

\item{data}{A data frame containing the dataset to be used for testing the models.}

\item{parOptions}{Default: list()}

\item{projectPath}{Default: 'PKmodels'}

\item{FLAGrun}{Default: \code{TRUE}}

\item{tool}{Default: 'MONOLIX'}

\item{toolVersion}{Default: \code{NULL}}

\item{ncores}{Default: 1}

\item{Nparallel}{Default: 1}

\item{NPROCESSORS}{Default: 1}

\item{NPROCESSORSpar}{Default: 1}

\item{setting}{Default: \code{NULL}}
}
\value{
Nothing is returned directly by the function, tested PK models are written to \code{projectPath}
}
\description{
Test PK models using IQRtools
}
\seealso{
Other NLME Modeling: 
\code{\link{generate_InitialParametersIQRnlme}()},
\code{\link{generate_NLMEalgorithmSettings}()},
\code{\link{modelSpec_MMVmalariaProject}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Daniel Kaschek (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{NLME Modeling}
