% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{assess_PDcomboModel}
\alias{assess_PDcomboModel}
\title{assess_PDcomboModel}
\usage{
assess_PDcomboModel(
  ModelToAssess,
  fileDataGeneral,
  hiDoseGr,
  LLOQ,
  modelFile = NULL,
  PDname = "Parasitemia Total",
  abs0inputs = NULL,
  abs0Tk0param = NULL
)
}
\arguments{
\item{ModelToAssess}{IQRnlmeProject path, IQRnlmeProjectMulti path, IQRsysFit object or IQRsysFitMulti object, in case of IQRnlmeProjectMulti or IQRsysFitMulti best model will be used}

\item{hiDoseGr}{character vector with TRTNAME of highest dose group(s) for calculation of assessment criteria for these separately}

\item{LLOQ}{Numeric, lower limit of quantification for PD measurement}

\item{modelFile}{path to the PKPD combo model file}

\item{PDname}{Default: 'Parasitemia Total'}

\item{abs0inputs}{Numeric vector: Number or vector matching INPUT(S) in the MODEL file, default \code{NULL}}

\item{abs0Tk0param}{Names of abs0param eg Tk0, default: \code{NULL}}

\item{fileDataGeneralpath}{to general dataset for the observations (to make sure that BLQ values are included). Expects PD on linear scale.}
}
\value{
data.frame with assessment criteria:
\RMSEtime2REC Sum of squares for time to recrudescence
\BIC Bayesian Information Criterion
\RMSEtime2RecHI Sum of squares for time to recrudescence in highest dose group
\RMSEparasitemiaHI Sum of squares for parasitemia data in highest dose group
\RMSEparasitemiaCL Sum of squares for clearance phase
\Robustness Robustness criterion: sqrt(sum((OBJvalues-minOBJ)^2)/nOBJ)
\Ntime2recrud Number of recrudescence cases
\Nind Number of individuals
\Npar Number of parameters
}
\description{
Assess a PKPD combo model(s) and produce key summary statistics
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
