% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{backwardElimination}
\alias{backwardElimination}
\title{backwardElimination}
\usage{
backwardElimination(
  model,
  dosing,
  data,
  modelSpec,
  covariateToTest,
  COVcentering = NULL,
  projectPath = "PKmodels",
  alpha = 0.05,
  tool = "MONOLIX",
  toolVersion = NULL,
  ncores = 1,
  Nparallel = 1,
  setting = NULL
)
}
\arguments{
\item{model}{IQRmodel object, see \code{IQRmodel()}}

\item{dosing}{IQRdosing object, see \code{IQRdosing()}}

\item{data}{IQRdatageneral object, see \code{IQRdataGeneral()}}

\item{modelSpec}{IQR modelSpec, see \code{modelSpec_IQRest}}

\item{covariateToTest}{character vector of covariate names to test, e.g. \code{c("WT0", "AGE")}}

\item{COVcentering}{Named vector defining centering value for covariates i.e. COVcentering = c(WT0=55)}

\item{projectPath}{Path to the project folder, default: "PKmodels", covariate models will be written to disk here}

\item{alpha}{Default: 0.05}

\item{tool}{Default: 'MONOLIX'}

\item{toolVersion}{Default: \code{NULL}}

\item{ncores}{Default: 1}

\item{Nparallel}{Default: 1}

\item{setting}{Default: \code{NULL}}
}
\value{
list of paths to IQRnlmeProject folders selected by backwards elimination process
}
\description{
Perform backward elimination of covariates on an IQRnlme model given a provided dataset. Models with covariates are simulated and selected based on
comparison of models using one-tail chi sqaured test with provided \code{alpha} value
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
