% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-OldFunctions.R
\name{catCovToTxt}
\alias{catCovToTxt}
\title{catCovToTxt}
\usage{
catCovToTxt(data, cat0ColNames = NULL)
}
\arguments{
\item{cat0ColNames}{Default: NULL}
}
\description{
catCovToTxt
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{DataPreparation}
