% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{censorGametocyte}
\alias{censorGametocyte}
\title{censorGametocyte}
\usage{
censorGametocyte(
  data,
  threshold,
  nameAll = "Parasitemia asexual + gametocyte",
  nameGam = "Parasitemia female gametocyte",
  FLAGfollowing = TRUE
)
}
\arguments{
\item{nameAll}{Default: 'Parasitemia asexual + gametocyte'}

\item{nameGam}{Default: 'Parasitemia female gametocyte'}

\item{FLAGfollowing}{Default: TRUE}
}
\description{
censorGametocyte
}
\seealso{
Other Data Preparation: 
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan)
}
\concept{Data Preparation}
