% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_DataModelMMV}
\alias{compare_DataModelMMV}
\title{compare_DataModelMMV}
\usage{
compare_DataModelMMV(
  modelFolder,
  dataFile = NULL,
  outputFolder = ".",
  ylabel = NULL,
  stratify = "TRTNAME",
  catToStratify = NULL,
  fileName = "DataPredComparison",
  logY = FALSE,
  figLimX = NULL,
  figLimY = NULL,
  predToPlot = NULL,
  doseCOL = "DOSELEVEL",
  doseMultCOL = "DOSEMULT",
  ActivityPath = NULL
)
}
\arguments{
\item{dataFile}{Default: \code{NULL}}

\item{outputFolder}{Default: '.'}

\item{ylabel}{Default: \code{NULL}}

\item{stratify}{Default: 'TRTNAME'}

\item{catToStratify}{Default: \code{NULL}}

\item{fileName}{Default: 'DataPredComparison'}

\item{logY}{Default: \code{FALSE}}

\item{figLimX}{Default: \code{NULL}}

\item{figLimY}{Default: \code{NULL}}

\item{predToPlot}{Default: \code{NULL}}

\item{doseCOL}{Default: 'DOSELEVEL'}

\item{doseMultCOL}{Default: 'DOSEMULT'}

\item{ActivityPath}{Default: \code{NULL}}
}
\value{

}
\description{
compare_DataModelMMV
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
