% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_DataPopPred}
\alias{compare_DataPopPred}
\title{compare_DataPopPred}
\usage{
compare_DataPopPred(
  PKmodelFolder,
  PDmodelFolder,
  dataFile,
  TRTsubset = NULL,
  GRmodelFolder = NULL,
  exclude = NULL,
  yLim = NULL,
  IndCovariates = NULL,
  ActivityPath = NULL,
  addArgs_sim_IQRmodel = NULL
)
}
\arguments{
\item{TRTsubset}{Default: \code{NULL}}

\item{GRmodelFolder}{Default: \code{NULL}}

\item{exclude}{Default: \code{NULL}}

\item{yLim}{Default: \code{NULL}}

\item{IndCovariates}{Default: \code{NULL}}

\item{ActivityPath}{Default: \code{NULL}}
}
\description{
compare_DataPopPred
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
