% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_IndPredFits}
\alias{compare_IndPredFits}
\title{compare_IndPredFits}
\usage{
compare_IndPredFits(
  modelFolder,
  outputFolder,
  stratify = "TRTNAME",
  fileName = "DataPredComparison",
  logY = FALSE,
  figLimX = NULL,
  figLimY = NULL,
  predToPlot = NULL,
  withVehicle = FALSE,
  ActivityPath = NULL
)
}
\arguments{
\item{stratify}{Default: 'TRTNAME'}

\item{fileName}{Default: 'DataPredComparison'}

\item{logY}{Default: \code{FALSE}}

\item{figLimX}{Default: \code{NULL}}

\item{figLimY}{Default: \code{NULL}}

\item{predToPlot}{Default: \code{NULL}}

\item{withVehicle}{Default: \code{FALSE}}

\item{ActivityPath}{Default: \code{NULL}}
}
\value{

}
\description{
compare_IndPredFits
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Aline Fuchs (MMV), Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
