% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_ModelEstimateRobustness}
\alias{compare_ModelEstimateRobustness}
\title{compare_ModelEstimateRobustness}
\usage{
compare_ModelEstimateRobustness(
  FitList,
  filename,
  ModelAlignment = "Vertical",
  PDddiPara = c("Alpha", "Beta", "Alpha12", "Alpha21", "Beta12", "Beta21", "Gamma")
)
}
\arguments{
\item{FitList}{A path to a folder containing IQRNLME project folders or
an IQRnlmeProjectMulti object.}

\item{filename}{Character string denoting the folder name where outputs will
be saved. Note: the argument name is misleading, this is the folder name,
file names will be included by the function by default.}

\item{ModelAlignment}{Default: 'Vertical'}

\item{PDddiPara}{Default: c("Alpha", "Beta", "Alpha12", "Alpha21", "Beta12", "Beta21",
"Gamma")}
}
\value{
Nothing is returned directly by the function, visualizations of
model comparisons are saved as .png to \code{filename}
}
\description{
Compare model estimates and robustness of provided DDI models
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Aline Fuchs (MMV), Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
