% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{compare_PredFitsSYS}
\alias{compare_PredFitsSYS}
\title{compare_PredFitsSYS}
\usage{
compare_PredFitsSYS(
  fit,
  pathname = "ComparePredObs",
  logY = FALSE,
  stratify = "CONDITION",
  data = NULL,
  modelFile = NULL,
  FLAGrefit = FALSE
)
}
\arguments{
\item{pathname}{Default: 'ComparePredObs'}

\item{logY}{Default: \code{FALSE}}

\item{stratify}{Default: 'CONDITION'}

\item{data}{Default: \code{NULL}}

\item{modelFile}{Default: \code{NULL}}

\item{FLAGrefit}{Default: \code{FALSE}}
}
\value{

}
\description{
compare_PredFitsSYS
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
