% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{convert_SCIDconcentration}
\alias{convert_SCIDconcentration}
\title{convert_SCIDconcentration}
\usage{
convert_SCIDconcentration(
  ConcToConv,
  HuRbp,
  HuPPB,
  MouseRbp,
  MousePPB,
  HuH = 0.45,
  MouseH = 0.45,
  huRBCinSCID = 0.6,
  scidH = 0.8,
  ConvertTo = c("Human Blood", "Human Plasma", "Mouse Blood", "Mouse Plasma", "Free")[1]
)
}
\arguments{
\item{ConcToConv}{Numeric vector containing concentration data to convert}

\item{HuRbp}{Numeric: value of human ratio blood:plasma}

\item{HuPPB}{Numeric: \% value of human plasma protein binding as a proportion}

\item{MouseRbp}{Numeric: value of mice ratio blood:plasma}

\item{MousePPB}{Numeric: \% value of mice plasma protein binding as a proportion}

\item{HuH}{Numeric: Value of \% human hematocrit as a proportion. Default: 0.45}

\item{MouseH}{Numeric: Value of \% mouse hematocrit as a proportion. Default: 0.45}

\item{huRBCinSCID}{Numeric: value of \% human red blood cells in SCID as proportion. Default: 0.6}

\item{scidH}{Numeric: Value of \% SCID  hematocrit as a proportion. Default: 0.8}

\item{ConvertTo}{Character string defining matrix to convert to. One of 'Human Blood', 'Human Plasma', 'Mouse Blood', 'Mouse Plasma' or 'Free'. Default: Human Blood}
}
\value{
Numeric vector containing converted concentration data
}
\description{
convert_SCIDconcentration to selected matrix
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
