% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-NLMEmodeling.R
\name{generate_InitialParametersIQRnlme}
\alias{generate_InitialParametersIQRnlme}
\title{generate_InitialParametersIQRnlme}
\usage{
generate_InitialParametersIQRnlme(projectPath, IIVvalues0 = NULL)
}
\arguments{
\item{IIVvalues0}{Default: \code{NULL}}
}
\description{
generate_InitialParametersIQRnlme
}
\seealso{
Other NLME Modeling: 
\code{\link{generate_NLMEalgorithmSettings}()},
\code{\link{modelSpec_MMVmalariaProject}()},
\code{\link{testPKmodelsIQR}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{NLME Modeling}
