% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{get_LagTime}
\alias{get_LagTime}
\title{get_LagTime}
\usage{
get_LagTime(
  dataSim,
  GR,
  EMAX,
  effCOL = "Eff",
  Parasitemia = "PL",
  Plog = TRUE,
  lagTimeTYPE = "MIC"
)
}
\arguments{
\item{dataSim}{Simulation data generated by \code{sim_IQRmodel()} or equivalent}

\item{GR}{Numeric, value of growth rate.}

\item{EMAX}{Numeric, value of EMAX.}

\item{effCOL}{Character string denoting name of eff column in \code{dataSim}. Default: 'Eff'}

\item{Parasitemia}{Character string denoting name of parasitaemia column in \code{dataSim}.Default: "PL"}

\item{Plog}{Logical, is parasitaemia in \code{dataSim} in log-scale? Default: \code{TRUE}}

\item{lagTimeTYPE}{Character string: Which threshold to assess lag time against? Must be One of \code{Baseline}, \code{MIC}, \code{MPC90} or \verb{1Log}. Default: "MIC"}
}
\value{
Estimate of lag time.
}
\description{
Estimates lag-time of drug effect from simulated time v parasitaemia data
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getCure}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Aline Fuchs (MMV), Mohammed H. Cherkaoui (MMV)
}
\concept{Key Parameters}
