% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-GeneralFunctions.R
\name{get_MMVmalariaPath}
\alias{get_MMVmalariaPath}
\title{Retrieve Package Path}
\usage{
get_MMVmalariaPath(subdir = NULL)
}
\description{
Retrieve Package Path
}
\seealso{
Other General Functions: 
\code{\link{saveActivityInfo}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{General Functions}
