% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-KeyParameters.R
\name{getCure}
\alias{getCure}
\title{getCure}
\usage{
getCure(
  dataSim,
  paraCOL = "Parasitemia",
  LLOQ = 10,
  CureThreshold = 1/5000,
  Day = 28
)
}
\arguments{
\item{dataSim}{Simulation data generated by \code{sim_IQRmodel()} or equivalent}

\item{paraCOL}{Character string denoting name of parasitaemia column in \code{dataSim} Default: "PL"}

\item{LLOQ}{Numeric: LLOQ of parasitaemia measurements. Default: 10}

\item{CureThreshold}{Numeric: Cure threshold Default: 1/5000}

\item{Day}{Numeric: Which day to determine if cure has occured by. Default: 28}
}
\value{
Vector of cure events or time at which <LLOQ occurred or maximum time, if cure does not occur.
}
\description{
Evaluate simulation data to see if cure has occured.
}
\seealso{
Other Key Parameters: 
\code{\link{assess_MICMPC90PRR48}()},
\code{\link{convert_SCIDconcentration}()},
\code{\link{estimate_AUC}()},
\code{\link{estimate_LambdaZ}()},
\code{\link{evaluate_APR}()},
\code{\link{evaluate_APR28}()},
\code{\link{formulaPKpar_MacroToMicro}()},
\code{\link{formulaPKpar_MacroToPrimary}()},
\code{\link{formulaPKpar_MicroToMacro}()},
\code{\link{formulaPKpar_MicroToPrimary}()},
\code{\link{formulaPKpar_PrimaryToMacro}()},
\code{\link{formulaPKpar_PrimaryToMicro}()},
\code{\link{formulaPKpar_PrimaryToSecondary}()},
\code{\link{formula_SCIDrbp}()},
\code{\link{getApparentKeys}()},
\code{\link{getKeysEMAX}()},
\code{\link{getTimeAboveMIC}()},
\code{\link{getTimeAboveMICsim}()},
\code{\link{getTimeAboveMPC90}()},
\code{\link{getTimeAboveMPC90sim}()},
\code{\link{getTimeKRaboveGR}()},
\code{\link{getTimeRecrudescence}()},
\code{\link{get_LagTime}()},
\code{\link{get_NumberOfPKcpt}()},
\code{\link{get_PRRtot}()},
\code{\link{plot_keyPDparametersMMV}()}
}
\author{
Aline Fuchs (MMV)
}
\concept{Key Parameters}
