% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-ModelComparison.R
\name{getEstimateTable}
\alias{getEstimateTable}
\title{getEstimateTable}
\usage{
getEstimateTable(modelFolder)
}
\arguments{
\item{modelFolder}{Character string denoting path to IQR project}
}
\value{
Table of estimates
}
\description{
Generates tables of fixed and random effects estimates and errors for provided model
}
\seealso{
Other ModelComparison: 
\code{\link{table_ModelEstimates}()}
}
\author{
Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{ModelComparison}
