% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{getPopParameters_MMVmalariaXLS}
\alias{getPopParameters_MMVmalariaXLS}
\title{getPopParameters_MMVmalariaXLS}
\usage{
getPopParameters_MMVmalariaXLS(filename, IndCovariates = NULL)
}
\arguments{
\item{IndCovariates}{Default: \code{NULL}}
}
\description{
getPopParameters_MMVmalariaXLS
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
