% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{import_SCIDpkpdData}
\alias{import_SCIDpkpdData}
\title{import_SCIDpkpdData}
\usage{
import_SCIDpkpdData(
  dataFile,
  compound,
  PKsheet = NULL,
  PKrange = NULL,
  PKmapping,
  PKlloq = 1,
  PKlloqIdentifier = c("BLQ", "<LLOQ", "< LLOQ"),
  PKfactor = 0.001,
  PDsheet = NULL,
  PDrange = NULL,
  PDmapping,
  PDlloq = 0.01,
  PDlloqIdentifier = c("BLQ", "<0.01", "< 0,01", "< 0.01"),
  PDname = "Parasitemia Total",
  study = NULL,
  DayOfFirstDrugAdmin = NULL,
  centerNumber = -1,
  centerName = "",
  visitNumber = -1,
  nrdoses = NULL,
  intervaldose = NULL,
  route = NULL,
  DoseTimeSheet = "DosingTimeTable",
  DoseTimeRange = NULL,
  DoseTimeMapping,
  DoseSheet = "IndividualMetadataDrug1",
  DoseRange = NULL,
  DoseMapping = NULL,
  positiveQC,
  importVersion = centerName
)
}
\arguments{
\item{PKsheet}{Default: NULL}

\item{PKrange}{Default: NULL}

\item{PKlloq}{Default: 1}

\item{PKlloqIdentifier}{Default: c("BLQ", "<LLOQ", "< LLOQ")}

\item{PKfactor}{Default: 0.001}

\item{PDsheet}{Default: NULL}

\item{PDrange}{Default: NULL}

\item{PDlloq}{Default: 0.01}

\item{PDlloqIdentifier}{Default: c("BLQ", "<0.01", "< 0,01", "< 0.01")}

\item{PDname}{Default: 'Parasitemia Total'}

\item{study}{Default: NULL}

\item{DayOfFirstDrugAdmin}{Default: NULL}

\item{centerNumber}{Default: -1}

\item{centerName}{Default: ''}

\item{visitNumber}{Default: -1}

\item{nrdoses}{Default: NULL}

\item{intervaldose}{Default: NULL}

\item{route}{Default: NULL}

\item{DoseTimeSheet}{Default: 'DosingTimeTable'}

\item{DoseTimeRange}{Default: NULL}

\item{DoseSheet}{Default: 'IndividualMetadataDrug1'}

\item{DoseRange}{Default: NULL}

\item{DoseMapping}{Default: NULL}

\item{importVersion}{Default: centerName}
}
\description{
import_SCIDpkpdData
}
\seealso{
Other Data Preparation: 
\code{\link{Check_MissingDatabyNAME}()},
\code{\link{censorGametocyte}()},
\code{\link{check_dataGeneralMMV}()},
\code{\link{check_dataGeneralMMV_Center}()},
\code{\link{convertGametocytes}()},
\code{\link{convert_Gametocytes}()},
\code{\link{convert_catCovToTxt}()},
\code{\link{import_SCIDpkpdDataCombo}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{list_Center}()},
\code{\link{load_AfricanPediatricMalariaPopulation2to5}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()},
\code{\link{summarizeReplicates}()},
\code{\link{swapName_MMVnameToName}()},
\code{\link{swapName_NameToMMVname}()},
\code{\link{transform_dataFrame_LongToWide}()},
\code{\link{transform_dataFrame_WideToLong}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Catalina Barcelo (MMV), Mohammed H. Cherkaoui (MMV), \email{cherkaouim@mmv.org}
}
\concept{Data Preparation}
