% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-DataPreparation.R
\name{import_SCIDpkpdDataCombo}
\alias{import_SCIDpkpdDataCombo}
\title{import_SCIDpkpdDataCombo}
\usage{
import_SCIDpkpdDataCombo(
  dataFile,
  Compound1 = NULL,
  Compound2 = NULL,
  PKsheets = NULL,
  PKranges = NULL,
  PKmapping,
  PKlloq = c(0.001, 0.001),
  PKlloqIdentifier = c("BLQ", "<LLOQ", "< LLOQ", "BLOQ"),
  PKfactor = c(0.001, 0.001),
  PDsheet = NULL,
  PDrange = NULL,
  PDmapping,
  PDlloq = 0.01,
  PDlloqIdentifier = c("BLQ", "<0.01"),
  PDname = "Parasitemia Total",
  study = NULL,
  studyID = NULL,
  DayOfFirstDrugAdmin = NULL,
  centerNumber = -1,
  centerName = "",
  visitNumber = -1,
  intervaldose = NULL,
  DoseTimeSheet = "DosingTimeTable",
  DoseTimeRange = NULL,
  DoseTimeMapping,
  DoseSheet = "DrugTreatmentTable",
  DoseRange = NULL,
  DoseMapping = NULL,
  positiveQC = NULL,
  route = NULL,
  importVersion = "oldTAD"
)
}
\arguments{
\item{dataFile}{Path to the Excel file containing the data.}

\item{Compound1}{Default: NULL}

\item{Compound2}{Default: NULL}

\item{PKsheets}{Name of the single sheet recording PK data}

\item{PKranges}{Default: NULL}

\item{PKmapping}{Character vector mapping PKsheet headers (left) to desired name in converted data format (right) (eg c("StudyID"= "STUDY", "Concentration" = "VALUE"))}

\item{PKlloq}{Default: c(0.001, 0.001)}

\item{PKlloqIdentifier}{Default: c("BLQ", "<LLOQ", "< LLOQ", "BLOQ")}

\item{PKfactor}{Default: c(0.001, 0.001)}

\item{PDsheet}{Default: NULL}

\item{PDrange}{Default: NULL}

\item{PDmapping}{Character vector mapping PKsheet headers (left) to desired name in converted data format (right) (eg c("StudyID"= "STUDY", "Concentration" = "VALUE"))}

\item{PDlloq}{Default: 0.01}

\item{PDlloqIdentifier}{Default: c("BLQ", "<0.01")}

\item{PDname}{Default: 'Parasitemia Total'}

\item{study}{Default: NULL}

\item{studyID}{Default: NULL}

\item{DayOfFirstDrugAdmin}{Default: NULL}

\item{centerNumber}{Default: -1}

\item{centerName}{Default: ''}

\item{visitNumber}{Default: -1}

\item{intervaldose}{Default: NULL}

\item{DoseTimeSheet}{Default: 'DosingTimeTable'}

\item{DoseTimeRange}{Default: NULL}

\item{DoseTimeMapping}{Character vector mapping DoseTimesheet headers (left) to desired name in converted final dataset (right) (eg c("StudyID"="STUDY", "DosingTime"="NT"))}

\item{DoseSheet}{Default: 'DrugTreatmentTable'}

\item{DoseRange}{Default: NULL}

\item{DoseMapping}{Default: NULL}

\item{positiveQC}{Character vector identifying the name of positive control (must match upper or lower case as in original dataset)}

\item{route}{Default: NULL}

\item{importVersion}{Default: 'oldTAD'}
}
\value{
Imported SCID data
}
\description{
Imports SCID PK and PD data from an Excel file.
}
\seealso{
Other Data Preparation: 
\code{\link{censorGametocyte}()},
\code{\link{convert_Gametocytes}()},
\code{\link{import_SCIDpkpdData}()},
\code{\link{import_SCIDpkpdDataCombo_GSK}()},
\code{\link{import_SCIDpkpdDataCombo_TAD}()},
\code{\link{import_SCIDpkpdDataCombo_oldTAD}()},
\code{\link{import_SCIDpkpdData_GSK}()},
\code{\link{import_SCIDpkpdData_TAD}()},
\code{\link{import_SCIDpkpdData_oldTAD}()},
\code{\link{load_MalariaPopulation}()},
\code{\link{load_PediatricParameters}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Data Preparation}
