% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{plotVPC_MMVdataVPC}
\alias{plotVPC_MMVdataVPC}
\title{MMV wrapper for plotVPC_IQRdataVPC}
\usage{
plotVPC_MMVdataVPC(
  dataVPC,
  stratifyBy = NULL,
  filename = NULL,
  FLAGlogY = FALSE,
  FLAGuseTAD = FALSE,
  FLAGdataPlotOnly = FALSE,
  FLAGnoDataPoints = FALSE,
  FLAGpc = FALSE,
  FLAGplotBins = FALSE,
  FLAGplotN = FALSE,
  title = NULL,
  subtitle = NULL,
  BIN.column = NULL,
  BIN.breaks = NULL,
  BIN.groupsize = NULL,
  BIN.lambda = 1,
  BIN.resolution = 0.1,
  CIlevel = 95,
  percentiles = c(5, 95),
  alphaDataPoints = 0.4,
  alphaPredInter = 0.4,
  ActivityPath = NULL,
  Caption = NULL,
  setting = NULL
)
}
\arguments{
\item{dataVPC}{Named list with simulated (sim) and potentially typical predictions for observed (obs) data}

\item{stratifyBy}{Column name used to stratify VPC plots. Default: \code{NULL}}

\item{filename}{Filename for export of the VPC to a PDF. If NULL object including plotting data is returned. Default: \code{NULL}}

\item{FLAGlogY}{logical. If TRUE the Y axis will be shown in logscale. Default: \code{FALSE}}

\item{FLAGuseTAD}{logical. If TRUE then observed time after previous dose (TAD) is used for x-axis. If FALSE then actual TIME is used. Default: \code{FALSE}}

\item{FLAGdataPlotOnly}{logical. If TRUE then no simulation is done and only the data and lines for the percentiles are plotted. This allows to find suitable settings for the binning parameters BIN.groupsize and BIN.lambda. Default: \code{FALSE}}

\item{FLAGnoDataPoints}{If TRUE then data points will be ommitted but percentiles for data will be shown. Default: \code{FALSE}}

\item{FLAGpc}{Flag whether to apply prediction correction. Can only be performed if typical prediction are present in dataVPC. Default: \code{FALSE}}

\item{FLAGplotBins}{Flag whether to plot the bin boundaries as vertical lines. Default: \code{FALSE}}

\item{FLAGplotN}{Flag whether to display number of subjects in panel. Default: \code{FALSE}}

\item{title}{Character string to be plotted as plot title (on each page). Default: \code{NULL}}

\item{subtitle}{Character string to be plotted as plot subtitle (on each page). Default: \code{NULL}}

\item{BIN.column}{Column name of integer column assigning the observations to time bins. Defaults to NULL such that bins are automatically generated or the user-provided BIN.breaks are used. Default: \code{NULL}}

\item{BIN.breaks}{Numerical vector with bin borders that are used for the observation times. Defaults to NULL such that bins are automatically generated. This input is ignored if BIN.column is provided. Default: \code{NULL}}

\item{BIN.groupsize}{Smallest expected group size for the binning. By default the round(0.5 x number of subjects) in the dataset is used. Default: \code{NULL}}

\item{BIN.lambda}{Penalization of intra-group variance, set to 1 to have more groups and set to 0 to get less but larger groups. Default: \code{NULL}}

\item{BIN.resolution}{Gaps between groups of data points greater than resolution lead to separation of groupsDefault: \code{NULL}}

\item{CIlevel}{Confidence interval level for percentiles (in percent). Default: \code{NULL}}

\item{percentiles}{numeric vector with 1, 2 or 3 entries between 0 and 100 to be used as percentiles for data and simulation plots. 1 entry: only this percentile is plotted, e.g. for 50 the median is visualized. 2 entries: low, and high percentiles were provided and median should also be plotted. 3 entries: low, medium and high percentiles were provided in this order. Default: \code{NULL}}

\item{alphaDataPoints}{Alpha transparency level for data points. Default: \code{NULL}}

\item{alphaPredInter}{Default: \code{NULL}}

\item{ActivityPath}{Default: \code{NULL}}

\item{Caption}{Default: \code{NULL}}

\item{setting}{Default: \code{NULL}}
}
\value{
List of generated VPC figures
}
\description{
Upgrades the corresponding function in IQRtools. Check its documentation for a description of the arguments.
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()},
\code{\link{table_EstimatesComboNLME}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{Model Assessment}
