% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{simulate_ComboMouse2Human}
\alias{simulate_ComboMouse2Human}
\title{simulate_ComboMouse2Human}
\usage{
simulate_ComboMouse2Human(
  TreatmentGroups,
  PKmodelFolders,
  PDmodelFolders,
  InteractionModelFolder,
  ComboModel,
  PbaseSample = function(x) {     sample_Distribution(Mean = 10^6.72, SD = 0.1, Dist =
    "L", nSample = x) },
  GRsample = function(x) {     sample_Distribution(Mean = 0.048, SD = 0.1, Dist = "L",
    nSample = x) },
  doseCovariate = list(drug1 = NULL, drug2 = NULL),
  covariates = NULL,
  evalDay = 28,
  simtime = create_PKPDsimtime(24 * evalDay),
  nTrial = 10,
  nSubj = 10,
  LLOQ.PD,
  cureThreshold,
  replace = TRUE,
  setting = NULL
)
}
\arguments{
\item{PbaseSample}{Default: function(x) {
   sample_Distribution(Mean = 10^6.72, SD = 0.1, Dist = "L",
       nSample = x)
}}

\item{GRsample}{Default: function(x) {
   sample_Distribution(Mean = 0.048, SD = 0.1, Dist = "L", nSample = x)
}}

\item{doseCovariate}{Default: list(drug1 = NULL, drug2 = NULL)}

\item{covariates}{Default: \code{NULL}}

\item{evalDay}{Default: 28}

\item{simtime}{Default: create_PKPDsimtime(24 * evalDay)}

\item{nTrial}{Default: 10}

\item{nSubj}{Default: 10}

\item{replace}{Default: TRUE}

\item{setting}{Default: \code{NULL}}
}
\value{

}
\description{
simulate_ComboMouse2Human
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_PKPDseparate}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
