% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-Simulations.R
\name{simulate_PKPDseparate}
\alias{simulate_PKPDseparate}
\title{simulate_PKPDseparate}
\usage{
simulate_PKPDseparate(
  Doses,
  Ndoses,
  PKmodelFolder,
  PDmodelFolder,
  PKPDmodelFile,
  filename = NULL,
  EC50adj = 1,
  DOSEcovariate = NULL,
  covariates = NULL,
  simtime = 0:800,
  Npop = 10,
  Nind = 10,
  FLAGsamplePK = 1,
  FLAGreturnObject = FALSE,
  LLOQ.PK = NULL,
  LLOQ.PD = NULL,
  cureThreshold = NULL,
  PLbaseSample = NULL,
  setting = NULL
)
}
\arguments{
\item{Doses}{Numeric vector of doses to simulate}

\item{Ndoses}{Integer vector of same length as \code{Doses} giving the number of doses}

\item{PKmodelFolder}{String denoting path to IQRnlme project with PK model}

\item{PDmodelFolder}{String denoting path to IQRnlme project with PD model}

\item{PKPDmodelFile}{IQRmodel object or path to it, containing both PK and PD}

\item{filename}{Optional string denoting filepath to write figures to. Default: \code{NULL}}

\item{EC50adj}{Numeric scalar to adjust EC50 value. Default: 1}

\item{DOSEcovariate}{Named string denoting the name of the dose covariate.
Default: \code{NULL}}

\item{covariates}{Named list of covariate vectors or sampling functions. Default: \code{NULL}}

\item{simtime}{Numeric vector of simulation time in hours.
Default: MMVbase::create_PKPDsimtime(24 * evalDay)}

\item{Npop}{integer, number of virtual populations to simulate. Default: 10}

\item{Nind}{Integer, number of individuals to simulate per population. Default: 10}

\item{FLAGsamplePK}{Numeric, value for sample flag, see \code{?IQRtools::sample_IQRnlmeProject}. Default: 0}

\item{FLAGreturnObject}{Logical, return the simulation data? Default: FALSE}

\item{LLOQ.PK}{Numeric, lower limit of quantification of concentration. Default: \code{NULL}}

\item{LLOQ.PD}{Numeric, lower limit of quantification of parasitaemia. Default: \code{NULL}}

\item{cureThreshold}{Numeric, value of parasitaemia to treat as cure threshold. Default: \code{NULL}}

\item{PLbaseSample}{Function from which to sample baseline parasitaemia. Default: \code{NULL}}

\item{setting}{Optional named list of settings for solver. Default: \code{NULL}}
}
\value{
if \code{FLAGreturnObject} = TRUE, data.frame with simulation output.
}
\description{
Simulate PK/PD models using separate PK and PD NLME models. In other words,
loads PK and PD parameters from separate IQRnlme projects, then simulates a provided
combined PKPD structural model with the provided PK and PD models
}
\seealso{
Other Simulations: 
\code{\link{applyCureThreshold}()},
\code{\link{generateTrialSim}()},
\code{\link{generate_GPFfromOldIQRproject}()},
\code{\link{generate_GPFfromParametersTable}()},
\code{\link{getModelParameters_MMVmalariaCSV}()},
\code{\link{getModelParameters_MMVmalariaProject}()},
\code{\link{getModelParameters_MMVmalariaXLS}()},
\code{\link{getPopParameters_MMVmalariaCSV}()},
\code{\link{getPopParameters_MMVmalariaProject}()},
\code{\link{getPopParameters_MMVmalariaXLS}()},
\code{\link{predict_HumanChallenge}()},
\code{\link{sample_CustomParameters}()},
\code{\link{sample_MMVmalariaCSV}()},
\code{\link{sample_MMVmalariaCSV_NEW}()},
\code{\link{sample_MMVmalariaProject}()},
\code{\link{sample_MMVmalariaXLS}()},
\code{\link{simTimeRecrudescenceCombo}()},
\code{\link{simulate_ComboMouse2Human}()},
\code{\link{simulate_VirtualTrials}()}
}
\author{
Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{Simulations}
