% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-DataExploration.R
\name{summary_PKdata_byStudyTRT}
\alias{summary_PKdata_byStudyTRT}
\title{summary_PKdata_byStudyTRT}
\usage{
summary_PKdata_byStudyTRT(
  dataGen,
  filename = NULL,
  NewTRTNAME = NULL,
  CovToTRTNAME = NULL,
  CovNewTRTNAME = NULL
)
}
\arguments{
\item{dataGen}{IQRdataGeneral object}

\item{filename}{Default: NULL}

\item{NewTRTNAME}{Character vector denoting new treatment name. Default: NULL}

\item{CovToTRTNAME}{Default: NULL}

\item{CovNewTRTNAME}{Default: NULL}
}
\value{
Data frame with summary of PK data by study and treatment
}
\description{
Summarize PK data by study and treatment
}
\seealso{
Other DataExploration: 
\code{\link{barplotACPR}()},
\code{\link{display_ComboTreatments}()},
\code{\link{plot_2DconcVsACPR}()},
\code{\link{plot_PKPDdataMMV}()},
\code{\link{plot_PKPDdataMMVhuCh}()},
\code{\link{plot_PKPDdataMMVhuCh_Ratio_Gam}()},
\code{\link{plot_PKPDhuChCombo}()},
\code{\link{plot_PKPDscidCombo}()},
\code{\link{plot_PKdataMMV}()},
\code{\link{scatterConcACPR}()},
\code{\link{summary_PDdata_byStudyTRT}()},
\code{\link{summary_PKPDdataCombo}()},
\code{\link{summary_PKPDdataSimplified}()},
\code{\link{summary_PKdataMMV}()},
\code{\link{summarybyGroup_MMVnca}()}
}
\author{
Mohammed H. Cherkaoui (MMV)
}
\concept{DataExploration}
