% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-SYSFITmodeling.R
\name{table_BestFitEstimation}
\alias{table_BestFitEstimation}
\title{table_BestFitEstimation}
\usage{
table_BestFitEstimation(
  fit,
  filename = NULL,
  FLAGout = FALSE,
  title = NULL,
  footer = NULL
)
}
\arguments{
\item{filename}{Default: NULL}

\item{FLAGout}{Default: FALSE}

\item{title}{Default: NULL}

\item{footer}{Default: NULL}
}
\description{
table_BestFitEstimation
}
\seealso{
Other SYSFITmodeling: 
\code{\link{generate_InitialParametersIQRsys}()},
\code{\link{generate_SYSalgorithmSettings}()}
}
\author{
Aline Fuchs (MMV), Anne Kümmel (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{SYSFITmodeling}
