% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{table_covariateSummaryResults}
\alias{table_covariateSummaryResults}
\title{table_covariateSummaryResults}
\usage{
table_covariateSummaryResults(
  tableInputPath,
  tableOutputPath = NULL,
  report = TRUE,
  FUNnameModel = function(x) {
     gsub("MODEL_", "", x)
 }
)
}
\arguments{
\item{tableOutputPath}{(Default: \code{NULL}).}

\item{report}{Default: TRUE}

\item{FUNnameModel}{Default: function(x) {
gsub("MODEL_", "", x)
}}
}
\description{
table_covariateSummaryResults
}
\seealso{
Other Reporting: 
\code{\link{generate_EstimateTableForReport}()},
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_PKmodelHypos}()},
\code{\link{reportText_convertListToText}()},
\code{\link{table_PKsummaryCompare}()},
\code{\link{table_PKsummaryResults}()},
\code{\link{table_strToLaTeX}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Reporting}
