% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ModelAssessment.R
\name{table_EstimatesComboNLME}
\alias{table_EstimatesComboNLME}
\title{table_EstimatesComboNLME}
\usage{
table_EstimatesComboNLME(
  x,
  filename = NULL,
  FLAGout = FALSE,
  title = NULL,
  footer = NULL
)
}
\arguments{
\item{filename}{Default: \code{NULL}}

\item{FLAGout}{Default: \code{FALSE}}

\item{title}{Default: \code{NULL}}

\item{footer}{Default: \code{NULL}}
}
\description{
table_EstimatesComboNLME
}
\seealso{
Other Model Assessment: 
\code{\link{add_CovariateToModelSpec}()},
\code{\link{add_OneCovariateToModelSpec}()},
\code{\link{assess_PDcomboModel}()},
\code{\link{backwardElimination}()},
\code{\link{compare_DataModelMMV}()},
\code{\link{compare_DataPopPred}()},
\code{\link{compare_IndPredFits}()},
\code{\link{compare_ModelEstimateRobustness}()},
\code{\link{compare_PredFitsSYS}()},
\code{\link{forwardSelection}()},
\code{\link{likelihoodRatioTestIQR}()},
\code{\link{plotSummaryComparisonBinary}()},
\code{\link{plotSummaryComparisonContinuous}()},
\code{\link{plotVPC_MMVdataVPC}()},
\code{\link{remove_OneCovariateToModelSpec}()},
\code{\link{tablePopPDmalaria}()}
}
\author{
Aline Fuchs (MMV)
}
\concept{Model Assessment}
