% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-Reporting.R
\name{table_PKsummaryResults}
\alias{table_PKsummaryResults}
\title{table_PKsummaryResults}
\usage{
table_PKsummaryResults(
  tableInputPath,
  tableOutputPath = NULL,
  colToAdd = c("Nr. of Compartments", "Clearance", "Absorption", "Time Lag",
    "Error Model"),
  Metric = "BIC",
  xTitle = NULL,
  xFooter = NULL,
  report = TRUE
)
}
\arguments{
\item{tableInputPath}{String denoting path to IQRtable .txt file saved with \code{IQRtools::IQRoutputTable()}}

\item{tableOutputPath}{(Default: \code{NULL}).}

\item{colToAdd}{Character vector denoting which columns should be included in
the output table. Default: c("Nr. of Compartments", "Clearance", "Absorption", "Time Lag",
"Error Model")}

\item{Metric}{Which metric (OBJ, BIC, AIC) to compare and arrange models with. Default: 'BIC'}

\item{xTitle}{Optional string for adding title to output table (Default: \code{NULL}).}

\item{xFooter}{Optional string for adding footer to output table (Default: \code{NULL}).}

\item{report}{Logical, TRUE: add reporting tags FALSE: do not add
(for screen output only) NULL: same as TRUE if xtable provided.Default: TRUE}
}
\description{
Read a PK model in .txt format as generated by
IQRtools, restructure it into a tidier format and save it
}
\seealso{
Other Reporting: 
\code{\link{generate_EstimateTableForReport}()},
\code{\link{parameter_strToLaTeX}()},
\code{\link{reportText_PKmodelHypos}()},
\code{\link{reportText_convertListToText}()},
\code{\link{table_PKsummaryCompare}()},
\code{\link{table_covariateSummaryResults}()},
\code{\link{table_strToLaTeX}()}
}
\author{
Mohammed H. Cherkaoui (MMV, \email{cherkaouim@mmv.org})
}
\concept{Reporting}
