% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-NLMEmodeling.R
\name{testPKmodelsIQR}
\alias{testPKmodelsIQR}
\title{testPKmodelsIQR}
\usage{
testPKmodelsIQR(
  testSetup,
  data,
  parOptions = list(),
  projectPath = "PKmodels",
  FLAGrun = TRUE,
  tool = "MONOLIX",
  toolVersion = NULL,
  ncores = 1,
  Nparallel = 1,
  NPROCESSORS = 1,
  NPROCESSORSpar = 1,
  setting = NULL
)
}
\arguments{
\item{parOptions}{Default: list()}

\item{projectPath}{Default: 'PKmodels'}

\item{FLAGrun}{Default: \code{TRUE}}

\item{tool}{Default: 'MONOLIX'}

\item{toolVersion}{Default: \code{NULL}}

\item{ncores}{Default: 1}

\item{Nparallel}{Default: 1}

\item{NPROCESSORS}{Default: 1}

\item{NPROCESSORSpar}{Default: 1}

\item{setting}{Default: \code{NULL}}
}
\value{

}
\description{
testPKmodelsIQR
}
\seealso{
Other NLME Modeling: 
\code{\link{generate_InitialParametersIQRnlme}()},
\code{\link{generate_NLMEalgorithmSettings}()},
\code{\link{modelSpec_MMVmalariaProject}()}
}
\author{
Aline Fuchs (MMV), Anne Kuemmel (IntiQuan), Daniel Kaschek (IntiQuan), Mohammed H. Cherkaoui (MMV)
}
\concept{NLME Modeling}
